+++ gcc/rust/parse/rust-parse-impl.h
@@ -34,7 +34,6 @@
 #define INCLUDE_ALGORITHM
 #include "rust-diagnostics.h"
 #include "rust-dir-owner.h"
-#include "rust-attribute-values.h"
 #include "rust-keyword-values.h"
 #include "rust-edition.h"
 #include "rust-parse-error.h"
@@ -42,105 +41,6 @@
 #include "optional.h"
 
 namespace Rust {
-// Left binding powers of operations.
-enum binding_powers
-{
-  // Highest priority
-  LBP_HIGHEST = 100,
-
-  LBP_PATH = 95,
-
-  LBP_METHOD_CALL = 90,
-
-  LBP_FIELD_EXPR = 85,
-
-  LBP_FUNCTION_CALL = 80,
-  LBP_ARRAY_REF = LBP_FUNCTION_CALL,
-
-  LBP_QUESTION_MARK = 75, // unary postfix - counts as left
-
-  LBP_UNARY_PLUS = 70,		    // Used only when the null denotation is +
-  LBP_UNARY_MINUS = LBP_UNARY_PLUS, // Used only when the null denotation is -
-  LBP_UNARY_ASTERISK = LBP_UNARY_PLUS, // deref operator - unary prefix
-  LBP_UNARY_EXCLAM = LBP_UNARY_PLUS,
-  LBP_UNARY_AMP = LBP_UNARY_PLUS,
-  LBP_UNARY_AMP_MUT = LBP_UNARY_PLUS,
-
-  LBP_AS = 65,
-
-  LBP_MUL = 60,
-  LBP_DIV = LBP_MUL,
-  LBP_MOD = LBP_MUL,
-
-  LBP_PLUS = 55,
-  LBP_MINUS = LBP_PLUS,
-
-  LBP_L_SHIFT = 50,
-  LBP_R_SHIFT = LBP_L_SHIFT,
-
-  LBP_AMP = 45,
-
-  LBP_CARET = 40,
-
-  LBP_PIPE = 35,
-
-  LBP_EQUAL = 30,
-  LBP_NOT_EQUAL = LBP_EQUAL,
-  LBP_SMALLER_THAN = LBP_EQUAL,
-  LBP_SMALLER_EQUAL = LBP_EQUAL,
-  LBP_GREATER_THAN = LBP_EQUAL,
-  LBP_GREATER_EQUAL = LBP_EQUAL,
-
-  LBP_LOGICAL_AND = 25,
-
-  LBP_LOGICAL_OR = 20,
-
-  LBP_DOT_DOT = 15,
-  LBP_DOT_DOT_EQ = LBP_DOT_DOT,
-
-  // TODO: note all these assig operators are RIGHT associative!
-  LBP_ASSIG = 10,
-  LBP_PLUS_ASSIG = LBP_ASSIG,
-  LBP_MINUS_ASSIG = LBP_ASSIG,
-  LBP_MULT_ASSIG = LBP_ASSIG,
-  LBP_DIV_ASSIG = LBP_ASSIG,
-  LBP_MOD_ASSIG = LBP_ASSIG,
-  LBP_AMP_ASSIG = LBP_ASSIG,
-  LBP_PIPE_ASSIG = LBP_ASSIG,
-  LBP_CARET_ASSIG = LBP_ASSIG,
-  LBP_L_SHIFT_ASSIG = LBP_ASSIG,
-  LBP_R_SHIFT_ASSIG = LBP_ASSIG,
-
-  // return, break, and closures as lowest priority?
-  LBP_RETURN = 5,
-  LBP_BREAK = LBP_RETURN,
-  LBP_CLOSURE = LBP_RETURN, // unary prefix operators
-
-#if 0
-  // rust precedences
-  // used for closures
-  PREC_CLOSURE = -40,
-  // used for break, continue, return, and yield
-  PREC_JUMP = -30,
-  // used for range (although weird comment in rustc about this)
-  PREC_RANGE = -10,
-  // used for binary operators mentioned below - also cast, colon (type),
-  // assign, assign_op
-  PREC_BINOP = FROM_ASSOC_OP,
-  // used for box, address_of, let, unary (again, weird comment on let)
-  PREC_PREFIX = 50,
-  // used for await, call, method call, field, index, try,
-  // inline asm, macro invocation
-  PREC_POSTFIX = 60,
-  // used for array, repeat, tuple, literal, path, paren, if,
-  // while, for, 'loop', match, block, try block, async, struct
-  PREC_PAREN = 99,
-  PREC_FORCE_PAREN = 100,
-#endif
-
-  // lowest priority
-  LBP_LOWEST = 0
-};
 
 /* HACK-y special handling for skipping a right angle token at the end of
  * generic arguments.
@@ -412,503 +312,6 @@ Parser<ManagedTokenSource>::parse_crate ()
     new AST::Crate (std::move (items), std::move (inner_attrs)));
 }
 
-// Parse a contiguous block of inner attributes.
-template <typename ManagedTokenSource>
-AST::AttrVec
-Parser<ManagedTokenSource>::parse_inner_attributes ()
-{
-  AST::AttrVec inner_attributes;
-
-  auto has_valid_inner_attribute_prefix = [&] () {
-    auto id = lexer.peek_token ()->get_id ();
-    /* Outer attribute `#[` is not allowed, only accepts `#!` */
-    return (id == HASH && lexer.peek_token (1)->get_id () == EXCLAM)
-	   || id == INNER_DOC_COMMENT;
-  };
-
-  while (has_valid_inner_attribute_prefix ())
-    {
-      auto inner_attr = parse_inner_attribute ();
-
-      /* Ensure only valid inner attributes are added to the inner_attributes
-       * list */
-      if (inner_attr)
-	{
-	  inner_attributes.push_back (std::move (inner_attr.value ()));
-	}
-      else
-	{
-	  /* If no more valid inner attributes, break out of loop (only
-	   * contiguous inner attributes parsed). */
-	  break;
-	}
-    }
-
-  inner_attributes.shrink_to_fit ();
-  return inner_attributes;
-}
-
-// Parse a inner or outer doc comment into an doc attribute
-template <typename ManagedTokenSource>
-std::tuple<AST::SimplePath, std::unique_ptr<AST::AttrInput>, location_t>
-Parser<ManagedTokenSource>::parse_doc_comment ()
-{
-  const_TokenPtr token = lexer.peek_token ();
-  location_t locus = token->get_locus ();
-  AST::SimplePathSegment segment (Values::Attributes::DOC, locus);
-  std::vector<AST::SimplePathSegment> segments;
-  segments.push_back (std::move (segment));
-  AST::SimplePath attr_path (std::move (segments), false, locus);
-  AST::LiteralExpr lit_expr (token->get_str (), AST::Literal::STRING,
-			     PrimitiveCoreType::CORETYPE_STR, {}, locus);
-  std::unique_ptr<AST::AttrInput> attr_input (
-    new AST::AttrInputLiteral (std::move (lit_expr)));
-  lexer.skip_token ();
-  return std::make_tuple (std::move (attr_path), std::move (attr_input), locus);
-}
-
-// Parse a single inner attribute.
-template <typename ManagedTokenSource>
-tl::expected<AST::Attribute, Parse::Error::Attribute>
-Parser<ManagedTokenSource>::parse_inner_attribute ()
-{
-  if (lexer.peek_token ()->get_id () == INNER_DOC_COMMENT)
-    {
-      auto values = parse_doc_comment ();
-      auto path = std::move (std::get<0> (values));
-      auto input = std::move (std::get<1> (values));
-      auto loc = std::get<2> (values);
-      return AST::Attribute (std::move (path), std::move (input), loc, true);
-    }
-
-  rust_assert (lexer.peek_token ()->get_id () == HASH);
-
-  lexer.skip_token ();
-
-  if (lexer.peek_token ()->get_id () != EXCLAM)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "expected %<!%> or %<[%> for inner attribute");
-      add_error (std::move (error));
-
-      return Parse::Error::Attribute::make_malformed ();
-    }
-  lexer.skip_token ();
-
-  if (!skip_token (LEFT_SQUARE))
-    return Parse::Error::Attribute::make_malformed ();
-
-  auto body_res = parse_attribute_body ();
-  if (!body_res)
-    return Parse::Error::Attribute::make_malformed ();
-  auto body = std::move (body_res.value ());
-
-  auto actual_attribute
-    = AST::Attribute (std::move (body.path), std::move (body.input), body.locus,
-		      true);
-
-  if (!skip_token (RIGHT_SQUARE))
-    return Parse::Error::Attribute::make_malformed ();
-
-  return actual_attribute;
-}
-
-// Parses the body of an attribute (inner or outer).
-template <typename ManagedTokenSource>
-tl::expected<Parse::AttributeBody, Parse::Error::AttributeBody>
-Parser<ManagedTokenSource>::parse_attribute_body ()
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-
-  auto attr_path = parse_simple_path ();
-  // ensure path is valid to parse attribute input
-  if (!attr_path)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "empty simple path in attribute");
-      add_error (std::move (error));
-
-      // Skip past potential further info in attribute (i.e. attr_input)
-      skip_after_end_attribute ();
-      return Parse::Error::AttributeBody::make_invalid_path ();
-    }
-
-  auto attr_input = parse_attr_input ();
-  // AttrInput is allowed to be null, so no checks here
-  if (attr_input)
-    return Parse::AttributeBody{std::move (attr_path.value ()),
-				std::move (attr_input.value ()), locus};
-  else if (attr_input.error ().kind == Parse::Error::AttrInput::Kind::MISSING)
-    return Parse::AttributeBody{std::move (attr_path.value ()), nullptr, locus};
-  else
-    return Parse::Error::AttributeBody::make_invalid_attrinput ();
-}
-
-// Parses a SimplePath AST node, if it exists. Does nothing otherwise.
-template <typename ManagedTokenSource>
-tl::expected<AST::SimplePath, Parse::Error::SimplePath>
-Parser<ManagedTokenSource>::parse_simple_path ()
-{
-  bool has_opening_scope_resolution = false;
-  location_t locus = UNKNOWN_LOCATION;
-
-  using Parse::Utils::is_simple_path_segment;
-
-  // don't parse anything if not a path upfront
-  if (!is_simple_path_segment (lexer.peek_token ()->get_id ())
-      && !is_simple_path_segment (lexer.peek_token (1)->get_id ()))
-    return Parse::Error::SimplePath::make_malformed ();
-
-  /* Checks for opening scope resolution (i.e. global scope fully-qualified
-   * path) */
-  if (lexer.peek_token ()->get_id () == SCOPE_RESOLUTION)
-    {
-      has_opening_scope_resolution = true;
-
-      locus = lexer.peek_token ()->get_locus ();
-
-      lexer.skip_token ();
-    }
-
-  // Parse single required simple path segment
-  auto segment = parse_simple_path_segment ();
-
-  if (!segment)
-    return Parse::Error::SimplePath::make_malformed ();
-
-  // get location if not gotten already
-  if (locus == UNKNOWN_LOCATION)
-    locus = segment->get_locus ();
-
-  std::vector<AST::SimplePathSegment> segments;
-  segments.push_back (std::move (segment.value ()));
-
-  // Parse all other simple path segments
-  while (lexer.peek_token ()->get_id () == SCOPE_RESOLUTION)
-    {
-      auto new_segment = parse_simple_path_segment (1);
-
-      using Error = Parse::Error::SimplePathSegment::Kind;
-      // Return path as currently constructed if segment in error state.
-      if (!new_segment)
-	{
-	  if (new_segment.error ().kind == Error::INVALID_SIMPLE_PATH_TOKEN)
-	    break; /* Could be end of path */
-	  else	   /* Any other error is an hard error */
-	    return Parse::Error::SimplePath::make_malformed ();
-	}
-
-      segments.push_back (std::move (new_segment.value ()));
-    }
-
-  return AST::SimplePath (std::move (segments), has_opening_scope_resolution,
-			  locus);
-  /* TODO: now that is_simple_path_segment exists, could probably start
-   * actually making errors upon parse failure of segments and whatever */
-}
-
-/* Parses a single SimplePathSegment (does not handle the scope resolution
- * operators)
- * Starts parsing at an offset of base_peek */
-template <typename ManagedTokenSource>
-tl::expected<AST::SimplePathSegment, Parse::Error::SimplePathSegment>
-Parser<ManagedTokenSource>::parse_simple_path_segment (int base_peek)
-{
-  using namespace Values;
-  const_TokenPtr t = lexer.peek_token (base_peek);
-  switch (t->get_id ())
-    {
-    case IDENTIFIER:
-      lexer.skip_token (base_peek);
-
-      return AST::SimplePathSegment (t->get_str (), t->get_locus ());
-    case SUPER:
-      lexer.skip_token (base_peek);
-
-      return AST::SimplePathSegment (Keywords::SUPER, t->get_locus ());
-    case SELF:
-      lexer.skip_token (base_peek);
-
-      return AST::SimplePathSegment (Keywords::SELF, t->get_locus ());
-    case CRATE:
-      lexer.skip_token (base_peek);
-
-      return AST::SimplePathSegment (Keywords::CRATE, t->get_locus ());
-    case DOLLAR_SIGN:
-      if (lexer.peek_token (base_peek + 1)->get_id () == CRATE)
-	{
-	  lexer.skip_token (base_peek + 1);
-
-	  return AST::SimplePathSegment ("$crate", t->get_locus ());
-	}
-      gcc_fallthrough ();
-    default:
-      // do nothing but inactivates warning from gcc when compiling
-      /* could put the rust_error_at thing here but fallthrough (from failing
-       * $crate condition) isn't completely obvious if it is. */
-
-      return Parse::Error::SimplePathSegment::make_invalid_token_or_path_end ();
-    }
-  rust_unreachable ();
-}
-
-// Parses a PathIdentSegment - an identifier segment of a non-SimplePath path.
-template <typename ManagedTokenSource>
-tl::expected<AST::PathIdentSegment, Parse::Error::PathIdentSegment>
-Parser<ManagedTokenSource>::parse_path_ident_segment ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case IDENTIFIER:
-      lexer.skip_token ();
-
-      return AST::PathIdentSegment (t->get_str (), t->get_locus ());
-    case SUPER:
-      lexer.skip_token ();
-
-      return AST::PathIdentSegment (Values::Keywords::SUPER, t->get_locus ());
-    case SELF:
-      lexer.skip_token ();
-
-      return AST::PathIdentSegment (Values::Keywords::SELF, t->get_locus ());
-    case SELF_ALIAS:
-      lexer.skip_token ();
-
-      return AST::PathIdentSegment (Values::Keywords::SELF_ALIAS,
-				    t->get_locus ());
-    case CRATE:
-      lexer.skip_token ();
-
-      return AST::PathIdentSegment (Values::Keywords::CRATE, t->get_locus ());
-    case DOLLAR_SIGN:
-      if (lexer.peek_token (1)->get_id () == CRATE)
-	{
-	  lexer.skip_token (1);
-
-	  return AST::PathIdentSegment ("$crate", t->get_locus ());
-	}
-      gcc_fallthrough ();
-    default:
-      /* do nothing but inactivates warning from gcc when compiling
-       * could put the error_at thing here but fallthrough (from failing $crate
-       * condition) isn't completely obvious if it is. */
-
-      // test prevent error
-      return Parse::Error::PathIdentSegment::make_invalid_token ();
-    }
-  rust_unreachable ();
-}
-
-// Parses an AttrInput AST node (polymorphic, as AttrInput is abstract)
-template <typename ManagedTokenSource>
-tl::expected<std::unique_ptr<AST::AttrInput>, Parse::Error::AttrInput>
-Parser<ManagedTokenSource>::parse_attr_input ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case LEFT_PAREN:
-    case LEFT_SQUARE:
-    case LEFT_CURLY:
-      {
-	auto dtoken_tree = parse_delim_token_tree ();
-	if (!dtoken_tree)
-	  return Parse::Error::AttrInput::make_bad_token_tree ();
-
-	// must be a delimited token tree, so parse that
-	std::unique_ptr<AST::AttrInput> input_tree (
-	  new AST::DelimTokenTree (dtoken_tree.value ()));
-
-	return tl::expected<std::unique_ptr<AST::AttrInput>,
-			    Parse::Error::AttrInput>{std::move (input_tree)};
-      }
-    case EQUAL:
-      {
-	// = LiteralExpr
-	lexer.skip_token ();
-
-	t = lexer.peek_token ();
-
-	// attempt to parse macro
-	// TODO: macros may/may not be allowed in attributes
-	// this is needed for "#[doc = include_str!(...)]"
-	if (Parse::Utils::is_simple_path_segment (t->get_id ()))
-	  {
-	    std::unique_ptr<AST::MacroInvocation> invoke
-	      = parse_macro_invocation ({});
-
-	    if (!invoke)
-	      return Parse::Error::AttrInput::make_bad_macro_invocation ();
-
-	    return std::unique_ptr<AST::AttrInput> (
-	      new AST::AttrInputMacro (std::move (invoke)));
-	  }
-
-	/* Ensure token is a "literal expression" (literally only a literal
-	 * token of any type) */
-	if (!t->is_literal ())
-	  {
-	    Error error (
-	      t->get_locus (),
-	      "unknown token %qs in attribute body - literal expected",
-	      t->get_token_description ());
-	    add_error (std::move (error));
-
-	    skip_after_end_attribute ();
-	    return Parse::Error::AttrInput::make_malformed ();
-	  }
-
-	AST::Literal::LitType lit_type = AST::Literal::STRING;
-	// Crappy mapping of token type to literal type
-	switch (t->get_id ())
-	  {
-	  case INT_LITERAL:
-	    lit_type = AST::Literal::INT;
-	    break;
-	  case FLOAT_LITERAL:
-	    lit_type = AST::Literal::FLOAT;
-	    break;
-	  case CHAR_LITERAL:
-	    lit_type = AST::Literal::CHAR;
-	    break;
-	  case BYTE_CHAR_LITERAL:
-	    lit_type = AST::Literal::BYTE;
-	    break;
-	  case BYTE_STRING_LITERAL:
-	    lit_type = AST::Literal::BYTE_STRING;
-	    break;
-	  case RAW_STRING_LITERAL:
-	    lit_type = AST::Literal::RAW_STRING;
-	    break;
-	  case STRING_LITERAL:
-	  default:
-	    lit_type = AST::Literal::STRING;
-	    break; // TODO: raw string? don't eliminate it from lexer?
-	  }
-
-	// create actual LiteralExpr
-	AST::LiteralExpr lit_expr (t->get_str (), lit_type, t->get_type_hint (),
-				   {}, t->get_locus ());
-	lexer.skip_token ();
-
-	std::unique_ptr<AST::AttrInput> attr_input_lit (
-	  new AST::AttrInputLiteral (std::move (lit_expr)));
-
-	// do checks or whatever? none required, really
-
-	// FIXME: shouldn't a skip token be required here?
-
-	return tl::expected<std::unique_ptr<AST::AttrInput>,
-			    Parse::Error::AttrInput>{
-	  std::move (attr_input_lit)};
-      }
-      break;
-    case RIGHT_PAREN:
-    case RIGHT_SQUARE:
-    case RIGHT_CURLY:
-    case END_OF_FILE:
-      // means AttrInput is missing, which is allowed
-      return Parse::Error::AttrInput::make_missing_attrinput ();
-    default:
-      add_error (
-	Error (t->get_locus (),
-	       "unknown token %qs in attribute body - attribute input or "
-	       "none expected",
-	       t->get_token_description ()));
-
-      skip_after_end_attribute ();
-      return Parse::Error::AttrInput::make_malformed ();
-    }
-  rust_unreachable ();
-  // TODO: find out how to stop gcc error on "no return value"
-}
-
-// Parses a delimited token tree
-template <typename ManagedTokenSource>
-tl::expected<AST::DelimTokenTree, Parse::Error::DelimTokenTree>
-Parser<ManagedTokenSource>::parse_delim_token_tree ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-  lexer.skip_token ();
-  location_t initial_loc = t->get_locus ();
-
-  // save delim type to ensure it is reused later
-  AST::DelimType delim_type = AST::PARENS;
-
-  // Map tokens to DelimType
-  switch (t->get_id ())
-    {
-    case LEFT_PAREN:
-      delim_type = AST::PARENS;
-      break;
-    case LEFT_SQUARE:
-      delim_type = AST::SQUARE;
-      break;
-    case LEFT_CURLY:
-      delim_type = AST::CURLY;
-      break;
-    default:
-      add_error (Error (t->get_locus (),
-			"unexpected token %qs - expecting delimiters (for a "
-			"delimited token tree)",
-			t->get_token_description ()));
-
-      return Parse::Error::DelimTokenTree::make_expected_delimiter ();
-    }
-
-  // parse actual token tree vector - 0 or more
-  std::vector<std::unique_ptr<AST::TokenTree>> token_trees_in_tree;
-  auto delim_open
-    = std::unique_ptr<AST::Token> (new AST::Token (std::move (t)));
-  token_trees_in_tree.push_back (std::move (delim_open));
-
-  // repeat loop until finding the matching delimiter
-  t = lexer.peek_token ();
-  while (!Parse::Utils::token_id_matches_delims (t->get_id (), delim_type)
-	 && t->get_id () != END_OF_FILE)
-    {
-      auto tok_tree = parse_token_tree ();
-      if (!tok_tree)
-	return Parse::Error::DelimTokenTree::make_invalid_token_tree ();
-
-      token_trees_in_tree.push_back (std::move (tok_tree.value ()));
-
-      // lexer.skip_token();
-      t = lexer.peek_token ();
-    }
-  auto delim_close
-    = std::unique_ptr<AST::Token> (new AST::Token (std::move (t)));
-  token_trees_in_tree.push_back (std::move (delim_close));
-
-  AST::DelimTokenTree token_tree (delim_type, std::move (token_trees_in_tree),
-				  initial_loc);
-
-  // parse end delimiters
-  t = lexer.peek_token ();
-
-  if (Parse::Utils::token_id_matches_delims (t->get_id (), delim_type))
-    {
-      // tokens match opening delimiter, so skip.
-      lexer.skip_token ();
-      return token_tree;
-    }
-  else
-    {
-      // tokens don't match opening delimiters, so produce error
-      Error error (t->get_locus (),
-		   "unexpected token %qs - expecting closing delimiter %qs "
-		   "(for a delimited token tree)",
-		   t->get_token_description (),
-		   (delim_type == AST::PARENS
-		      ? ")"
-		      : (delim_type == AST::SQUARE ? "]" : "}")));
-      add_error (std::move (error));
-
-      return Parse::Error::DelimTokenTree::make_mismatched_delimiters ();
-    }
-}
-
 // Parses an identifier/keyword as a Token
 template <typename ManagedTokenSource>
 tl::expected<std::unique_ptr<AST::Token>, Parse::Error::Token>
@@ -925,50 +328,6 @@ Parser<ManagedTokenSource>::parse_identifier_or_keyword_token ()
     {
       add_error (Error (t->get_locus (), "expected keyword or identifier"));
       return Parse::Error::Token::make_malformed ();
-    }
-}
-
-/* Parses a TokenTree syntactical production. This is either a delimited token
- * tree or a non-delimiter token. */
-template <typename ManagedTokenSource>
-tl::expected<std::unique_ptr<AST::TokenTree>, Parse::Error::TokenTree>
-Parser<ManagedTokenSource>::parse_token_tree ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-
-  switch (t->get_id ())
-    {
-    case LEFT_PAREN:
-    case LEFT_SQUARE:
-    case LEFT_CURLY:
-      {
-	// Parse delimited token tree
-	auto delim_token_tree = parse_delim_token_tree ();
-	if (!delim_token_tree)
-	  return Parse::Error::TokenTree::
-	    make_malformed_delimited_token_tree ();
-
-	// TODO: use move rather than copy constructor
-	return std::unique_ptr<AST::DelimTokenTree> (
-	  new AST::DelimTokenTree (delim_token_tree.value ()));
-      }
-    case RIGHT_PAREN:
-    case RIGHT_SQUARE:
-    case RIGHT_CURLY:
-      // error - should not be called when this a token
-      add_error (Error (t->get_locus (), "unexpected closing delimiter %qs",
-			t->get_token_description ()));
-
-      add_error (Error (Error::Kind::Hint, t->get_locus (),
-			"token tree requires either paired delimiters or "
-			"non-delimiter tokens"));
-
-      lexer.skip_token ();
-      return Parse::Error::TokenTree::make_malformed ();
-    default:
-      // parse token itself as TokenTree
-      lexer.skip_token ();
-      return std::unique_ptr<AST::Token> (new AST::Token (std::move (t)));
     }
 }
 
@@ -1108,112 +467,6 @@ Parser<ManagedTokenSource>::parse_item (bool called_from_statement)
       return Parse::Error::Item::make_malformed ();
       break;
     }
-}
-
-// Parses a contiguous block of outer attributes.
-template <typename ManagedTokenSource>
-AST::AttrVec
-Parser<ManagedTokenSource>::parse_outer_attributes ()
-{
-  AST::AttrVec outer_attributes;
-
-  auto has_valid_attribute_prefix = [&] () {
-    auto id = lexer.peek_token ()->get_id ();
-    /* We allow inner attributes `#!` and catch the error later  */
-    return id == HASH || id == OUTER_DOC_COMMENT || id == INNER_DOC_COMMENT;
-  };
-
-  while (has_valid_attribute_prefix ()) /* For error handling.  */
-    {
-      auto outer_attr = parse_outer_attribute ();
-
-      /* Ensure only valid outer attributes are added to the outer_attributes
-       * list */
-      if (outer_attr)
-	{
-	  outer_attributes.push_back (std::move (outer_attr.value ()));
-	}
-      else
-	{
-	  /* If no more valid outer attributes, break out of loop (only
-	   * contiguous outer attributes parsed). */
-	  break;
-	}
-    }
-
-  outer_attributes.shrink_to_fit ();
-  return outer_attributes;
-
-  /* TODO: this shares basically all code with parse_inner_attributes except
-   * function call - find way of making it more modular? function pointer? */
-}
-
-// Parse a single outer attribute.
-template <typename ManagedTokenSource>
-tl::expected<AST::Attribute, Parse::Error::Attribute>
-Parser<ManagedTokenSource>::parse_outer_attribute ()
-{
-  if (lexer.peek_token ()->get_id () == OUTER_DOC_COMMENT)
-    {
-      auto values = parse_doc_comment ();
-      auto path = std::move (std::get<0> (values));
-      auto input = std::move (std::get<1> (values));
-      auto loc = std::get<2> (values);
-      return AST::Attribute (std::move (path), std::move (input), loc, false);
-    }
-
-  if (lexer.peek_token ()->get_id () == INNER_DOC_COMMENT)
-    {
-      Error error (
-	lexer.peek_token ()->get_locus (), ErrorCode::E0753,
-	"expected outer doc comment, inner doc (%<//!%> or %</*!%>) only "
-	"allowed at start of item "
-	"and before any outer attribute or doc (%<#[%>, %<///%> or %</**%>)");
-      add_error (std::move (error));
-      lexer.skip_token ();
-      return Parse::Error::Attribute::make_unexpected_inner ();
-    }
-
-  /* OuterAttribute -> '#' '[' Attr ']' */
-
-  if (lexer.peek_token ()->get_id () != HASH)
-    return Parse::Error::Attribute::make_malformed ();
-
-  lexer.skip_token ();
-
-  TokenId id = lexer.peek_token ()->get_id ();
-  if (id != LEFT_SQUARE)
-    {
-      if (id == EXCLAM)
-	{
-	  // this is inner attribute syntax, so throw error
-	  // inner attributes were either already parsed or not allowed here.
-	  Error error (
-	    lexer.peek_token ()->get_locus (),
-	    "token %<!%> found, indicating inner attribute definition. Inner "
-	    "attributes are not possible at this location");
-	  add_error (std::move (error));
-	}
-      return Parse::Error::Attribute::make_unexpected_inner ();
-    }
-
-  lexer.skip_token ();
-
-  auto body_res = parse_attribute_body ();
-  if (!body_res)
-    return Parse::Error::Attribute::make_malformed_body ();
-  auto body = std::move (body_res.value ());
-
-  auto actual_attribute
-    = AST::Attribute (std::move (body.path), std::move (body.input), body.locus,
-		      false);
-
-  if (lexer.peek_token ()->get_id () != RIGHT_SQUARE)
-    return Parse::Error::Attribute::make_malformed ();
-
-  lexer.skip_token ();
-
-  return actual_attribute;
 }
 
 // Parses a VisItem (item that can have non-default visibility).
@@ -1726,561 +979,6 @@ Parser<ManagedTokenSource>::parse_decl_macro_def (AST::Visibility vis,
 			t->get_token_description ()));
       return nullptr;
     }
-}
-
-// Parses a semi-coloned (except for full block) macro invocation item.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MacroInvocation>
-Parser<ManagedTokenSource>::parse_macro_invocation_semi (
-  AST::AttrVec outer_attrs)
-{
-  location_t macro_locus = lexer.peek_token ()->get_locus ();
-  auto path = parse_simple_path ();
-  if (!path)
-    return nullptr;
-
-  if (!skip_token (EXCLAM))
-    {
-      // skip after somewhere?
-      return nullptr;
-    }
-
-  // save delim type to ensure it is reused later
-  AST::DelimType delim_type = AST::PARENS;
-
-  // Map tokens to DelimType
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case LEFT_PAREN:
-      delim_type = AST::PARENS;
-      break;
-    case LEFT_SQUARE:
-      delim_type = AST::SQUARE;
-      break;
-    case LEFT_CURLY:
-      delim_type = AST::CURLY;
-      break;
-    default:
-      add_error (Error (t->get_locus (),
-			"unexpected token %qs - expecting delimiters (for a "
-			"macro invocation semi body)",
-			t->get_token_description ()));
-
-      return nullptr;
-    }
-  location_t tok_tree_locus = t->get_locus ();
-  lexer.skip_token ();
-
-  // parse actual token trees
-  std::vector<std::unique_ptr<AST::TokenTree>> token_trees;
-  auto delim_open
-    = std::unique_ptr<AST::Token> (new AST::Token (std::move (t)));
-  token_trees.push_back (std::move (delim_open));
-
-  t = lexer.peek_token ();
-  // parse token trees until the initial delimiter token is found again
-  while (!Parse::Utils::token_id_matches_delims (t->get_id (), delim_type)
-	 && t->get_id () != END_OF_FILE)
-    {
-      auto tree = parse_token_tree ();
-      if (!tree)
-	return nullptr;
-
-      token_trees.push_back (std::move (tree.value ()));
-
-      t = lexer.peek_token ();
-    }
-  auto delim_close
-    = std::unique_ptr<AST::Token> (new AST::Token (std::move (t)));
-  token_trees.push_back (std::move (delim_close));
-
-  AST::DelimTokenTree delim_tok_tree (delim_type, std::move (token_trees),
-				      tok_tree_locus);
-  AST::MacroInvocData invoc_data (std::move (path.value ()),
-				  std::move (delim_tok_tree));
-
-  // parse end delimiters
-  t = lexer.peek_token ();
-  if (Parse::Utils::token_id_matches_delims (t->get_id (), delim_type))
-    {
-      // tokens match opening delimiter, so skip.
-      lexer.skip_token ();
-
-      if (delim_type != AST::CURLY)
-	{
-	  // skip semicolon at end of non-curly macro invocation semis
-	  if (!skip_token (SEMICOLON))
-	    {
-	      // as this is the end, allow recovery (probably) - may change
-
-	      return AST::MacroInvocation::Regular (std::move (invoc_data),
-						    std::move (outer_attrs),
-						    macro_locus, true);
-	    }
-	}
-
-      // DEBUG:
-      rust_debug ("skipped token is '%s', next token (current peek) is '%s'",
-		  t->get_token_description (),
-		  lexer.peek_token ()->get_token_description ());
-
-      return AST::MacroInvocation::Regular (std::move (invoc_data),
-					    std::move (outer_attrs),
-					    macro_locus, true);
-    }
-  else
-    {
-      // tokens don't match opening delimiters, so produce error
-      Error error (t->get_locus (),
-		   "unexpected token %qs - expecting closing delimiter %qs "
-		   "(for a macro invocation semi)",
-		   t->get_token_description (),
-		   (delim_type == AST::PARENS
-		      ? ")"
-		      : (delim_type == AST::SQUARE ? "]" : "}")));
-      add_error (std::move (error));
-
-      /* return empty macro invocation despite possibly parsing mostly valid one
-       * - TODO is this a good idea? */
-      return nullptr;
-    }
-}
-
-// Parses a non-semicoloned macro invocation (i.e. as pattern or expression).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MacroInvocation>
-Parser<ManagedTokenSource>::parse_macro_invocation (AST::AttrVec outer_attrs)
-{
-  // parse macro path
-  auto macro_path = parse_simple_path ();
-  if (!macro_path)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse macro invocation path");
-      add_error (std::move (error));
-
-      // skip?
-      return nullptr;
-    }
-
-  if (!skip_token (EXCLAM))
-    {
-      // skip after somewhere?
-      return nullptr;
-    }
-
-  // parse internal delim token tree
-  auto delim_tok_tree = parse_delim_token_tree ();
-  if (!delim_tok_tree)
-    return nullptr;
-
-  location_t macro_locus = macro_path->get_locus ();
-
-  return AST::MacroInvocation::Regular (
-    AST::MacroInvocData (std::move (macro_path.value ()),
-			 std::move (delim_tok_tree.value ())),
-    std::move (outer_attrs), macro_locus);
-}
-
-// Parses a macro rule definition - does not parse semicolons.
-template <typename ManagedTokenSource>
-AST::MacroRule
-Parser<ManagedTokenSource>::parse_macro_rule ()
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-
-  // parse macro matcher
-  AST::MacroMatcher matcher = parse_macro_matcher ();
-
-  if (matcher.is_error ())
-    return AST::MacroRule::create_error (locus);
-
-  if (!skip_token (MATCH_ARROW))
-    {
-      // skip after somewhere?
-      return AST::MacroRule::create_error (locus);
-    }
-
-  // parse transcriber (this is just a delim token tree)
-  location_t token_tree_loc = lexer.peek_token ()->get_locus ();
-  auto delim_token_tree = parse_delim_token_tree ();
-  if (!delim_token_tree)
-    return AST::MacroRule::create_error (token_tree_loc);
-
-  AST::MacroTranscriber transcriber (delim_token_tree.value (), token_tree_loc);
-
-  return AST::MacroRule (std::move (matcher), std::move (transcriber), locus);
-}
-
-// Parses a macro matcher (part of a macro rule definition).
-template <typename ManagedTokenSource>
-AST::MacroMatcher
-Parser<ManagedTokenSource>::parse_macro_matcher ()
-{
-  // save delim type to ensure it is reused later
-  AST::DelimType delim_type = AST::PARENS;
-
-  // DEBUG
-  rust_debug ("begun parsing macro matcher");
-
-  // Map tokens to DelimType
-  const_TokenPtr t = lexer.peek_token ();
-  location_t locus = t->get_locus ();
-  switch (t->get_id ())
-    {
-    case LEFT_PAREN:
-      delim_type = AST::PARENS;
-      break;
-    case LEFT_SQUARE:
-      delim_type = AST::SQUARE;
-      break;
-    case LEFT_CURLY:
-      delim_type = AST::CURLY;
-      break;
-    default:
-      add_error (Error (
-	t->get_locus (),
-	"unexpected token %qs - expecting delimiters (for a macro matcher)",
-	t->get_token_description ()));
-
-      return AST::MacroMatcher::create_error (t->get_locus ());
-    }
-  lexer.skip_token ();
-
-  // parse actual macro matches
-  std::vector<std::unique_ptr<AST::MacroMatch>> matches;
-  // Set of possible preceding macro matches to make sure follow-set
-  // restrictions are respected.
-  // TODO: Consider using std::reference_wrapper instead of raw pointers?
-  std::vector<const AST::MacroMatch *> last_matches;
-
-  t = lexer.peek_token ();
-  // parse token trees until the initial delimiter token is found again
-  while (!Parse::Utils::token_id_matches_delims (t->get_id (), delim_type))
-    {
-      std::unique_ptr<AST::MacroMatch> match = parse_macro_match ();
-
-      if (match == nullptr)
-	{
-	  Error error (
-	    t->get_locus (),
-	    "failed to parse macro match for macro matcher - found %qs",
-	    t->get_token_description ());
-	  add_error (std::move (error));
-
-	  return AST::MacroMatcher::create_error (t->get_locus ());
-	}
-
-      if (matches.size () > 0)
-	{
-	  const auto *last_match = matches.back ().get ();
-
-	  // We want to check if we are dealing with a zeroable repetition
-	  bool zeroable = false;
-	  if (last_match->get_macro_match_type ()
-	      == AST::MacroMatch::MacroMatchType::Repetition)
-	    {
-	      auto repetition
-		= static_cast<const AST::MacroMatchRepetition *> (last_match);
-
-	      if (repetition->get_op ()
-		  != AST::MacroMatchRepetition::MacroRepOp::ONE_OR_MORE)
-		zeroable = true;
-	    }
-
-	  if (!zeroable)
-	    last_matches.clear ();
-
-	  last_matches.emplace_back (last_match);
-
-	  for (auto last : last_matches)
-	    if (!is_match_compatible (*last, *match))
-	      return AST::MacroMatcher::create_error (
-		match->get_match_locus ());
-	}
-
-      matches.push_back (std::move (match));
-
-      // DEBUG
-      rust_debug ("pushed back a match in macro matcher");
-
-      t = lexer.peek_token ();
-    }
-
-  // parse end delimiters
-  t = lexer.peek_token ();
-  if (Parse::Utils::token_id_matches_delims (t->get_id (), delim_type))
-    {
-      // tokens match opening delimiter, so skip.
-      lexer.skip_token ();
-
-      return AST::MacroMatcher (delim_type, std::move (matches), locus);
-    }
-  else
-    {
-      // tokens don't match opening delimiters, so produce error
-      Error error (t->get_locus (),
-		   "unexpected token %qs - expecting closing delimiter %qs "
-		   "(for a macro matcher)",
-		   t->get_token_description (),
-		   (delim_type == AST::PARENS
-		      ? ")"
-		      : (delim_type == AST::SQUARE ? "]" : "}")));
-      add_error (std::move (error));
-
-      /* return error macro matcher despite possibly parsing mostly correct one?
-       * TODO is this the best idea? */
-      return AST::MacroMatcher::create_error (t->get_locus ());
-    }
-}
-
-// Parses a macro match (syntax match inside a matcher in a macro rule).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MacroMatch>
-Parser<ManagedTokenSource>::parse_macro_match ()
-{
-  // branch based on token available
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case LEFT_PAREN:
-    case LEFT_SQUARE:
-    case LEFT_CURLY:
-      {
-	// must be macro matcher as delimited
-	AST::MacroMatcher matcher = parse_macro_matcher ();
-	if (matcher.is_error ())
-	  {
-	    Error error (lexer.peek_token ()->get_locus (),
-			 "failed to parse macro matcher in macro match");
-	    add_error (std::move (error));
-
-	    return nullptr;
-	  }
-	return std::unique_ptr<AST::MacroMatcher> (
-	  new AST::MacroMatcher (std::move (matcher)));
-      }
-    case DOLLAR_SIGN:
-      {
-	// have to do more lookahead to determine if fragment or repetition
-	const_TokenPtr t2 = lexer.peek_token (1);
-	switch (t2->get_id ())
-	  {
-	  case IDENTIFIER:
-	  case UNDERSCORE:
-	    // macro fragment
-	    return parse_macro_match_fragment ();
-	  case LEFT_PAREN:
-	    // macro repetition
-	    return parse_macro_match_repetition ();
-	  default:
-	    if (token_id_is_keyword (t2->get_id ()) && t2->get_id () != CRATE)
-	      {
-		// keyword as macro fragment
-		return parse_macro_match_fragment ();
-	      }
-	    else
-	      {
-		// error: unrecognised
-		add_error (Error (
-		  t2->get_locus (),
-		  "unrecognised token combination %<$%s%> at start of "
-		  "macro match - did you mean %<$identifier%> or %<$(%>?",
-		  t2->get_token_description ()));
-
-		// skip somewhere?
-		return nullptr;
-	      }
-	  }
-      }
-    case RIGHT_PAREN:
-    case RIGHT_SQUARE:
-    case RIGHT_CURLY:
-      // not allowed
-      add_error (Error (
-	t->get_locus (),
-	"closing delimiters like %qs are not allowed at the start of a macro "
-	"match",
-	t->get_token_description ()));
-
-      // skip somewhere?
-      return nullptr;
-    default:
-      // just the token
-      lexer.skip_token ();
-      return std::unique_ptr<AST::Token> (new AST::Token (std::move (t)));
-    }
-}
-
-// Parses a fragment macro match.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MacroMatchFragment>
-Parser<ManagedTokenSource>::parse_macro_match_fragment ()
-{
-  location_t fragment_locus = lexer.peek_token ()->get_locus ();
-  skip_token (DOLLAR_SIGN);
-
-  Identifier ident;
-  auto identifier = lexer.peek_token ();
-  if (identifier->get_id () == UNDERSCORE)
-    ident = {Values::Keywords::UNDERSCORE, identifier->get_locus ()};
-  else
-    ident = {identifier};
-
-  if (ident.empty ())
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "missing identifier in macro match fragment");
-      add_error (std::move (error));
-
-      return nullptr;
-    }
-  skip_token (identifier->get_id ());
-
-  if (!skip_token (COLON))
-    {
-      // skip after somewhere?
-      return nullptr;
-    }
-
-  // get MacroFragSpec for macro
-  const_TokenPtr t = expect_token (IDENTIFIER);
-  if (t == nullptr)
-    return nullptr;
-
-  AST::MacroFragSpec frag
-    = AST::MacroFragSpec::get_frag_spec_from_str (t->get_str ());
-  if (frag.is_error ())
-    {
-      Error error (t->get_locus (),
-		   "invalid fragment specifier %qs in fragment macro match",
-		   t->get_str ().c_str ());
-      add_error (std::move (error));
-
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::MacroMatchFragment> (
-    new AST::MacroMatchFragment (std::move (ident), frag, fragment_locus));
-}
-
-// Parses a repetition macro match.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MacroMatchRepetition>
-Parser<ManagedTokenSource>::parse_macro_match_repetition ()
-{
-  skip_token (DOLLAR_SIGN);
-  skip_token (LEFT_PAREN);
-
-  std::vector<std::unique_ptr<AST::MacroMatch>> matches;
-
-  // parse required first macro match
-  std::unique_ptr<AST::MacroMatch> initial_match = parse_macro_match ();
-  if (initial_match == nullptr)
-    {
-      Error error (
-	lexer.peek_token ()->get_locus (),
-	"could not parse required first macro match in macro match repetition");
-      add_error (std::move (error));
-
-      // skip after somewhere?
-      return nullptr;
-    }
-  matches.push_back (std::move (initial_match));
-
-  // parse optional later macro matches
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () != RIGHT_PAREN)
-    {
-      std::unique_ptr<AST::MacroMatch> match = parse_macro_match ();
-
-      if (match == nullptr)
-	{
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse macro match in macro match repetition");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-
-      matches.push_back (std::move (match));
-
-      t = lexer.peek_token ();
-    }
-
-  if (!skip_token (RIGHT_PAREN))
-    {
-      // skip after somewhere?
-      return nullptr;
-    }
-
-  t = lexer.peek_token ();
-  // see if separator token exists
-  std::unique_ptr<AST::Token> separator = nullptr;
-  switch (t->get_id ())
-    {
-    // repetition operators
-    case ASTERISK:
-    case PLUS:
-    case QUESTION_MARK:
-    // delimiters
-    case LEFT_PAREN:
-    case LEFT_CURLY:
-    case LEFT_SQUARE:
-    case RIGHT_PAREN:
-    case RIGHT_CURLY:
-    case RIGHT_SQUARE:
-      // separator does not exist, so still null and don't skip token
-      break;
-    default:
-      // separator does exist
-      separator = std::unique_ptr<AST::Token> (new AST::Token (std::move (t)));
-      lexer.skip_token ();
-      break;
-    }
-
-  // parse repetition operator
-  t = lexer.peek_token ();
-  AST::MacroMatchRepetition::MacroRepOp op = AST::MacroMatchRepetition::NONE;
-  switch (t->get_id ())
-    {
-    case ASTERISK:
-      op = AST::MacroMatchRepetition::ANY;
-      lexer.skip_token ();
-      break;
-    case PLUS:
-      op = AST::MacroMatchRepetition::ONE_OR_MORE;
-      lexer.skip_token ();
-      break;
-    case QUESTION_MARK:
-      op = AST::MacroMatchRepetition::ZERO_OR_ONE;
-      lexer.skip_token ();
-
-      if (separator != nullptr)
-	{
-	  add_error (
-	    Error (separator->get_locus (),
-		   "the %<?%> macro repetition operator does not take a "
-		   "separator"));
-	  separator = nullptr;
-	}
-
-      break;
-    default:
-      add_error (
-	Error (t->get_locus (),
-	       "expected macro repetition operator (%<*%>, %<+%>, or %<?%>) in "
-	       "macro match - found %qs",
-	       t->get_token_description ()));
-
-      // skip after somewhere?
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::MacroMatchRepetition> (
-    new AST::MacroMatchRepetition (std::move (matches), op,
-				   std::move (separator), t->get_locus ()));
 }
 
 /* Parses a visibility syntactical production (i.e. creating a non-default
@@ -6204,63 +4902,6 @@ Parser<ManagedTokenSource>::parse_let_stmt (AST::AttrVec outer_attrs,
 		      std::move (else_expr), std::move (outer_attrs), locus));
 }
 
-// Parses a type path.
-template <typename ManagedTokenSource>
-AST::TypePath
-Parser<ManagedTokenSource>::parse_type_path ()
-{
-  bool has_opening_scope_resolution = false;
-  location_t locus = lexer.peek_token ()->get_locus ();
-  if (lexer.peek_token ()->get_id () == SCOPE_RESOLUTION)
-    {
-      has_opening_scope_resolution = true;
-      lexer.skip_token ();
-    }
-
-  // create segment vector
-  std::vector<std::unique_ptr<AST::TypePathSegment>> segments;
-
-  // parse required initial segment
-  std::unique_ptr<AST::TypePathSegment> initial_segment
-    = parse_type_path_segment ();
-  if (initial_segment == nullptr)
-    {
-      // skip after somewhere?
-      // don't necessarily throw error but yeah
-      return AST::TypePath::create_error ();
-    }
-  segments.push_back (std::move (initial_segment));
-
-  // parse optional segments (as long as scope resolution operator exists)
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () == SCOPE_RESOLUTION)
-    {
-      // skip scope resolution operator
-      lexer.skip_token ();
-
-      // parse the actual segment - it is an error if it doesn't exist now
-      std::unique_ptr<AST::TypePathSegment> segment
-	= parse_type_path_segment ();
-      if (segment == nullptr)
-	{
-	  // skip after somewhere?
-	  Error error (t->get_locus (), "could not parse type path segment");
-	  add_error (std::move (error));
-
-	  return AST::TypePath::create_error ();
-	}
-
-      segments.push_back (std::move (segment));
-
-      t = lexer.peek_token ();
-    }
-
-  segments.shrink_to_fit ();
-
-  return AST::TypePath (std::move (segments), locus,
-			has_opening_scope_resolution);
-}
-
 template <typename ManagedTokenSource>
 tl::optional<AST::GenericArg>
 Parser<ManagedTokenSource>::parse_generic_arg ()
@@ -6487,446 +5128,6 @@ Parser<ManagedTokenSource>::parse_generic_args_binding ()
 
   return AST::GenericArgsBinding (std::move (ident), std::move (type),
 				  ident_tok->get_locus ());
-}
-
-/* Parses a single type path segment (not including opening scope resolution,
- * but includes any internal ones). Includes generic args or type path
- * functions too. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::TypePathSegment>
-Parser<ManagedTokenSource>::parse_type_path_segment ()
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-  // parse ident segment part
-  auto ident_segment_res = parse_path_ident_segment ();
-  if (!ident_segment_res)
-    {
-      // not necessarily an error
-      return nullptr;
-    }
-  auto ident_segment = ident_segment_res.value ();
-
-  /* lookahead to determine if variants exist - only consume scope resolution
-   * then */
-  bool has_separating_scope_resolution = false;
-  const_TokenPtr next = lexer.peek_token (1);
-  if (lexer.peek_token ()->get_id () == SCOPE_RESOLUTION
-      && (next->get_id () == LEFT_ANGLE || next->get_id () == LEFT_PAREN))
-    {
-      has_separating_scope_resolution = true;
-      lexer.skip_token ();
-    }
-
-  // branch into variants on next token
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case LEFT_SHIFT:
-    case LEFT_ANGLE:
-      {
-	// parse generic args
-	AST::GenericArgs generic_args = parse_path_generic_args ();
-
-	return std::unique_ptr<AST::TypePathSegmentGeneric> (
-	  new AST::TypePathSegmentGeneric (std::move (ident_segment),
-					   has_separating_scope_resolution,
-					   std::move (generic_args), locus));
-      }
-    case LEFT_PAREN:
-      {
-	// parse type path function
-	AST::TypePathFunction type_path_function
-	  = parse_type_path_function (locus);
-
-	if (type_path_function.is_error ())
-	  {
-	    // skip after somewhere?
-	    return nullptr;
-	  }
-
-	return std::unique_ptr<AST::TypePathSegmentFunction> (
-	  new AST::TypePathSegmentFunction (std::move (ident_segment),
-					    has_separating_scope_resolution,
-					    std::move (type_path_function),
-					    locus));
-      }
-    default:
-      // neither of them
-      return std::unique_ptr<AST::TypePathSegment> (
-	new AST::TypePathSegment (std::move (ident_segment),
-				  has_separating_scope_resolution, locus));
-    }
-  rust_unreachable ();
-}
-
-// Parses a function call representation inside a type path.
-template <typename ManagedTokenSource>
-AST::TypePathFunction
-Parser<ManagedTokenSource>::parse_type_path_function (location_t id_location)
-{
-  if (!skip_token (LEFT_PAREN))
-    {
-      // skip somewhere?
-      return AST::TypePathFunction::create_error ();
-    }
-
-  // parse function inputs
-  std::vector<std::unique_ptr<AST::Type>> inputs;
-
-  while (lexer.peek_token ()->get_id () != RIGHT_PAREN)
-    {
-      std::unique_ptr<AST::Type> type = parse_type ();
-      if (type == nullptr)
-	{
-	  /* this is an error as there should've been a ')' there if there
-	   * wasn't a type */
-	  Error error (
-	    lexer.peek_token ()->get_locus (),
-	    "failed to parse type in parameters of type path function");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return AST::TypePathFunction::create_error ();
-	}
-
-      inputs.push_back (std::move (type));
-
-      // skip commas, including trailing commas
-      if (lexer.peek_token ()->get_id () != COMMA)
-	break;
-
-      lexer.skip_token ();
-    }
-
-  if (!skip_token (RIGHT_PAREN))
-    {
-      // skip somewhere?
-      return AST::TypePathFunction::create_error ();
-    }
-
-  // parse optional return type
-  std::unique_ptr<AST::Type> return_type = parse_function_return_type ();
-
-  inputs.shrink_to_fit ();
-  return AST::TypePathFunction (std::move (inputs), id_location,
-				std::move (return_type));
-}
-
-// Parses a path inside an expression that allows generic arguments.
-template <typename ManagedTokenSource>
-AST::PathInExpression
-Parser<ManagedTokenSource>::parse_path_in_expression ()
-{
-  location_t locus = UNKNOWN_LOCATION;
-  bool has_opening_scope_resolution = false;
-  if (lexer.peek_token ()->get_id () == SCOPE_RESOLUTION)
-    {
-      has_opening_scope_resolution = true;
-
-      locus = lexer.peek_token ()->get_locus ();
-
-      lexer.skip_token ();
-    }
-
-  // create segment vector
-  std::vector<AST::PathExprSegment> segments;
-
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-    }
-
-  // parse required initial segment
-  AST::PathExprSegment initial_segment = parse_path_expr_segment ();
-  if (initial_segment.is_error ())
-    {
-      // skip after somewhere?
-      // don't necessarily throw error but yeah
-      return AST::PathInExpression::create_error ();
-    }
-  segments.push_back (std::move (initial_segment));
-
-  // parse optional segments (as long as scope resolution operator exists)
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () == SCOPE_RESOLUTION)
-    {
-      // skip scope resolution operator
-      lexer.skip_token ();
-
-      // parse the actual segment - it is an error if it doesn't exist now
-      AST::PathExprSegment segment = parse_path_expr_segment ();
-      if (segment.is_error ())
-	{
-	  // skip after somewhere?
-	  Error error (t->get_locus (),
-		       "could not parse path expression segment");
-	  add_error (std::move (error));
-
-	  return AST::PathInExpression::create_error ();
-	}
-
-      segments.push_back (std::move (segment));
-
-      t = lexer.peek_token ();
-    }
-
-  segments.shrink_to_fit ();
-
-  return AST::PathInExpression (std::move (segments), {}, locus,
-				has_opening_scope_resolution);
-}
-
-/* Parses a single path in expression path segment (including generic
- * arguments). */
-template <typename ManagedTokenSource>
-AST::PathExprSegment
-Parser<ManagedTokenSource>::parse_path_expr_segment ()
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-  // parse ident segment
-  auto ident_result = parse_path_ident_segment ();
-  if (!ident_result)
-    {
-      // not necessarily an error?
-      return AST::PathExprSegment::create_error ();
-    }
-  auto ident = ident_result.value ();
-
-  // parse generic args (and turbofish), if they exist
-  /* use lookahead to determine if they actually exist (don't want to
-   * accidently parse over next ident segment) */
-  if (lexer.peek_token ()->get_id () == SCOPE_RESOLUTION
-      && (lexer.peek_token (1)->get_id () == LEFT_ANGLE
-	  || lexer.peek_token (1)->get_id () == LEFT_SHIFT))
-    {
-      // skip scope resolution
-      lexer.skip_token ();
-
-      // Let parse_path_generic_args split "<<" tokens
-      AST::GenericArgs generic_args = parse_path_generic_args ();
-
-      return AST::PathExprSegment (std::move (ident), locus,
-				   std::move (generic_args));
-    }
-
-  // return a generic parameter-less expr segment if not found
-  return AST::PathExprSegment (std::move (ident), locus);
-}
-
-/* Parses a fully qualified path in expression (i.e. a pattern). FIXME does
- * not parse outer attrs. */
-template <typename ManagedTokenSource>
-AST::QualifiedPathInExpression
-Parser<ManagedTokenSource>::parse_qualified_path_in_expression (
-  location_t pratt_parsed_loc)
-{
-  /* Note: the Rust grammar is defined in such a way that it is impossible to
-   * determine whether a prospective qualified path is a
-   * QualifiedPathInExpression or QualifiedPathInType in all cases by the
-   * rules themselves (the only possible difference is a TypePathSegment with
-   * function, and lookahead to find this is too difficult). However, as this
-   * is a pattern and QualifiedPathInType is a type, I believe it that their
-   * construction will not be confused (due to rules regarding patterns vs
-   * types).
-   * As such, this function will not attempt to minimise errors created by
-   * their confusion. */
-
-  // parse the qualified path type (required)
-  AST::QualifiedPathType qual_path_type
-    = parse_qualified_path_type (pratt_parsed_loc);
-  if (qual_path_type.is_error ())
-    {
-      // TODO: should this create a parse error?
-      return AST::QualifiedPathInExpression::create_error ();
-    }
-  location_t locus = qual_path_type.get_locus ();
-
-  // parse path segments
-  std::vector<AST::PathExprSegment> segments;
-
-  // parse initial required segment
-  if (!expect_token (SCOPE_RESOLUTION))
-    {
-      // skip after somewhere?
-
-      return AST::QualifiedPathInExpression::create_error ();
-    }
-  AST::PathExprSegment initial_segment = parse_path_expr_segment ();
-  if (initial_segment.is_error ())
-    {
-      // skip after somewhere?
-      Error error (lexer.peek_token ()->get_locus (),
-		   "required initial path expression segment in "
-		   "qualified path in expression could not be parsed");
-      add_error (std::move (error));
-
-      return AST::QualifiedPathInExpression::create_error ();
-    }
-  segments.push_back (std::move (initial_segment));
-
-  // parse optional segments (as long as scope resolution operator exists)
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () == SCOPE_RESOLUTION)
-    {
-      // skip scope resolution operator
-      lexer.skip_token ();
-
-      // parse the actual segment - it is an error if it doesn't exist now
-      AST::PathExprSegment segment = parse_path_expr_segment ();
-      if (segment.is_error ())
-	{
-	  // skip after somewhere?
-	  Error error (t->get_locus (),
-		       "could not parse path expression segment in qualified "
-		       "path in expression");
-	  add_error (std::move (error));
-
-	  return AST::QualifiedPathInExpression::create_error ();
-	}
-
-      segments.push_back (std::move (segment));
-
-      t = lexer.peek_token ();
-    }
-
-  segments.shrink_to_fit ();
-
-  // FIXME: outer attr parsing
-  return AST::QualifiedPathInExpression (std::move (qual_path_type),
-					 std::move (segments), {}, locus);
-}
-
-// Parses the type syntactical construction at the start of a qualified path.
-template <typename ManagedTokenSource>
-AST::QualifiedPathType
-Parser<ManagedTokenSource>::parse_qualified_path_type (
-  location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  /* TODO: should this actually be error? is there anywhere where this could
-   * be valid? */
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-
-      if (lexer.peek_token ()->get_id () == LEFT_SHIFT)
-	lexer.split_current_token (LEFT_ANGLE, LEFT_ANGLE);
-
-      // skip after somewhere?
-      if (!skip_token (LEFT_ANGLE))
-	return AST::QualifiedPathType::create_error ();
-    }
-
-  // parse type (required)
-  std::unique_ptr<AST::Type> type = parse_type ();
-  if (type == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "could not parse type in qualified path type");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return AST::QualifiedPathType::create_error ();
-    }
-
-  // parse optional as clause
-  AST::TypePath as_type_path = AST::TypePath::create_error ();
-  if (lexer.peek_token ()->get_id () == AS)
-    {
-      lexer.skip_token ();
-
-      // parse type path, which is required now
-      as_type_path = parse_type_path ();
-      if (as_type_path.is_error ())
-	{
-	  Error error (
-	    lexer.peek_token ()->get_locus (),
-	    "could not parse type path in as clause in qualified path type");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return AST::QualifiedPathType::create_error ();
-	}
-    }
-
-  /* NOTE: should actually be a right-angle token, so
-   * skip_generics_right_angle shouldn't be required */
-  if (!skip_token (RIGHT_ANGLE))
-    {
-      // skip after somewhere?
-      return AST::QualifiedPathType::create_error ();
-    }
-
-  return AST::QualifiedPathType (std::move (type), locus,
-				 std::move (as_type_path));
-}
-
-// Parses a fully qualified path in type (i.e. a type).
-template <typename ManagedTokenSource>
-AST::QualifiedPathInType
-Parser<ManagedTokenSource>::parse_qualified_path_in_type ()
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-  // parse the qualified path type (required)
-  AST::QualifiedPathType qual_path_type = parse_qualified_path_type ();
-  if (qual_path_type.is_error ())
-    {
-      // TODO: should this create a parse error?
-      return AST::QualifiedPathInType::create_error ();
-    }
-
-  // parse initial required segment
-  if (!expect_token (SCOPE_RESOLUTION))
-    {
-      // skip after somewhere?
-
-      return AST::QualifiedPathInType::create_error ();
-    }
-  std::unique_ptr<AST::TypePathSegment> initial_segment
-    = parse_type_path_segment ();
-  if (initial_segment == nullptr)
-    {
-      // skip after somewhere?
-      Error error (lexer.peek_token ()->get_locus (),
-		   "required initial type path segment in qualified path in "
-		   "type could not be parsed");
-      add_error (std::move (error));
-
-      return AST::QualifiedPathInType::create_error ();
-    }
-
-  // parse optional segments (as long as scope resolution operator exists)
-  std::vector<std::unique_ptr<AST::TypePathSegment>> segments;
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () == SCOPE_RESOLUTION)
-    {
-      // skip scope resolution operator
-      lexer.skip_token ();
-
-      // parse the actual segment - it is an error if it doesn't exist now
-      std::unique_ptr<AST::TypePathSegment> segment
-	= parse_type_path_segment ();
-      if (segment == nullptr)
-	{
-	  // skip after somewhere?
-	  Error error (
-	    t->get_locus (),
-	    "could not parse type path segment in qualified path in type");
-	  add_error (std::move (error));
-
-	  return AST::QualifiedPathInType::create_error ();
-	}
-
-      segments.push_back (std::move (segment));
-
-      t = lexer.peek_token ();
-    }
-
-  segments.shrink_to_fit ();
-
-  return AST::QualifiedPathInType (std::move (qual_path_type),
-				   std::move (initial_segment),
-				   std::move (segments), locus);
 }
 
 // Parses a self param. Also handles self param not existing.
@@ -7176,482 +5377,6 @@ Parser<ManagedTokenSource>::parse_expr_stmt (AST::AttrVec outer_attrs,
     new AST::ExprStmt (std::move (expr), locus, has_semi));
 }
 
-// Parses a block expression, including the curly braces at start and end.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::BlockExpr>
-Parser<ManagedTokenSource>::parse_block_expr (
-  AST::AttrVec outer_attrs, tl::optional<AST::LoopLabel> label,
-  location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      if (!skip_token (LEFT_CURLY))
-	{
-	  skip_after_end_block ();
-	  return nullptr;
-	}
-    }
-
-  AST::AttrVec inner_attrs = parse_inner_attributes ();
-
-  // parse statements and expression
-  std::vector<std::unique_ptr<AST::Stmt>> stmts;
-  std::unique_ptr<AST::Expr> expr = nullptr;
-
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () != RIGHT_CURLY)
-    {
-      ExprOrStmt expr_or_stmt = parse_stmt_or_expr ();
-      if (expr_or_stmt.is_error ())
-	{
-	  skip_after_end_block ();
-	  return nullptr;
-	}
-
-      t = lexer.peek_token ();
-
-      if (expr_or_stmt.stmt != nullptr)
-	{
-	  stmts.push_back (std::move (expr_or_stmt.stmt));
-	}
-      else
-	{
-	  // assign to expression and end parsing inside
-	  expr = std::move (expr_or_stmt.expr);
-	  break;
-	}
-    }
-
-  location_t end_locus = t->get_locus ();
-
-  if (!skip_token (RIGHT_CURLY))
-    {
-      Error error (t->get_locus (),
-		   "error may be from having an expression (as opposed to "
-		   "statement) in the body of the function but not last");
-      add_error (std::move (error));
-
-      skip_after_end_block ();
-      return nullptr;
-    }
-
-  // grammar allows for empty block expressions
-
-  stmts.shrink_to_fit ();
-
-  return std::unique_ptr<AST::BlockExpr> (
-    new AST::BlockExpr (std::move (stmts), std::move (expr),
-			std::move (inner_attrs), std::move (outer_attrs),
-			std::move (label), locus, end_locus));
-}
-
-/* Parse an anonymous const expression. This can be a regular const expression
- * or an underscore for deferred const inference */
-template <typename ManagedTokenSource>
-tl::expected<AST::AnonConst, AnonConstError>
-Parser<ManagedTokenSource>::parse_anon_const ()
-{
-  auto current = lexer.peek_token ();
-  auto locus = current->get_locus ();
-
-  // Special case deferred inference constants
-  if (maybe_skip_token (UNDERSCORE))
-    return AST::AnonConst (locus);
-
-  auto expr = parse_expr ();
-
-  if (!expr)
-    return tl::make_unexpected (AnonConstError::InvalidSizeExpr);
-
-  return AST::AnonConst (std::move (expr), locus);
-}
-
-/* Parse a "const block", a block preceded by the `const` keyword whose
- * statements can be const evaluated and used in constant contexts */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ConstBlock>
-Parser<ManagedTokenSource>::parse_const_block_expr (AST::AttrVec outer_attrs,
-						    location_t locus)
-{
-  auto block = parse_block_expr ();
-
-  if (!block)
-    {
-      add_error (Error (locus, "failed to parse inner block in const block"));
-      skip_after_end_block ();
-
-      return nullptr;
-    }
-
-  auto block_locus = block->get_locus ();
-
-  return std::make_unique<AST::ConstBlock> (AST::AnonConst (std::move (block),
-							    block_locus),
-					    locus, std::move (outer_attrs));
-}
-
-/* Parses a "grouped" expression (expression in parentheses), used to control
- * precedence. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::GroupedExpr>
-Parser<ManagedTokenSource>::parse_grouped_expr (AST::AttrVec outer_attrs)
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-  skip_token (LEFT_PAREN);
-
-  AST::AttrVec inner_attrs = parse_inner_attributes ();
-
-  // parse required expr inside parentheses
-  std::unique_ptr<AST::Expr> expr_in_parens = parse_expr ();
-  if (expr_in_parens == nullptr)
-    {
-      // skip after somewhere?
-      // error?
-      return nullptr;
-    }
-
-  if (!skip_token (RIGHT_PAREN))
-    {
-      // skip after somewhere?
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::GroupedExpr> (
-    new AST::GroupedExpr (std::move (expr_in_parens), std::move (inner_attrs),
-			  std::move (outer_attrs), locus));
-}
-
-// Parses a closure expression (closure definition).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ClosureExpr>
-Parser<ManagedTokenSource>::parse_closure_expr (AST::AttrVec outer_attrs)
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-  // detect optional "move"
-  bool has_move = false;
-  if (lexer.peek_token ()->get_id () == MOVE)
-    {
-      lexer.skip_token ();
-      has_move = true;
-    }
-
-  // handle parameter list
-  std::vector<AST::ClosureParam> params;
-
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case OR:
-      // skip token, no parameters
-      lexer.skip_token ();
-      break;
-    case PIPE:
-      // actually may have parameters
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-
-      while (t->get_id () != PIPE)
-	{
-	  AST::ClosureParam param = parse_closure_param ();
-	  if (param.is_error ())
-	    {
-	      // TODO is this really an error?
-	      Error error (t->get_locus (), "could not parse closure param");
-	      add_error (std::move (error));
-
-	      break;
-	    }
-	  params.push_back (std::move (param));
-
-	  if (lexer.peek_token ()->get_id () != COMMA)
-	    {
-	      lexer.skip_token ();
-	      // not an error but means param list is done
-	      break;
-	    }
-	  // skip comma
-	  lexer.skip_token ();
-
-	  t = lexer.peek_token ();
-	}
-      params.shrink_to_fit ();
-      break;
-    default:
-      add_error (Error (t->get_locus (),
-			"unexpected token %qs in closure expression - expected "
-			"%<|%> or %<||%>",
-			t->get_token_description ()));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // again branch based on next token
-  t = lexer.peek_token ();
-  if (t->get_id () == RETURN_TYPE)
-    {
-      // must be return type closure with block expr
-
-      // skip "return type" token
-      lexer.skip_token ();
-
-      // parse actual type, which is required
-      std::unique_ptr<AST::TypeNoBounds> type = parse_type_no_bounds ();
-      if (type == nullptr)
-	{
-	  // error
-	  Error error (t->get_locus (), "failed to parse type for closure");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      // parse block expr, which is required
-      std::unique_ptr<AST::BlockExpr> block = parse_block_expr ();
-      if (block == nullptr)
-	{
-	  // error
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse block expr in closure");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      return std::unique_ptr<AST::ClosureExprInnerTyped> (
-	new AST::ClosureExprInnerTyped (std::move (type), std::move (block),
-					std::move (params), locus, has_move,
-					std::move (outer_attrs)));
-    }
-  else
-    {
-      // must be expr-only closure
-
-      // parse expr, which is required
-      std::unique_ptr<AST::Expr> expr = parse_expr ();
-      if (expr == nullptr)
-	{
-	  Error error (t->get_locus (),
-		       "failed to parse expression in closure");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      return std::unique_ptr<AST::ClosureExprInner> (
-	new AST::ClosureExprInner (std::move (expr), std::move (params), locus,
-				   has_move, std::move (outer_attrs)));
-    }
-}
-
-// Parses a literal token (to literal expression).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::LiteralExpr>
-Parser<ManagedTokenSource>::parse_literal_expr (AST::AttrVec outer_attrs)
-{
-  // TODO: change if literal representation in lexer changes
-
-  std::string literal_value;
-  AST::Literal::LitType type = AST::Literal::STRING;
-
-  // branch based on token
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case CHAR_LITERAL:
-      type = AST::Literal::CHAR;
-      literal_value = t->get_str ();
-      lexer.skip_token ();
-      break;
-    case STRING_LITERAL:
-      type = AST::Literal::STRING;
-      literal_value = t->get_str ();
-      lexer.skip_token ();
-      break;
-    case BYTE_CHAR_LITERAL:
-      type = AST::Literal::BYTE;
-      literal_value = t->get_str ();
-      lexer.skip_token ();
-      break;
-    case BYTE_STRING_LITERAL:
-      type = AST::Literal::BYTE_STRING;
-      literal_value = t->get_str ();
-      lexer.skip_token ();
-      break;
-    case RAW_STRING_LITERAL:
-      type = AST::Literal::RAW_STRING;
-      literal_value = t->get_str ();
-      lexer.skip_token ();
-      break;
-    case INT_LITERAL:
-      type = AST::Literal::INT;
-      literal_value = t->get_str ();
-      lexer.skip_token ();
-      break;
-    case FLOAT_LITERAL:
-      type = AST::Literal::FLOAT;
-      literal_value = t->get_str ();
-      lexer.skip_token ();
-      break;
-    // case BOOL_LITERAL
-    // use true and false keywords rather than "bool literal" Rust terminology
-    case TRUE_LITERAL:
-      type = AST::Literal::BOOL;
-      literal_value = Values::Keywords::TRUE_LITERAL;
-      lexer.skip_token ();
-      break;
-    case FALSE_LITERAL:
-      type = AST::Literal::BOOL;
-      literal_value = Values::Keywords::FALSE_LITERAL;
-      lexer.skip_token ();
-      break;
-    default:
-      // error - cannot be a literal expr
-      add_error (Error (t->get_locus (),
-			"unexpected token %qs when parsing literal expression",
-			t->get_token_description ()));
-
-      // skip?
-      return nullptr;
-    }
-
-  // create literal based on stuff in switch
-  return std::unique_ptr<AST::LiteralExpr> (
-    new AST::LiteralExpr (std::move (literal_value), std::move (type),
-			  t->get_type_hint (), std::move (outer_attrs),
-			  t->get_locus ()));
-}
-
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::BoxExpr>
-Parser<ManagedTokenSource>::parse_box_expr (AST::AttrVec outer_attrs,
-					    location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (BOX);
-    }
-
-  ParseRestrictions restrictions;
-  restrictions.expr_can_be_null = false;
-
-  std::unique_ptr<AST::Expr> expr = parse_expr (AST::AttrVec (), restrictions);
-
-  return std::unique_ptr<AST::BoxExpr> (
-    new AST::BoxExpr (std::move (expr), std::move (outer_attrs), locus));
-}
-
-// Parses a return expression (including any expression to return).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ReturnExpr>
-Parser<ManagedTokenSource>::parse_return_expr (AST::AttrVec outer_attrs,
-					       location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (RETURN_KW);
-    }
-
-  // parse expression to return, if it exists
-  ParseRestrictions restrictions;
-  restrictions.expr_can_be_null = true;
-  std::unique_ptr<AST::Expr> returned_expr
-    = parse_expr (AST::AttrVec (), restrictions);
-
-  return std::unique_ptr<AST::ReturnExpr> (
-    new AST::ReturnExpr (std::move (returned_expr), std::move (outer_attrs),
-			 locus));
-}
-
-// Parses a try expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::TryExpr>
-Parser<ManagedTokenSource>::parse_try_expr (AST::AttrVec outer_attrs,
-					    location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (TRY);
-    }
-
-  std::unique_ptr<AST::BlockExpr> block_expr = parse_block_expr ();
-
-  if (!block_expr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse try block expression");
-      add_error (std::move (error));
-
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::TryExpr> (
-    new AST::TryExpr (std::move (block_expr), std::move (outer_attrs), locus));
-}
-
-/* Parses a break expression (including any label to break to AND any return
- * expression). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::BreakExpr>
-Parser<ManagedTokenSource>::parse_break_expr (AST::AttrVec outer_attrs,
-					      location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (BREAK);
-    }
-
-  auto parsed_label = parse_lifetime (false);
-  auto label = (parsed_label)
-		 ? tl::optional<AST::Lifetime> (parsed_label.value ())
-		 : tl::nullopt;
-
-  // parse break return expression if it exists
-  ParseRestrictions restrictions;
-  restrictions.expr_can_be_null = true;
-  std::unique_ptr<AST::Expr> return_expr
-    = parse_expr (AST::AttrVec (), restrictions);
-
-  return std::unique_ptr<AST::BreakExpr> (
-    new AST::BreakExpr (std::move (label), std::move (return_expr),
-			std::move (outer_attrs), locus));
-}
-
-// Parses a continue expression (including any label to continue from).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ContinueExpr>
-Parser<ManagedTokenSource>::parse_continue_expr (AST::AttrVec outer_attrs,
-						 location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (CONTINUE);
-    }
-
-  auto parsed_label = parse_lifetime (false);
-  auto label = (parsed_label)
-		 ? tl::optional<AST::Lifetime> (parsed_label.value ())
-		 : tl::nullopt;
-
-  return std::unique_ptr<AST::ContinueExpr> (
-    new AST::ContinueExpr (std::move (label), std::move (outer_attrs), locus));
-}
-
 // Parses a loop label used in loop expressions.
 template <typename ManagedTokenSource>
 tl::expected<AST::LoopLabel, ParseLoopLabelError>
@@ -7677,804 +5402,6 @@ Parser<ManagedTokenSource>::parse_loop_label (const_TokenPtr tok)
 
   return tl::expected<AST::LoopLabel, ParseLoopLabelError> (
     AST::LoopLabel (std::move (label), tok->get_locus ()));
-}
-
-/* Parses an if expression of any kind, including with else, else if, else if
- * let, and neither. Note that any outer attributes will be ignored because if
- * expressions don't support them. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::IfExpr>
-Parser<ManagedTokenSource>::parse_if_expr (AST::AttrVec outer_attrs,
-					   location_t pratt_parsed_loc)
-{
-  // TODO: make having outer attributes an error?
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      if (!skip_token (IF))
-	{
-	  skip_after_end_block ();
-	  return nullptr;
-	}
-    }
-
-  // detect accidental if let
-  if (lexer.peek_token ()->get_id () == LET)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "if let expression probably exists, but is being parsed "
-		   "as an if expression. This may be a parser error");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  /* parse required condition expr - HACK to prevent struct expr from being
-   * parsed */
-  ParseRestrictions no_struct_expr;
-  no_struct_expr.can_be_struct_expr = false;
-  std::unique_ptr<AST::Expr> condition = parse_expr ({}, no_struct_expr);
-  if (condition == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse condition expression in if expression");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // parse required block expr
-  std::unique_ptr<AST::BlockExpr> if_body = parse_block_expr ();
-  if (if_body == nullptr)
-    return nullptr;
-
-  // branch to parse end or else (and then else, else if, or else if let)
-  if (lexer.peek_token ()->get_id () != ELSE)
-    {
-      // single selection - end of if expression
-      return std::unique_ptr<AST::IfExpr> (
-	new AST::IfExpr (std::move (condition), std::move (if_body),
-			 std::move (outer_attrs), locus));
-    }
-  else
-    {
-      // double or multiple selection - branch on end, else if, or else if let
-
-      // skip "else"
-      lexer.skip_token ();
-
-      // branch on whether next token is '{' or 'if'
-      const_TokenPtr t = lexer.peek_token ();
-      switch (t->get_id ())
-	{
-	case LEFT_CURLY:
-	  {
-	    // double selection - else
-	    // parse else block expr (required)
-	    std::unique_ptr<AST::BlockExpr> else_body = parse_block_expr ();
-	    if (else_body == nullptr)
-	      {
-		Error error (lexer.peek_token ()->get_locus (),
-			     "failed to parse else body block expression in "
-			     "if expression");
-		add_error (std::move (error));
-
-		// skip somewhere?
-		return nullptr;
-	      }
-
-	    return std::unique_ptr<AST::IfExprConseqElse> (
-	      new AST::IfExprConseqElse (std::move (condition),
-					 std::move (if_body),
-					 std::move (else_body),
-					 std::move (outer_attrs), locus));
-	  }
-	case IF:
-	  {
-	    // multiple selection - else if or else if let
-	    // branch on whether next token is 'let' or not
-	    if (lexer.peek_token (1)->get_id () == LET)
-	      {
-		// parse if let expr (required)
-		std::unique_ptr<AST::IfLetExpr> if_let_expr
-		  = parse_if_let_expr ();
-		if (if_let_expr == nullptr)
-		  {
-		    Error error (lexer.peek_token ()->get_locus (),
-				 "failed to parse (else) if let expression "
-				 "after if expression");
-		    add_error (std::move (error));
-
-		    // skip somewhere?
-		    return nullptr;
-		  }
-
-		return std::unique_ptr<AST::IfExprConseqElse> (
-		  new AST::IfExprConseqElse (std::move (condition),
-					     std::move (if_body),
-					     std::move (if_let_expr),
-					     std::move (outer_attrs), locus));
-	      }
-	    else
-	      {
-		// parse if expr (required)
-		std::unique_ptr<AST::IfExpr> if_expr = parse_if_expr ();
-		if (if_expr == nullptr)
-		  {
-		    Error error (lexer.peek_token ()->get_locus (),
-				 "failed to parse (else) if expression after "
-				 "if expression");
-		    add_error (std::move (error));
-
-		    // skip somewhere?
-		    return nullptr;
-		  }
-
-		return std::unique_ptr<AST::IfExprConseqElse> (
-		  new AST::IfExprConseqElse (std::move (condition),
-					     std::move (if_body),
-					     std::move (if_expr),
-					     std::move (outer_attrs), locus));
-	      }
-	  }
-	default:
-	  // error - invalid token
-	  add_error (Error (t->get_locus (),
-			    "unexpected token %qs after else in if expression",
-			    t->get_token_description ()));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-    }
-}
-
-/* Parses an if let expression of any kind, including with else, else if, else
- * if let, and none. Note that any outer attributes will be ignored as if let
- * expressions don't support them. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::IfLetExpr>
-Parser<ManagedTokenSource>::parse_if_let_expr (AST::AttrVec outer_attrs,
-					       location_t pratt_parsed_loc)
-{
-  // TODO: make having outer attributes an error?
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      if (!skip_token (IF))
-	{
-	  skip_after_end_block ();
-	  return nullptr;
-	}
-    }
-
-  // detect accidental if expr parsed as if let expr
-  if (lexer.peek_token ()->get_id () != LET)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "if expression probably exists, but is being parsed as an "
-		   "if let expression. This may be a parser error");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-  lexer.skip_token ();
-
-  // parse match arm patterns (which are required)
-  std::vector<std::unique_ptr<AST::Pattern>> match_arm_patterns
-    = parse_match_arm_patterns (EQUAL);
-  if (match_arm_patterns.empty ())
-    {
-      Error error (
-	lexer.peek_token ()->get_locus (),
-	"failed to parse any match arm patterns in if let expression");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  if (!skip_token (EQUAL))
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // parse expression (required) - HACK to prevent struct expr being parsed
-  ParseRestrictions no_struct_expr;
-  no_struct_expr.can_be_struct_expr = false;
-  std::unique_ptr<AST::Expr> scrutinee_expr = parse_expr ({}, no_struct_expr);
-  if (scrutinee_expr == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse scrutinee expression in if let expression");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-  /* TODO: check for expression not being a struct expression or lazy boolean
-   * expression here? or actually probably in semantic analysis. */
-
-  // parse block expression (required)
-  std::unique_ptr<AST::BlockExpr> if_let_body = parse_block_expr ();
-  if (if_let_body == nullptr)
-    {
-      Error error (
-	lexer.peek_token ()->get_locus (),
-	"failed to parse if let body block expression in if let expression");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // branch to parse end or else (and then else, else if, or else if let)
-  if (lexer.peek_token ()->get_id () != ELSE)
-    {
-      // single selection - end of if let expression
-      return std::unique_ptr<AST::IfLetExpr> (
-	new AST::IfLetExpr (std::move (match_arm_patterns),
-			    std::move (scrutinee_expr), std::move (if_let_body),
-			    std::move (outer_attrs), locus));
-    }
-  else
-    {
-      // double or multiple selection - branch on end, else if, or else if let
-
-      // skip "else"
-      lexer.skip_token ();
-
-      // branch on whether next token is '{' or 'if'
-      const_TokenPtr t = lexer.peek_token ();
-      switch (t->get_id ())
-	{
-	case LEFT_CURLY:
-	  {
-	    // double selection - else
-	    // parse else block expr (required)
-	    std::unique_ptr<AST::BlockExpr> else_body = parse_block_expr ();
-	    if (else_body == nullptr)
-	      {
-		Error error (lexer.peek_token ()->get_locus (),
-			     "failed to parse else body block expression in "
-			     "if let expression");
-		add_error (std::move (error));
-
-		// skip somewhere?
-		return nullptr;
-	      }
-
-	    return std::unique_ptr<AST::IfLetExprConseqElse> (
-	      new AST::IfLetExprConseqElse (std::move (match_arm_patterns),
-					    std::move (scrutinee_expr),
-					    std::move (if_let_body),
-					    std::move (else_body),
-					    std::move (outer_attrs), locus));
-	  }
-	case IF:
-	  {
-	    // multiple selection - else if or else if let
-	    // branch on whether next token is 'let' or not
-	    if (lexer.peek_token (1)->get_id () == LET)
-	      {
-		// parse if let expr (required)
-		std::unique_ptr<AST::IfLetExpr> if_let_expr
-		  = parse_if_let_expr ();
-		if (if_let_expr == nullptr)
-		  {
-		    Error error (lexer.peek_token ()->get_locus (),
-				 "failed to parse (else) if let expression "
-				 "after if let expression");
-		    add_error (std::move (error));
-
-		    // skip somewhere?
-		    return nullptr;
-		  }
-
-		return std::unique_ptr<AST::IfLetExprConseqElse> (
-		  new AST::IfLetExprConseqElse (
-		    std::move (match_arm_patterns), std::move (scrutinee_expr),
-		    std::move (if_let_body), std::move (if_let_expr),
-		    std::move (outer_attrs), locus));
-	      }
-	    else
-	      {
-		// parse if expr (required)
-		std::unique_ptr<AST::IfExpr> if_expr = parse_if_expr ();
-		if (if_expr == nullptr)
-		  {
-		    Error error (lexer.peek_token ()->get_locus (),
-				 "failed to parse (else) if expression after "
-				 "if let expression");
-		    add_error (std::move (error));
-
-		    // skip somewhere?
-		    return nullptr;
-		  }
-
-		return std::unique_ptr<AST::IfLetExprConseqElse> (
-		  new AST::IfLetExprConseqElse (
-		    std::move (match_arm_patterns), std::move (scrutinee_expr),
-		    std::move (if_let_body), std::move (if_expr),
-		    std::move (outer_attrs), locus));
-	      }
-	  }
-	default:
-	  // error - invalid token
-	  add_error (
-	    Error (t->get_locus (),
-		   "unexpected token %qs after else in if let expression",
-		   t->get_token_description ()));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-    }
-}
-
-/* TODO: possibly decide on different method of handling label (i.e. not
- * parameter) */
-
-/* Parses a "loop" infinite loop expression. Label is not parsed and should be
- * parsed via parse_labelled_loop_expr, which would call this. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::LoopExpr>
-Parser<ManagedTokenSource>::parse_loop_expr (AST::AttrVec outer_attrs,
-					     tl::optional<AST::LoopLabel> label,
-					     location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      if (label)
-	locus = label->get_locus ();
-      else
-	locus = lexer.peek_token ()->get_locus ();
-
-      if (!skip_token (LOOP))
-	{
-	  skip_after_end_block ();
-	  return nullptr;
-	}
-    }
-  else
-    {
-      if (label)
-	locus = label->get_locus ();
-    }
-
-  // parse loop body, which is required
-  std::unique_ptr<AST::BlockExpr> loop_body = parse_block_expr ();
-  if (loop_body == nullptr)
-    return nullptr;
-
-  return std::unique_ptr<AST::LoopExpr> (
-    new AST::LoopExpr (std::move (loop_body), locus, std::move (label),
-		       std::move (outer_attrs)));
-}
-
-/* Parses a "while" loop expression. Label is not parsed and should be parsed
- * via parse_labelled_loop_expr, which would call this. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::WhileLoopExpr>
-Parser<ManagedTokenSource>::parse_while_loop_expr (
-  AST::AttrVec outer_attrs, tl::optional<AST::LoopLabel> label,
-  location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      if (label)
-	locus = label->get_locus ();
-      else
-	locus = lexer.peek_token ()->get_locus ();
-
-      if (!skip_token (WHILE))
-	{
-	  skip_after_end_block ();
-	  return nullptr;
-	}
-    }
-  else
-    {
-      if (label)
-	locus = label->get_locus ();
-    }
-
-  // ensure it isn't a while let loop
-  if (lexer.peek_token ()->get_id () == LET)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "appears to be while let loop but is being parsed by "
-		   "while loop - this may be a compiler issue");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // parse loop predicate (required) with HACK to prevent struct expr parsing
-  ParseRestrictions no_struct_expr;
-  no_struct_expr.can_be_struct_expr = false;
-  std::unique_ptr<AST::Expr> predicate = parse_expr ({}, no_struct_expr);
-  if (predicate == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse predicate expression in while loop");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-  /* TODO: check that it isn't struct expression here? actually, probably in
-   * semantic analysis */
-
-  // parse loop body (required)
-  std::unique_ptr<AST::BlockExpr> body = parse_block_expr ();
-  if (body == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse loop body block expression in while loop");
-      add_error (std::move (error));
-
-      // skip somewhere
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::WhileLoopExpr> (
-    new AST::WhileLoopExpr (std::move (predicate), std::move (body), locus,
-			    std::move (label), std::move (outer_attrs)));
-}
-
-/* Parses a "while let" loop expression. Label is not parsed and should be
- * parsed via parse_labelled_loop_expr, which would call this. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::WhileLetLoopExpr>
-Parser<ManagedTokenSource>::parse_while_let_loop_expr (
-  AST::AttrVec outer_attrs, tl::optional<AST::LoopLabel> label)
-{
-  location_t locus = UNKNOWN_LOCATION;
-  if (label)
-    locus = label->get_locus ();
-  else
-    locus = lexer.peek_token ()->get_locus ();
-  maybe_skip_token (WHILE);
-
-  /* check for possible accidental recognition of a while loop as a while let
-   * loop */
-  if (lexer.peek_token ()->get_id () != LET)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "appears to be a while loop but is being parsed by "
-		   "while let loop - this may be a compiler issue");
-      add_error (std::move (error));
-
-      // skip somewhere
-      return nullptr;
-    }
-  // as this token is definitely let now, save the computation of comparison
-  lexer.skip_token ();
-
-  // parse predicate patterns
-  std::vector<std::unique_ptr<AST::Pattern>> predicate_patterns
-    = parse_match_arm_patterns (EQUAL);
-  // ensure that there is at least 1 pattern
-  if (predicate_patterns.empty ())
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "should be at least 1 pattern");
-      add_error (std::move (error));
-      return nullptr;
-    }
-
-  if (!skip_token (EQUAL))
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-
-  /* parse predicate expression, which is required (and HACK to prevent struct
-   * expr) */
-  ParseRestrictions no_struct_expr;
-  no_struct_expr.can_be_struct_expr = false;
-  std::unique_ptr<AST::Expr> predicate_expr = parse_expr ({}, no_struct_expr);
-  if (predicate_expr == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse predicate expression in while let loop");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-  /* TODO: ensure that struct expression is not parsed? Actually, probably in
-   * semantic analysis. */
-
-  // parse loop body, which is required
-  std::unique_ptr<AST::BlockExpr> body = parse_block_expr ();
-  if (body == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse block expr (loop body) of while let loop");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::WhileLetLoopExpr> (new AST::WhileLetLoopExpr (
-    std::move (predicate_patterns), std::move (predicate_expr),
-    std::move (body), locus, std::move (label), std::move (outer_attrs)));
-}
-
-/* Parses a "for" iterative loop. Label is not parsed and should be parsed via
- * parse_labelled_loop_expr, which would call this. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ForLoopExpr>
-Parser<ManagedTokenSource>::parse_for_loop_expr (
-  AST::AttrVec outer_attrs, tl::optional<AST::LoopLabel> label)
-{
-  location_t locus = UNKNOWN_LOCATION;
-  if (label)
-    locus = label->get_locus ();
-  else
-    locus = lexer.peek_token ()->get_locus ();
-  maybe_skip_token (FOR);
-
-  // parse pattern, which is required
-  std::unique_ptr<AST::Pattern> pattern = parse_pattern ();
-  if (pattern == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse iterator pattern in for loop");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  if (!skip_token (IN))
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-
-  /* parse iterator expression, which is required - also HACK to prevent
-   * struct expr */
-  ParseRestrictions no_struct_expr;
-  no_struct_expr.can_be_struct_expr = false;
-  std::unique_ptr<AST::Expr> expr = parse_expr ({}, no_struct_expr);
-  if (expr == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse iterator expression in for loop");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-  // TODO: check to ensure this isn't struct expr? Or in semantic analysis.
-
-  // parse loop body, which is required
-  std::unique_ptr<AST::BlockExpr> body = parse_block_expr ();
-  if (body == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse loop body block expression in for loop");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::ForLoopExpr> (
-    new AST::ForLoopExpr (std::move (pattern), std::move (expr),
-			  std::move (body), locus, std::move (label),
-			  std::move (outer_attrs)));
-}
-
-// Parses a loop expression with label (any kind of loop - disambiguates).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::parse_labelled_loop_expr (const_TokenPtr tok,
-						      AST::AttrVec outer_attrs)
-{
-  /* TODO: decide whether it should not work if there is no label, or parse it
-   * with no label at the moment, I will make it not work with no label
-   * because that's the implication. */
-
-  if (tok->get_id () != LIFETIME)
-    {
-      Error error (tok->get_locus (),
-		   "expected lifetime in labelled loop expr (to parse loop "
-		   "label) - found %qs",
-		   tok->get_token_description ());
-      add_error (std::move (error));
-
-      // skip?
-      return nullptr;
-    }
-
-  // parse loop label (required)
-  // TODO: Convert this return type to tl::expected instead of tl::optional
-  auto parsed_label = parse_loop_label (tok);
-  if (!parsed_label)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse loop label in labelled loop expr");
-      add_error (std::move (error));
-
-      // skip?
-      return nullptr;
-    }
-
-  auto label = parsed_label
-		 ? tl::optional<AST::LoopLabel> (parsed_label.value ())
-		 : tl::nullopt;
-
-  // branch on next token
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case LOOP:
-      return parse_loop_expr (std::move (outer_attrs), std::move (label));
-    case FOR:
-      return parse_for_loop_expr (std::move (outer_attrs), std::move (label));
-    case WHILE:
-      // further disambiguate into while vs while let
-      if (lexer.peek_token (1)->get_id () == LET)
-	{
-	  return parse_while_let_loop_expr (std::move (outer_attrs),
-					    std::move (label));
-	}
-      else
-	{
-	  return parse_while_loop_expr (std::move (outer_attrs),
-					std::move (label));
-	}
-    case LEFT_CURLY:
-      return parse_block_expr (std::move (outer_attrs), std::move (label));
-    default:
-      // error
-      add_error (Error (t->get_locus (),
-			"unexpected token %qs when parsing labelled loop",
-			t->get_token_description ()));
-
-      // skip?
-      return nullptr;
-    }
-}
-
-// Parses a match expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MatchExpr>
-Parser<ManagedTokenSource>::parse_match_expr (AST::AttrVec outer_attrs,
-					      location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (MATCH_KW);
-    }
-
-  /* parse scrutinee expression, which is required (and HACK to prevent struct
-   * expr) */
-  ParseRestrictions no_struct_expr;
-  no_struct_expr.can_be_struct_expr = false;
-  std::unique_ptr<AST::Expr> scrutinee = parse_expr ({}, no_struct_expr);
-  if (scrutinee == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse scrutinee expression in match expression");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-  /* TODO: check for scrutinee expr not being struct expr? or do so in
-   * semantic analysis */
-
-  if (!skip_token (LEFT_CURLY))
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // parse inner attributes (if they exist)
-  AST::AttrVec inner_attrs = parse_inner_attributes ();
-
-  // parse match arms (if they exist)
-  // std::vector<std::unique_ptr<AST::MatchCase> > match_arms;
-  std::vector<AST::MatchCase> match_arms;
-
-  // parse match cases
-  while (lexer.peek_token ()->get_id () != RIGHT_CURLY)
-    {
-      // parse match arm itself, which is required
-      AST::MatchArm arm = parse_match_arm ();
-      if (arm.is_error ())
-	{
-	  // TODO is this worth throwing everything away?
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse match arm in match arms");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-
-      if (!skip_token (MATCH_ARROW))
-	{
-	  // skip after somewhere?
-	  // TODO is returning here a good idea? or is break better?
-	  return nullptr;
-	}
-
-      ParseRestrictions restrictions;
-      restrictions.expr_can_be_stmt = true;
-
-      std::unique_ptr<AST::Expr> expr = parse_expr ({}, restrictions);
-
-      if (expr == nullptr)
-	{
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse expr in match arm in match expr");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      bool is_expr_without_block = expr->is_expr_without_block ();
-
-      match_arms.push_back (AST::MatchCase (std::move (arm), std::move (expr)));
-
-      // handle comma presence
-      if (lexer.peek_token ()->get_id () != COMMA)
-	{
-	  if (!is_expr_without_block)
-	    {
-	      // allowed even if not final case
-	      continue;
-	    }
-	  else if (is_expr_without_block
-		   && lexer.peek_token ()->get_id () != RIGHT_CURLY)
-	    {
-	      // not allowed if not final case
-	      Error error (lexer.peek_token ()->get_locus (),
-			   "exprwithoutblock requires comma after match case "
-			   "expression in match arm (if not final case)");
-	      add_error (std::move (error));
-
-	      return nullptr;
-	    }
-	  else
-	    {
-	      // otherwise, must be final case, so fine
-	      break;
-	    }
-	}
-      lexer.skip_token ();
-    }
-
-  if (!skip_token (RIGHT_CURLY))
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-
-  match_arms.shrink_to_fit ();
-
-  return std::unique_ptr<AST::MatchExpr> (
-    new AST::MatchExpr (std::move (scrutinee), std::move (match_arms),
-			std::move (inner_attrs), std::move (outer_attrs),
-			locus));
 }
 
 // Parses the "pattern" part of the match arm (the 'case x:' equivalent).
@@ -8603,217 +5530,6 @@ Parser<ManagedTokenSource>::parse_match_arm_patterns (TokenId end_token_id)
   return patterns;
 }
 
-// Parses an async block expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::AsyncBlockExpr>
-Parser<ManagedTokenSource>::parse_async_block_expr (AST::AttrVec outer_attrs)
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-  skip_token (ASYNC);
-
-  // detect optional move token
-  bool has_move = false;
-  if (lexer.peek_token ()->get_id () == MOVE)
-    {
-      lexer.skip_token ();
-      has_move = true;
-    }
-
-  // parse block expression (required)
-  std::unique_ptr<AST::BlockExpr> block_expr = parse_block_expr ();
-  if (block_expr == nullptr)
-    {
-      Error error (
-	lexer.peek_token ()->get_locus (),
-	"failed to parse block expression of async block expression");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::AsyncBlockExpr> (
-    new AST::AsyncBlockExpr (std::move (block_expr), has_move,
-			     std::move (outer_attrs), locus));
-}
-
-// Parses an unsafe block expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::UnsafeBlockExpr>
-Parser<ManagedTokenSource>::parse_unsafe_block_expr (
-  AST::AttrVec outer_attrs, location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (UNSAFE);
-    }
-
-  // parse block expression (required)
-  std::unique_ptr<AST::BlockExpr> block_expr = parse_block_expr ();
-  if (block_expr == nullptr)
-    {
-      Error error (
-	lexer.peek_token ()->get_locus (),
-	"failed to parse block expression of unsafe block expression");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::UnsafeBlockExpr> (
-    new AST::UnsafeBlockExpr (std::move (block_expr), std::move (outer_attrs),
-			      locus));
-}
-
-// Parses an array definition expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArrayExpr>
-Parser<ManagedTokenSource>::parse_array_expr (AST::AttrVec outer_attrs,
-					      location_t pratt_parsed_loc)
-{
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (LEFT_SQUARE);
-    }
-
-  // parse optional inner attributes
-  AST::AttrVec inner_attrs = parse_inner_attributes ();
-
-  // parse the "array elements" section, which is optional
-  if (lexer.peek_token ()->get_id () == RIGHT_SQUARE)
-    {
-      // no array elements
-      lexer.skip_token ();
-
-      std::vector<std::unique_ptr<AST::Expr>> exprs;
-      auto array_elems
-	= std::make_unique<AST::ArrayElemsValues> (std::move (exprs), locus);
-      return std::make_unique<AST::ArrayExpr> (std::move (array_elems),
-					       std::move (inner_attrs),
-					       std::move (outer_attrs), locus);
-    }
-  else
-    {
-      // should have array elements
-      // parse initial expression, which is required for either
-      std::unique_ptr<AST::Expr> initial_expr = parse_expr ();
-      if (initial_expr == nullptr)
-	{
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "could not parse expression in array expression "
-		       "(even though arrayelems seems to be present)");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      if (lexer.peek_token ()->get_id () == SEMICOLON)
-	{
-	  // copy array elems
-	  lexer.skip_token ();
-
-	  // parse copy amount expression (required)
-	  std::unique_ptr<AST::Expr> copy_amount = parse_expr ();
-	  if (copy_amount == nullptr)
-	    {
-	      Error error (lexer.peek_token ()->get_locus (),
-			   "could not parse copy amount expression in array "
-			   "expression (arrayelems)");
-	      add_error (std::move (error));
-
-	      // skip somewhere?
-	      return nullptr;
-	    }
-
-	  skip_token (RIGHT_SQUARE);
-
-	  std::unique_ptr<AST::ArrayElemsCopied> copied_array_elems (
-	    new AST::ArrayElemsCopied (std::move (initial_expr),
-				       std::move (copy_amount), locus));
-	  return std::unique_ptr<AST::ArrayExpr> (
-	    new AST::ArrayExpr (std::move (copied_array_elems),
-				std::move (inner_attrs),
-				std::move (outer_attrs), locus));
-	}
-      else if (lexer.peek_token ()->get_id () == RIGHT_SQUARE)
-	{
-	  // single-element array expression
-	  std::vector<std::unique_ptr<AST::Expr>> exprs;
-	  exprs.reserve (1);
-	  exprs.push_back (std::move (initial_expr));
-	  exprs.shrink_to_fit ();
-
-	  skip_token (RIGHT_SQUARE);
-
-	  std::unique_ptr<AST::ArrayElemsValues> array_elems (
-	    new AST::ArrayElemsValues (std::move (exprs), locus));
-	  return std::unique_ptr<AST::ArrayExpr> (
-	    new AST::ArrayExpr (std::move (array_elems),
-				std::move (inner_attrs),
-				std::move (outer_attrs), locus));
-	}
-      else if (lexer.peek_token ()->get_id () == COMMA)
-	{
-	  // multi-element array expression (or trailing comma)
-	  std::vector<std::unique_ptr<AST::Expr>> exprs;
-	  exprs.push_back (std::move (initial_expr));
-
-	  const_TokenPtr t = lexer.peek_token ();
-	  while (t->get_id () == COMMA)
-	    {
-	      lexer.skip_token ();
-
-	      // quick break if right square bracket
-	      if (lexer.peek_token ()->get_id () == RIGHT_SQUARE)
-		break;
-
-	      // parse expression (required)
-	      std::unique_ptr<AST::Expr> expr = parse_expr ();
-	      if (expr == nullptr)
-		{
-		  Error error (lexer.peek_token ()->get_locus (),
-			       "failed to parse element in array expression");
-		  add_error (std::move (error));
-
-		  // skip somewhere?
-		  return nullptr;
-		}
-	      exprs.push_back (std::move (expr));
-
-	      t = lexer.peek_token ();
-	    }
-
-	  skip_token (RIGHT_SQUARE);
-
-	  exprs.shrink_to_fit ();
-
-	  std::unique_ptr<AST::ArrayElemsValues> array_elems (
-	    new AST::ArrayElemsValues (std::move (exprs), locus));
-	  return std::unique_ptr<AST::ArrayExpr> (
-	    new AST::ArrayExpr (std::move (array_elems),
-				std::move (inner_attrs),
-				std::move (outer_attrs), locus));
-	}
-      else
-	{
-	  // error
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "unexpected token %qs in array expression (arrayelems)",
-		       lexer.peek_token ()->get_token_description ());
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-    }
-}
-
 // Parses a single parameter used in a closure definition.
 template <typename ManagedTokenSource>
 AST::ClosureParam
@@ -8851,113 +5567,6 @@ Parser<ManagedTokenSource>::parse_closure_param ()
   location_t loc = pattern->get_locus ();
   return AST::ClosureParam (std::move (pattern), loc, std::move (type),
 			    std::move (outer_attrs));
-}
-
-// Parses a grouped or tuple expression (disambiguates).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ExprWithoutBlock>
-Parser<ManagedTokenSource>::parse_grouped_or_tuple_expr (
-  AST::AttrVec outer_attrs, location_t pratt_parsed_loc)
-{
-  // adjustment to allow Pratt parsing to reuse function without copy-paste
-  location_t locus = pratt_parsed_loc;
-  if (locus == UNKNOWN_LOCATION)
-    {
-      locus = lexer.peek_token ()->get_locus ();
-      skip_token (LEFT_PAREN);
-    }
-
-  // parse optional inner attributes
-  AST::AttrVec inner_attrs = parse_inner_attributes ();
-
-  if (lexer.peek_token ()->get_id () == RIGHT_PAREN)
-    {
-      // must be empty tuple
-      lexer.skip_token ();
-
-      // create tuple with empty tuple elems
-      return std::unique_ptr<AST::TupleExpr> (
-	new AST::TupleExpr (std::vector<std::unique_ptr<AST::Expr>> (),
-			    std::move (inner_attrs), std::move (outer_attrs),
-			    locus));
-    }
-
-  // parse first expression (required)
-  std::unique_ptr<AST::Expr> first_expr = parse_expr ();
-  if (first_expr == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse expression in grouped or tuple expression");
-      add_error (std::move (error));
-
-      // skip after somewhere?
-      return nullptr;
-    }
-
-  // detect whether grouped expression with right parentheses as next token
-  if (lexer.peek_token ()->get_id () == RIGHT_PAREN)
-    {
-      // must be grouped expr
-      lexer.skip_token ();
-
-      // create grouped expr
-      return std::unique_ptr<AST::GroupedExpr> (
-	new AST::GroupedExpr (std::move (first_expr), std::move (inner_attrs),
-			      std::move (outer_attrs), locus));
-    }
-  else if (lexer.peek_token ()->get_id () == COMMA)
-    {
-      // tuple expr
-      std::vector<std::unique_ptr<AST::Expr>> exprs;
-      exprs.push_back (std::move (first_expr));
-
-      // parse potential other tuple exprs
-      const_TokenPtr t = lexer.peek_token ();
-      while (t->get_id () == COMMA)
-	{
-	  lexer.skip_token ();
-
-	  // break out if right paren
-	  if (lexer.peek_token ()->get_id () == RIGHT_PAREN)
-	    break;
-
-	  // parse expr, which is now required
-	  std::unique_ptr<AST::Expr> expr = parse_expr ();
-	  if (expr == nullptr)
-	    {
-	      Error error (lexer.peek_token ()->get_locus (),
-			   "failed to parse expr in tuple expr");
-	      add_error (std::move (error));
-
-	      // skip somewhere?
-	      return nullptr;
-	    }
-	  exprs.push_back (std::move (expr));
-
-	  t = lexer.peek_token ();
-	}
-
-      // skip right paren
-      skip_token (RIGHT_PAREN);
-
-      return std::unique_ptr<AST::TupleExpr> (
-	new AST::TupleExpr (std::move (exprs), std::move (inner_attrs),
-			    std::move (outer_attrs), locus));
-    }
-  else
-    {
-      // error
-      const_TokenPtr t = lexer.peek_token ();
-      Error error (t->get_locus (),
-		   "unexpected token %qs in grouped or tuple expression "
-		   "(parenthesised expression) - expected %<)%> for grouped "
-		   "expr and %<,%> for tuple expr",
-		   t->get_token_description ());
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
 }
 
 // Parses a type (will further disambiguate any type).
@@ -10217,1040 +6826,6 @@ Parser<ManagedTokenSource>::parse_paren_prefixed_type_no_bounds ()
    * lost somehow */
 }
 
-/* Parses a literal pattern or range pattern. Assumes that literals passed in
- * are valid range pattern bounds. Do not pass in paths in expressions, for
- * instance. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Pattern>
-Parser<ManagedTokenSource>::parse_literal_or_range_pattern ()
-{
-  const_TokenPtr range_lower = lexer.peek_token ();
-  AST::Literal::LitType type = AST::Literal::STRING;
-  bool has_minus = false;
-
-  // get lit type
-  switch (range_lower->get_id ())
-    {
-    case CHAR_LITERAL:
-      type = AST::Literal::CHAR;
-      lexer.skip_token ();
-      break;
-    case BYTE_CHAR_LITERAL:
-      type = AST::Literal::BYTE;
-      lexer.skip_token ();
-      break;
-    case INT_LITERAL:
-      type = AST::Literal::INT;
-      lexer.skip_token ();
-      break;
-    case FLOAT_LITERAL:
-      type = AST::Literal::FLOAT;
-      lexer.skip_token ();
-      break;
-    case MINUS:
-      // branch on next token
-      range_lower = lexer.peek_token (1);
-      switch (range_lower->get_id ())
-	{
-	case INT_LITERAL:
-	  type = AST::Literal::INT;
-	  has_minus = true;
-	  lexer.skip_token (1);
-	  break;
-	case FLOAT_LITERAL:
-	  type = AST::Literal::FLOAT;
-	  has_minus = true;
-	  lexer.skip_token (1);
-	  break;
-	default:
-	  add_error (Error (range_lower->get_locus (),
-			    "token type %qs cannot be parsed as range pattern "
-			    "bound or literal after minus symbol",
-			    range_lower->get_token_description ()));
-
-	  return nullptr;
-	}
-      break;
-    default:
-      add_error (
-	Error (range_lower->get_locus (),
-	       "token type %qs cannot be parsed as range pattern bound",
-	       range_lower->get_token_description ()));
-
-      return nullptr;
-    }
-
-  const_TokenPtr next = lexer.peek_token ();
-  if (next->get_id () == DOT_DOT_EQ || next->get_id () == ELLIPSIS
-      || next->get_id () == DOT_DOT)
-    {
-      AST::RangeKind kind = AST::tokenid_to_rangekind (next->get_id ());
-      // range pattern
-      lexer.skip_token ();
-      std::unique_ptr<AST::RangePatternBound> lower (
-	new AST::RangePatternBoundLiteral (
-	  AST::Literal (range_lower->get_str (), type,
-			PrimitiveCoreType::CORETYPE_UNKNOWN),
-	  range_lower->get_locus (), has_minus));
-
-      std::unique_ptr<AST::RangePatternBound> upper
-	= parse_range_pattern_bound ();
-      if (upper == nullptr)
-	{
-	  Error error (next->get_locus (),
-		       "failed to parse range pattern bound in range pattern");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-
-      return std::unique_ptr<AST::RangePattern> (
-	new AST::RangePattern (std::move (lower), std::move (upper), kind,
-			       range_lower->get_locus ()));
-    }
-  else
-    {
-      // literal pattern
-      return std::unique_ptr<AST::LiteralPattern> (
-	new AST::LiteralPattern (range_lower->get_str (), type,
-				 range_lower->get_locus (),
-				 range_lower->get_type_hint (), has_minus));
-    }
-}
-
-// Parses a range pattern bound (value only).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::RangePatternBound>
-Parser<ManagedTokenSource>::parse_range_pattern_bound ()
-{
-  const_TokenPtr range_lower = lexer.peek_token ();
-  location_t range_lower_locus = range_lower->get_locus ();
-
-  // get lit type
-  switch (range_lower->get_id ())
-    {
-    case CHAR_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::RangePatternBoundLiteral> (
-	new AST::RangePatternBoundLiteral (
-	  AST::Literal (range_lower->get_str (), AST::Literal::CHAR,
-			range_lower->get_type_hint ()),
-	  range_lower_locus));
-    case BYTE_CHAR_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::RangePatternBoundLiteral> (
-	new AST::RangePatternBoundLiteral (
-	  AST::Literal (range_lower->get_str (), AST::Literal::BYTE,
-			range_lower->get_type_hint ()),
-	  range_lower_locus));
-    case INT_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::RangePatternBoundLiteral> (
-	new AST::RangePatternBoundLiteral (
-	  AST::Literal (range_lower->get_str (), AST::Literal::INT,
-			range_lower->get_type_hint ()),
-	  range_lower_locus));
-    case FLOAT_LITERAL:
-      lexer.skip_token ();
-      rust_debug ("warning: used deprecated float range pattern bound");
-      return std::unique_ptr<AST::RangePatternBoundLiteral> (
-	new AST::RangePatternBoundLiteral (
-	  AST::Literal (range_lower->get_str (), AST::Literal::FLOAT,
-			range_lower->get_type_hint ()),
-	  range_lower_locus));
-    case MINUS:
-      // branch on next token
-      range_lower = lexer.peek_token (1);
-      switch (range_lower->get_id ())
-	{
-	case INT_LITERAL:
-	  lexer.skip_token (1);
-	  return std::unique_ptr<AST::RangePatternBoundLiteral> (
-	    new AST::RangePatternBoundLiteral (
-	      AST::Literal (range_lower->get_str (), AST::Literal::INT,
-			    range_lower->get_type_hint ()),
-	      range_lower_locus, true));
-	case FLOAT_LITERAL:
-	  lexer.skip_token (1);
-	  rust_debug ("warning: used deprecated float range pattern bound");
-	  return std::unique_ptr<AST::RangePatternBoundLiteral> (
-	    new AST::RangePatternBoundLiteral (
-	      AST::Literal (range_lower->get_str (), AST::Literal::FLOAT,
-			    range_lower->get_type_hint ()),
-	      range_lower_locus, true));
-	default:
-	  add_error (Error (range_lower->get_locus (),
-			    "token type %qs cannot be parsed as range pattern "
-			    "bound after minus symbol",
-			    range_lower->get_token_description ()));
-
-	  return nullptr;
-	}
-    case IDENTIFIER:
-    case SUPER:
-    case SELF:
-    case SELF_ALIAS:
-    case CRATE:
-    case SCOPE_RESOLUTION:
-    case DOLLAR_SIGN:
-      {
-	// path in expression
-	AST::PathInExpression path = parse_path_in_expression ();
-	if (path.is_error ())
-	  {
-	    Error error (
-	      range_lower->get_locus (),
-	      "failed to parse path in expression range pattern bound");
-	    add_error (std::move (error));
-
-	    return nullptr;
-	  }
-	return std::unique_ptr<AST::RangePatternBoundPath> (
-	  new AST::RangePatternBoundPath (std::move (path)));
-      }
-    case LEFT_SHIFT:
-    case LEFT_ANGLE:
-      {
-	// qualified path in expression
-	AST::QualifiedPathInExpression path
-	  = parse_qualified_path_in_expression ();
-	if (path.is_error ())
-	  {
-	    Error error (range_lower->get_locus (),
-			 "failed to parse qualified path in expression range "
-			 "pattern bound");
-	    add_error (std::move (error));
-
-	    return nullptr;
-	  }
-	return std::unique_ptr<AST::RangePatternBoundQualPath> (
-	  new AST::RangePatternBoundQualPath (std::move (path)));
-      }
-    default:
-      add_error (
-	Error (range_lower->get_locus (),
-	       "token type %qs cannot be parsed as range pattern bound",
-	       range_lower->get_token_description ()));
-
-      return nullptr;
-    }
-}
-
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Pattern>
-Parser<ManagedTokenSource>::parse_pattern ()
-{
-  location_t start_locus = lexer.peek_token ()->get_locus ();
-
-  /* skip optional starting pipe */
-  maybe_skip_token (PIPE);
-
-  auto first = parse_pattern_no_alt ();
-
-  if (lexer.peek_token ()->get_id () != PIPE)
-    /* no alternates */
-    return first;
-
-  std::vector<std::unique_ptr<AST::Pattern>> alts;
-  if (first != nullptr)
-    alts.push_back (std::move (first));
-
-  do
-    {
-      lexer.skip_token ();
-      auto follow = parse_pattern_no_alt ();
-      if (follow != nullptr)
-	alts.push_back (std::move (follow));
-    }
-
-  while (lexer.peek_token ()->get_id () == PIPE);
-
-  if (alts.empty ())
-    return nullptr;
-
-  /* alternates */
-  return std::unique_ptr<AST::Pattern> (
-    new AST::AltPattern (std::move (alts), start_locus));
-}
-
-// Parses a pattern without alternates ('|')
-// (will further disambiguate any pattern).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Pattern>
-Parser<ManagedTokenSource>::parse_pattern_no_alt ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case TRUE_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::LiteralPattern> (
-	new AST::LiteralPattern (Values::Keywords::TRUE_LITERAL,
-				 AST::Literal::BOOL, t->get_locus (),
-				 t->get_type_hint ()));
-    case FALSE_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::LiteralPattern> (
-	new AST::LiteralPattern (Values::Keywords::FALSE_LITERAL,
-				 AST::Literal::BOOL, t->get_locus (),
-				 t->get_type_hint ()));
-    case CHAR_LITERAL:
-    case BYTE_CHAR_LITERAL:
-    case INT_LITERAL:
-    case FLOAT_LITERAL:
-      return parse_literal_or_range_pattern ();
-    case STRING_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::LiteralPattern> (
-	new AST::LiteralPattern (t->get_str (), AST::Literal::STRING,
-				 t->get_locus (), t->get_type_hint ()));
-    case BYTE_STRING_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::LiteralPattern> (
-	new AST::LiteralPattern (t->get_str (), AST::Literal::BYTE_STRING,
-				 t->get_locus (), t->get_type_hint ()));
-    case RAW_STRING_LITERAL:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::LiteralPattern> (
-	new AST::LiteralPattern (t->get_str (), AST::Literal::RAW_STRING,
-				 t->get_locus (), t->get_type_hint ()));
-    // raw string and raw byte string literals too if they are readded to
-    // lexer
-    case MINUS:
-      if (lexer.peek_token (1)->get_id () == INT_LITERAL)
-	{
-	  return parse_literal_or_range_pattern ();
-	}
-      else if (lexer.peek_token (1)->get_id () == FLOAT_LITERAL)
-	{
-	  return parse_literal_or_range_pattern ();
-	}
-      else
-	{
-	  Error error (t->get_locus (), "unexpected token %<-%> in pattern - "
-					"did you forget an integer literal");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-    case UNDERSCORE:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::WildcardPattern> (
-	new AST::WildcardPattern (t->get_locus ()));
-    case DOT_DOT:
-      lexer.skip_token ();
-      return std::unique_ptr<AST::RestPattern> (
-	new AST::RestPattern (t->get_locus ()));
-    case REF:
-    case MUT:
-      return parse_identifier_pattern ();
-    case IDENTIFIER:
-      /* if identifier with no scope resolution afterwards, identifier
-       * pattern. if scope resolution afterwards, path pattern (or range
-       * pattern or struct pattern or tuple struct pattern) or macro
-       * invocation */
-      return parse_ident_leading_pattern ();
-    case AMP:
-    case LOGICAL_AND:
-      // reference pattern
-      return parse_reference_pattern ();
-    case LEFT_PAREN:
-      // tuple pattern or grouped pattern
-      return parse_grouped_or_tuple_pattern ();
-    case LEFT_SQUARE:
-      // slice pattern
-      return parse_slice_pattern ();
-    case LEFT_SHIFT:
-    case LEFT_ANGLE:
-      {
-	// qualified path in expression or qualified range pattern bound
-	AST::QualifiedPathInExpression path
-	  = parse_qualified_path_in_expression ();
-
-	if (lexer.peek_token ()->get_id () == DOT_DOT_EQ
-	    || lexer.peek_token ()->get_id () == ELLIPSIS
-	    || lexer.peek_token ()->get_id () == DOT_DOT)
-	  {
-	    // qualified range pattern bound, so parse rest of range pattern
-	    AST::RangeKind kind
-	      = AST::tokenid_to_rangekind (lexer.peek_token ()->get_id ());
-	    lexer.skip_token ();
-
-	    std::unique_ptr<AST::RangePatternBoundQualPath> lower_bound (
-	      new AST::RangePatternBoundQualPath (std::move (path)));
-	    std::unique_ptr<AST::RangePatternBound> upper_bound
-	      = parse_range_pattern_bound ();
-
-	    return std::unique_ptr<AST::RangePattern> (
-	      new AST::RangePattern (std::move (lower_bound),
-				     std::move (upper_bound), kind,
-				     t->get_locus ()));
-	  }
-	else
-	  {
-	    // just qualified path in expression
-	    return std::unique_ptr<AST::QualifiedPathInExpression> (
-	      new AST::QualifiedPathInExpression (std::move (path)));
-	  }
-      }
-    case SUPER:
-    case SELF:
-    case SELF_ALIAS:
-    case CRATE:
-    case SCOPE_RESOLUTION:
-    case DOLLAR_SIGN:
-      {
-	// path in expression or range pattern bound
-	AST::PathInExpression path = parse_path_in_expression ();
-
-	const_TokenPtr next = lexer.peek_token ();
-	switch (next->get_id ())
-	  {
-	  case DOT_DOT_EQ:
-	  case DOT_DOT:
-	  case ELLIPSIS:
-	    {
-	      // qualified range pattern bound, so parse rest of range pattern
-	      AST::RangeKind kind = AST::tokenid_to_rangekind (next->get_id ());
-	      lexer.skip_token ();
-
-	      std::unique_ptr<AST::RangePatternBoundPath> lower_bound (
-		new AST::RangePatternBoundPath (std::move (path)));
-	      std::unique_ptr<AST::RangePatternBound> upper_bound
-		= parse_range_pattern_bound ();
-
-	      return std::unique_ptr<AST::RangePattern> (
-		new AST::RangePattern (std::move (lower_bound),
-				       std::move (upper_bound), kind,
-				       next->get_locus ()));
-	    }
-	  case EXCLAM:
-	    return parse_macro_invocation_partial (std::move (path),
-						   AST::AttrVec ());
-	  case LEFT_PAREN:
-	    {
-	      // tuple struct
-	      lexer.skip_token ();
-
-	      // parse items
-	      std::unique_ptr<AST::TupleStructItems> items
-		= parse_tuple_struct_items ();
-	      if (items == nullptr)
-		{
-		  Error error (lexer.peek_token ()->get_locus (),
-			       "failed to parse tuple struct items");
-		  add_error (std::move (error));
-
-		  return nullptr;
-		}
-
-	      if (!skip_token (RIGHT_PAREN))
-		{
-		  return nullptr;
-		}
-
-	      return std::unique_ptr<AST::TupleStructPattern> (
-		new AST::TupleStructPattern (std::move (path),
-					     std::move (items)));
-	    }
-	  case LEFT_CURLY:
-	    {
-	      // struct
-	      lexer.skip_token ();
-
-	      // parse elements (optional)
-	      AST::StructPatternElements elems = parse_struct_pattern_elems ();
-
-	      if (!skip_token (RIGHT_CURLY))
-		{
-		  return nullptr;
-		}
-
-	      return std::unique_ptr<AST::StructPattern> (
-		new AST::StructPattern (std::move (path), t->get_locus (),
-					std::move (elems)));
-	    }
-	  default:
-	    // assume path in expression
-	    return std::unique_ptr<AST::PathInExpression> (
-	      new AST::PathInExpression (std::move (path)));
-	  }
-      }
-    default:
-      add_error (Error (t->get_locus (), "unexpected token %qs in pattern",
-			t->get_token_description ()));
-
-      return nullptr;
-    }
-}
-
-// Parses a single or double reference pattern.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ReferencePattern>
-Parser<ManagedTokenSource>::parse_reference_pattern ()
-{
-  // parse double or single ref
-  bool is_double_ref = false;
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case AMP:
-      // still false
-      lexer.skip_token ();
-      break;
-    case LOGICAL_AND:
-      is_double_ref = true;
-      lexer.skip_token ();
-      break;
-    default:
-      add_error (Error (t->get_locus (),
-			"unexpected token %qs in reference pattern",
-			t->get_token_description ()));
-
-      return nullptr;
-    }
-
-  // parse mut (if it exists)
-  bool is_mut = false;
-  if (lexer.peek_token ()->get_id () == MUT)
-    {
-      is_mut = true;
-      lexer.skip_token ();
-    }
-
-  // parse pattern to get reference of (required)
-  std::unique_ptr<AST::Pattern> pattern = parse_pattern_no_alt ();
-  if (pattern == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse pattern in reference pattern");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  return std::unique_ptr<AST::ReferencePattern> (
-    new AST::ReferencePattern (std::move (pattern), is_mut, is_double_ref,
-			       t->get_locus ()));
-}
-
-/* Parses a grouped pattern or tuple pattern. Prefers grouped over tuple if
- * only a single element with no commas. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Pattern>
-Parser<ManagedTokenSource>::parse_grouped_or_tuple_pattern ()
-{
-  location_t paren_locus = lexer.peek_token ()->get_locus ();
-  skip_token (LEFT_PAREN);
-
-  // detect '..' token (ranged with no lower range)
-  if (lexer.peek_token ()->get_id () == DOT_DOT)
-    {
-      lexer.skip_token ();
-
-      // parse new patterns while next token is a comma
-      std::vector<std::unique_ptr<AST::Pattern>> patterns;
-
-      const_TokenPtr t = lexer.peek_token ();
-      while (t->get_id () == COMMA)
-	{
-	  lexer.skip_token ();
-
-	  // break if next token is ')'
-	  if (lexer.peek_token ()->get_id () == RIGHT_PAREN)
-	    {
-	      break;
-	    }
-
-	  // parse pattern, which is required
-	  std::unique_ptr<AST::Pattern> pattern = parse_pattern ();
-	  if (pattern == nullptr)
-	    {
-	      Error error (
-		lexer.peek_token ()->get_locus (),
-		"failed to parse pattern inside ranged tuple pattern");
-	      add_error (std::move (error));
-
-	      // skip somewhere?
-	      return nullptr;
-	    }
-	  patterns.push_back (std::move (pattern));
-
-	  t = lexer.peek_token ();
-	}
-
-      if (!skip_token (RIGHT_PAREN))
-	{
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      // create tuple pattern items with only upper pattern items
-      std::unique_ptr<AST::TuplePatternItemsHasRest> items (
-	new AST::TuplePatternItemsHasRest (
-	  std::vector<std::unique_ptr<AST::Pattern>> (), std::move (patterns)));
-      return std::unique_ptr<AST::TuplePattern> (
-	new AST::TuplePattern (std::move (items), paren_locus));
-    }
-  else if (lexer.peek_token ()->get_id () == RIGHT_PAREN)
-    {
-      skip_token (RIGHT_PAREN);
-      auto items = std::unique_ptr<AST::TuplePatternItemsNoRest> (
-	new AST::TuplePatternItemsNoRest (
-	  std::vector<std::unique_ptr<AST::Pattern>> ()));
-      return std::unique_ptr<AST::TuplePattern> (
-	new AST::TuplePattern (std::move (items), paren_locus));
-    }
-
-  // parse initial pattern (required)
-  std::unique_ptr<AST::Pattern> initial_pattern = parse_pattern ();
-  if (initial_pattern == nullptr)
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse pattern in grouped or tuple pattern");
-      add_error (std::move (error));
-
-      return nullptr;
-    }
-
-  // branch on whether next token is a comma or not
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case RIGHT_PAREN:
-      // grouped pattern
-      lexer.skip_token ();
-
-      return std::unique_ptr<AST::GroupedPattern> (
-	new AST::GroupedPattern (std::move (initial_pattern), paren_locus));
-    case COMMA:
-      {
-	// tuple pattern
-	lexer.skip_token ();
-
-	// create vector of patterns
-	std::vector<std::unique_ptr<AST::Pattern>> patterns;
-	patterns.push_back (std::move (initial_pattern));
-
-	t = lexer.peek_token ();
-	while (t->get_id () != RIGHT_PAREN && t->get_id () != DOT_DOT)
-	  {
-	    // parse pattern (required)
-	    std::unique_ptr<AST::Pattern> pattern = parse_pattern ();
-	    if (pattern == nullptr)
-	      {
-		Error error (t->get_locus (),
-			     "failed to parse pattern in tuple pattern");
-		add_error (std::move (error));
-
-		return nullptr;
-	      }
-	    patterns.push_back (std::move (pattern));
-
-	    if (lexer.peek_token ()->get_id () != COMMA)
-	      break;
-
-	    lexer.skip_token ();
-	    t = lexer.peek_token ();
-	  }
-
-	t = lexer.peek_token ();
-	if (t->get_id () == RIGHT_PAREN)
-	  {
-	    // non-ranged tuple pattern
-	    lexer.skip_token ();
-
-	    std::unique_ptr<AST::TuplePatternItemsNoRest> items (
-	      new AST::TuplePatternItemsNoRest (std::move (patterns)));
-	    return std::unique_ptr<AST::TuplePattern> (
-	      new AST::TuplePattern (std::move (items), paren_locus));
-	  }
-	else if (t->get_id () == DOT_DOT)
-	  {
-	    // ranged tuple pattern
-	    lexer.skip_token ();
-
-	    // parse upper patterns
-	    std::vector<std::unique_ptr<AST::Pattern>> upper_patterns;
-	    t = lexer.peek_token ();
-	    while (t->get_id () == COMMA)
-	      {
-		lexer.skip_token ();
-
-		// break if end
-		if (lexer.peek_token ()->get_id () == RIGHT_PAREN)
-		  break;
-
-		// parse pattern (required)
-		std::unique_ptr<AST::Pattern> pattern = parse_pattern ();
-		if (pattern == nullptr)
-		  {
-		    Error error (lexer.peek_token ()->get_locus (),
-				 "failed to parse pattern in tuple pattern");
-		    add_error (std::move (error));
-
-		    return nullptr;
-		  }
-		upper_patterns.push_back (std::move (pattern));
-
-		t = lexer.peek_token ();
-	      }
-
-	    if (!skip_token (RIGHT_PAREN))
-	      {
-		return nullptr;
-	      }
-
-	    std::unique_ptr<AST::TuplePatternItemsHasRest> items (
-	      new AST::TuplePatternItemsHasRest (std::move (patterns),
-						 std::move (upper_patterns)));
-	    return std::unique_ptr<AST::TuplePattern> (
-	      new AST::TuplePattern (std::move (items), paren_locus));
-	  }
-	else
-	  {
-	    // some kind of error
-	    Error error (t->get_locus (),
-			 "failed to parse tuple pattern (probably) or maybe "
-			 "grouped pattern");
-	    add_error (std::move (error));
-
-	    return nullptr;
-	  }
-      }
-    default:
-      // error
-      add_error (Error (t->get_locus (),
-			"unrecognised token %qs in grouped or tuple pattern "
-			"after first pattern",
-			t->get_token_description ()));
-
-      return nullptr;
-    }
-}
-
-/* Parses a slice pattern that can match arrays or slices. Parses the square
- * brackets too. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::SlicePattern>
-Parser<ManagedTokenSource>::parse_slice_pattern ()
-{
-  location_t square_locus = lexer.peek_token ()->get_locus ();
-  std::vector<std::unique_ptr<AST::Pattern>> patterns;
-  tl::optional<std::vector<std::unique_ptr<AST::Pattern>>> upper_patterns
-    = tl::nullopt;
-
-  // lambda function to determine which vector to push new patterns into
-  auto get_pattern_ref
-    = [&] () -> std::vector<std::unique_ptr<AST::Pattern>> & {
-    return upper_patterns.has_value () ? upper_patterns.value () : patterns;
-  };
-
-  skip_token (LEFT_SQUARE);
-
-  if (lexer.peek_token ()->get_id () == RIGHT_SQUARE)
-    {
-      skip_token (RIGHT_SQUARE);
-      std::unique_ptr<AST::SlicePatternItemsNoRest> items (
-	new AST::SlicePatternItemsNoRest (std::move (patterns)));
-      return std::unique_ptr<AST::SlicePattern> (
-	new AST::SlicePattern (std::move (items), square_locus));
-    }
-
-  // parse initial pattern (required)
-  if (lexer.peek_token ()->get_id () == DOT_DOT)
-    {
-      lexer.skip_token ();
-      upper_patterns = std::vector<std::unique_ptr<AST::Pattern>> ();
-    }
-  else
-    {
-      // Not a rest pattern `..`, parse normally
-      std::unique_ptr<AST::Pattern> initial_pattern = parse_pattern ();
-      if (initial_pattern == nullptr)
-	{
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse initial pattern in slice pattern");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-
-      patterns.push_back (std::move (initial_pattern));
-    }
-
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () == COMMA)
-    {
-      lexer.skip_token ();
-
-      // break if end bracket
-      if (lexer.peek_token ()->get_id () == RIGHT_SQUARE)
-	break;
-
-      if (lexer.peek_token ()->get_id () == DOT_DOT)
-	{
-	  if (upper_patterns.has_value ())
-	    {
-	      // DOT_DOT has been parsed before
-	      Error error (lexer.peek_token ()->get_locus (), "%s",
-			   "`..` can only be used once per slice pattern");
-	      add_error (std::move (error));
-
-	      return nullptr;
-	    }
-	  upper_patterns = std::vector<std::unique_ptr<AST::Pattern>> ();
-	  lexer.skip_token ();
-	  t = lexer.peek_token ();
-	  continue;
-	}
-
-      // parse pattern (required)
-      std::unique_ptr<AST::Pattern> pattern = parse_pattern ();
-      if (pattern == nullptr)
-	{
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse pattern in slice pattern");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-      get_pattern_ref ().push_back (std::move (pattern));
-
-      t = lexer.peek_token ();
-    }
-
-  if (!skip_token (RIGHT_SQUARE))
-    {
-      return nullptr;
-    }
-
-  if (upper_patterns.has_value ())
-    {
-      // Slice pattern with rest
-      std::unique_ptr<AST::SlicePatternItemsHasRest> items (
-	new AST::SlicePatternItemsHasRest (
-	  std::move (patterns), std::move (upper_patterns.value ())));
-      return std::unique_ptr<AST::SlicePattern> (
-	new AST::SlicePattern (std::move (items), square_locus));
-    }
-
-  // Rest-less slice pattern
-  std::unique_ptr<AST::SlicePatternItemsNoRest> items (
-    new AST::SlicePatternItemsNoRest (std::move (patterns)));
-  return std::unique_ptr<AST::SlicePattern> (
-    new AST::SlicePattern (std::move (items), square_locus));
-}
-
-/* Parses an identifier pattern (pattern that binds a value matched to a
- * variable). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::IdentifierPattern>
-Parser<ManagedTokenSource>::parse_identifier_pattern ()
-{
-  location_t locus = lexer.peek_token ()->get_locus ();
-
-  bool has_ref = false;
-  if (lexer.peek_token ()->get_id () == REF)
-    {
-      has_ref = true;
-      lexer.skip_token ();
-
-      // DEBUG
-      rust_debug ("parsed ref in identifier pattern");
-    }
-
-  bool has_mut = false;
-  if (lexer.peek_token ()->get_id () == MUT)
-    {
-      has_mut = true;
-      lexer.skip_token ();
-    }
-
-  // parse identifier (required)
-  const_TokenPtr ident_tok = expect_token (IDENTIFIER);
-  if (ident_tok == nullptr)
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-  Identifier ident{ident_tok};
-
-  // DEBUG
-  rust_debug ("parsed identifier in identifier pattern");
-
-  // parse optional pattern binding thing
-  std::unique_ptr<AST::Pattern> bind_pattern = nullptr;
-  if (lexer.peek_token ()->get_id () == PATTERN_BIND)
-    {
-      lexer.skip_token ();
-
-      // parse required pattern to bind
-      bind_pattern = parse_pattern_no_alt ();
-      if (bind_pattern == nullptr)
-	{
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse pattern to bind in identifier pattern");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-    }
-
-  // DEBUG
-  rust_debug ("about to return identifier pattern");
-
-  return std::unique_ptr<AST::IdentifierPattern> (
-    new AST::IdentifierPattern (std::move (ident), locus, has_ref, has_mut,
-				std::move (bind_pattern)));
-}
-
-/* Parses a pattern that opens with an identifier. This includes identifier
- * patterns, path patterns (and derivatives such as struct patterns, tuple
- * struct patterns, and macro invocations), and ranges. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Pattern>
-Parser<ManagedTokenSource>::parse_ident_leading_pattern ()
-{
-  // ensure first token is actually identifier
-  const_TokenPtr initial_tok = lexer.peek_token ();
-  if (initial_tok->get_id () != IDENTIFIER)
-    {
-      return nullptr;
-    }
-
-  // save initial identifier as it may be useful (but don't skip)
-  std::string initial_ident = initial_tok->get_str ();
-
-  // parse next tokens as a PathInExpression
-  AST::PathInExpression path = parse_path_in_expression ();
-
-  // branch on next token
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case EXCLAM:
-      return parse_macro_invocation_partial (std::move (path), AST::AttrVec ());
-    case LEFT_PAREN:
-      {
-	// tuple struct
-	lexer.skip_token ();
-
-	// DEBUG
-	rust_debug ("parsing tuple struct pattern");
-
-	// parse items
-	std::unique_ptr<AST::TupleStructItems> items
-	  = parse_tuple_struct_items ();
-	if (items == nullptr)
-	  {
-	    Error error (lexer.peek_token ()->get_locus (),
-			 "failed to parse tuple struct items");
-	    add_error (std::move (error));
-
-	    return nullptr;
-	  }
-
-	// DEBUG
-	rust_debug ("successfully parsed tuple struct items");
-
-	if (!skip_token (RIGHT_PAREN))
-	  {
-	    return nullptr;
-	  }
-
-	// DEBUG
-	rust_debug ("successfully parsed tuple struct pattern");
-
-	return std::unique_ptr<AST::TupleStructPattern> (
-	  new AST::TupleStructPattern (std::move (path), std::move (items)));
-      }
-    case LEFT_CURLY:
-      {
-	// struct
-	lexer.skip_token ();
-
-	// parse elements (optional)
-	AST::StructPatternElements elems = parse_struct_pattern_elems ();
-
-	if (!skip_token (RIGHT_CURLY))
-	  {
-	    return nullptr;
-	  }
-
-	// DEBUG
-	rust_debug ("successfully parsed struct pattern");
-
-	return std::unique_ptr<AST::StructPattern> (
-	  new AST::StructPattern (std::move (path), initial_tok->get_locus (),
-				  std::move (elems)));
-      }
-    case DOT_DOT_EQ:
-    case DOT_DOT:
-    case ELLIPSIS:
-      {
-	// range
-	AST::RangeKind kind
-	  = AST::tokenid_to_rangekind (lexer.peek_token ()->get_id ());
-
-	lexer.skip_token ();
-
-	std::unique_ptr<AST::RangePatternBoundPath> lower_bound (
-	  new AST::RangePatternBoundPath (std::move (path)));
-	std::unique_ptr<AST::RangePatternBound> upper_bound
-	  = parse_range_pattern_bound ();
-
-	return std::unique_ptr<AST::RangePattern> (
-	  new AST::RangePattern (std::move (lower_bound),
-				 std::move (upper_bound), kind,
-				 t->get_locus ()));
-      }
-    case PATTERN_BIND:
-      {
-	// only allow on single-segment paths
-	if (path.is_single_segment ())
-	  {
-	    // identifier with pattern bind
-	    lexer.skip_token ();
-
-	    std::unique_ptr<AST::Pattern> bind_pattern
-	      = parse_pattern_no_alt ();
-	    if (bind_pattern == nullptr)
-	      {
-		Error error (
-		  t->get_locus (),
-		  "failed to parse pattern to bind to identifier pattern");
-		add_error (std::move (error));
-
-		return nullptr;
-	      }
-	    return std::unique_ptr<AST::IdentifierPattern> (
-	      new AST::IdentifierPattern (std::move (initial_ident),
-					  initial_tok->get_locus (), false,
-					  false, std::move (bind_pattern)));
-	  }
-	Error error (
-	  t->get_locus (),
-	  "failed to parse pattern bind to a path, not an identifier");
-	add_error (std::move (error));
-
-	return nullptr;
-      }
-    default:
-      // assume identifier if single segment
-      if (path.is_single_segment ())
-	{
-	  return std::unique_ptr<AST::IdentifierPattern> (
-	    new AST::IdentifierPattern (std::move (initial_ident),
-					initial_tok->get_locus ()));
-	}
-      // return path otherwise
-      return std::unique_ptr<AST::PathInExpression> (
-	new AST::PathInExpression (std::move (path)));
-    }
-}
-
 // Parses tuple struct items if they exist. Does not parse parentheses.
 template <typename ManagedTokenSource>
 std::unique_ptr<AST::TupleStructItems>
@@ -11387,198 +6962,6 @@ Parser<ManagedTokenSource>::parse_tuple_struct_items ()
 			"unexpected token %qs in tuple struct items",
 			t->get_token_description ()));
 
-      return nullptr;
-    }
-}
-
-// Parses struct pattern elements if they exist.
-template <typename ManagedTokenSource>
-AST::StructPatternElements
-Parser<ManagedTokenSource>::parse_struct_pattern_elems ()
-{
-  std::vector<std::unique_ptr<AST::StructPatternField>> fields;
-
-  AST::AttrVec etc_attrs;
-  bool has_rest = false;
-
-  // try parsing struct pattern fields
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () != RIGHT_CURLY)
-    {
-      AST::AttrVec outer_attrs = parse_outer_attributes ();
-
-      // parse etc (must be last in struct pattern, so breaks)
-      if (lexer.peek_token ()->get_id () == DOT_DOT)
-	{
-	  lexer.skip_token ();
-	  etc_attrs = std::move (outer_attrs);
-	  has_rest = true;
-	  break;
-	}
-
-      std::unique_ptr<AST::StructPatternField> field
-	= parse_struct_pattern_field_partial (std::move (outer_attrs));
-      if (field == nullptr)
-	{
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse struct pattern field");
-	  add_error (std::move (error));
-
-	  // skip after somewhere?
-	  return AST::StructPatternElements::create_empty ();
-	}
-      fields.push_back (std::move (field));
-
-      if (lexer.peek_token ()->get_id () != COMMA)
-	break;
-
-      // skip comma
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-    }
-
-  if (has_rest)
-    return AST::StructPatternElements (std::move (fields),
-				       std::move (etc_attrs));
-  else
-    return AST::StructPatternElements (std::move (fields));
-}
-
-/* Parses a struct pattern field (tuple index/pattern, identifier/pattern, or
- * identifier). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::StructPatternField>
-Parser<ManagedTokenSource>::parse_struct_pattern_field ()
-{
-  // parse outer attributes (if they exist)
-  AST::AttrVec outer_attrs = parse_outer_attributes ();
-
-  return parse_struct_pattern_field_partial (std::move (outer_attrs));
-}
-
-/* Parses a struct pattern field (tuple index/pattern, identifier/pattern, or
- * identifier), with outer attributes passed in. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::StructPatternField>
-Parser<ManagedTokenSource>::parse_struct_pattern_field_partial (
-  AST::AttrVec outer_attrs)
-{
-  // branch based on next token
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case INT_LITERAL:
-      {
-	// tuple index
-	std::string index_str = t->get_str ();
-	int index = atoi (index_str.c_str ());
-
-	lexer.skip_token ();
-
-	if (!skip_token (COLON))
-	  {
-	    return nullptr;
-	  }
-
-	// parse required pattern
-	std::unique_ptr<AST::Pattern> pattern = parse_pattern ();
-	if (pattern == nullptr)
-	  {
-	    Error error (
-	      t->get_locus (),
-	      "failed to parse pattern in tuple index struct pattern field");
-	    add_error (std::move (error));
-
-	    return nullptr;
-	  }
-
-	return std::unique_ptr<AST::StructPatternFieldTuplePat> (
-	  new AST::StructPatternFieldTuplePat (index, std::move (pattern),
-					       std::move (outer_attrs),
-					       t->get_locus ()));
-      }
-    case IDENTIFIER:
-      // identifier-pattern OR only identifier
-      // branch on next token
-      switch (lexer.peek_token (1)->get_id ())
-	{
-	case COLON:
-	  {
-	    // identifier-pattern
-	    Identifier ident{t};
-	    lexer.skip_token ();
-
-	    skip_token (COLON);
-
-	    // parse required pattern
-	    std::unique_ptr<AST::Pattern> pattern = parse_pattern ();
-	    if (pattern == nullptr)
-	      {
-		Error error (t->get_locus (),
-			     "failed to parse pattern in struct pattern field");
-		add_error (std::move (error));
-
-		return nullptr;
-	      }
-
-	    return std::unique_ptr<AST::StructPatternFieldIdentPat> (
-	      new AST::StructPatternFieldIdentPat (std::move (ident),
-						   std::move (pattern),
-						   std::move (outer_attrs),
-						   t->get_locus ()));
-	  }
-	case COMMA:
-	case RIGHT_CURLY:
-	  {
-	    // identifier only
-	    Identifier ident = {t};
-	    lexer.skip_token ();
-
-	    return std::unique_ptr<AST::StructPatternFieldIdent> (
-	      new AST::StructPatternFieldIdent (std::move (ident), false, false,
-						std::move (outer_attrs),
-						t->get_locus ()));
-	  }
-	default:
-	  // error
-	  add_error (Error (t->get_locus (),
-			    "unrecognised token %qs in struct pattern field",
-			    t->get_token_description ()));
-
-	  return nullptr;
-	}
-    case REF:
-    case MUT:
-      {
-	// only identifier
-	bool has_ref = false;
-	if (t->get_id () == REF)
-	  {
-	    has_ref = true;
-	    lexer.skip_token ();
-	  }
-
-	bool has_mut = false;
-	if (lexer.peek_token ()->get_id () == MUT)
-	  {
-	    has_mut = true;
-	    lexer.skip_token ();
-	  }
-
-	const_TokenPtr ident_tok = expect_token (IDENTIFIER);
-	if (ident_tok == nullptr)
-	  {
-	    return nullptr;
-	  }
-	Identifier ident{ident_tok};
-
-	return std::unique_ptr<AST::StructPatternFieldIdent> (
-	  new AST::StructPatternFieldIdent (std::move (ident), has_ref, has_mut,
-					    std::move (outer_attrs),
-					    t->get_locus ()));
-      }
-    default:
-      // not necessarily an error
       return nullptr;
     }
 }
@@ -11811,2754 +7194,12 @@ Parser<ManagedTokenSource>::parse_stmt_or_expr ()
   return ExprOrStmt (std::move (expr));
 }
 
-// Parses a struct expression field.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::StructExprField>
-Parser<ManagedTokenSource>::parse_struct_expr_field ()
-{
-  AST::AttrVec outer_attrs = parse_outer_attributes ();
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case IDENTIFIER:
-      if (lexer.peek_token (1)->get_id () == COLON)
-	{
-	  // struct expr field with identifier and expr
-	  Identifier ident = {t};
-	  lexer.skip_token (1);
-
-	  // parse expression (required)
-	  std::unique_ptr<AST::Expr> expr = parse_expr ();
-	  if (expr == nullptr)
-	    {
-	      Error error (t->get_locus (),
-			   "failed to parse struct expression field with "
-			   "identifier and expression");
-	      add_error (std::move (error));
-
-	      return nullptr;
-	    }
-
-	  return std::unique_ptr<AST::StructExprFieldIdentifierValue> (
-	    new AST::StructExprFieldIdentifierValue (std::move (ident),
-						     std::move (expr),
-						     std::move (outer_attrs),
-						     t->get_locus ()));
-	}
-      else
-	{
-	  // struct expr field with identifier only
-	  Identifier ident{t};
-	  lexer.skip_token ();
-
-	  return std::unique_ptr<AST::StructExprFieldIdentifier> (
-	    new AST::StructExprFieldIdentifier (std::move (ident),
-						std::move (outer_attrs),
-						t->get_locus ()));
-	}
-    case INT_LITERAL:
-      {
-	// parse tuple index field
-	int index = atoi (t->get_str ().c_str ());
-	lexer.skip_token ();
-
-	if (!skip_token (COLON))
-	  {
-	    // skip somewhere?
-	    return nullptr;
-	  }
-
-	// parse field expression (required)
-	std::unique_ptr<AST::Expr> expr = parse_expr ();
-	if (expr == nullptr)
-	  {
-	    Error error (t->get_locus (),
-			 "failed to parse expr in struct (or enum) expr "
-			 "field with tuple index");
-	    add_error (std::move (error));
-
-	    return nullptr;
-	  }
-
-	return std::unique_ptr<AST::StructExprFieldIndexValue> (
-	  new AST::StructExprFieldIndexValue (index, std::move (expr),
-					      std::move (outer_attrs),
-					      t->get_locus ()));
-      }
-    case DOT_DOT:
-      /* this is a struct base and can't be parsed here, so just return
-       * nothing without erroring */
-
-      return nullptr;
-    default:
-      add_error (
-	Error (t->get_locus (),
-	       "unrecognised token %qs as first token of struct expr field - "
-	       "expected identifier or integer literal",
-	       t->get_token_description ()));
-
-      return nullptr;
-    }
-}
-
-// "Unexpected token" panic mode - flags gcc error at unexpected token
-// TODO: seems to be unused, remove?
-template <typename ManagedTokenSource>
-void
-Parser<ManagedTokenSource>::unexpected_token (const_TokenPtr t)
-{
-  Error error (t->get_locus (), "unexpected token %qs",
-	       t->get_token_description ());
-  add_error (std::move (error));
-}
-
-/* Crappy "error recovery" performed after error by skipping tokens until a
- * semi-colon is found */
-template <typename ManagedTokenSource>
-void
-Parser<ManagedTokenSource>::skip_after_semicolon ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-
-  while (t->get_id () != END_OF_FILE && t->get_id () != SEMICOLON)
-    {
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-    }
-
-  if (t->get_id () == SEMICOLON)
-    lexer.skip_token ();
-}
-
-/* Skips the current token */
-template <typename ManagedTokenSource>
-void
-Parser<ManagedTokenSource>::skip_token ()
-{
-  lexer.skip_token ();
-}
-
-/* Checks if current token has inputted id - skips it and returns true if so,
- * diagnoses an error and returns false otherwise. */
-template <typename ManagedTokenSource>
-bool
-Parser<ManagedTokenSource>::skip_token (TokenId token_id)
-{
-  return expect_token (token_id) != const_TokenPtr ();
-}
-
-/* Checks if current token is similar to inputted token - skips it and returns
- * true if so, diagnoses an error and returns false otherwise. */
-template <typename ManagedTokenSource>
-bool
-Parser<ManagedTokenSource>::skip_token (const_TokenPtr token)
-{
-  return expect_token (token) != const_TokenPtr ();
-}
-
-/* Checks if current token has inputted id - skips it and returns true if so,
- * returns false otherwise without diagnosing an error */
-template <typename ManagedTokenSource>
-bool
-Parser<ManagedTokenSource>::maybe_skip_token (TokenId token_id)
-{
-  if (lexer.peek_token ()->get_id () != token_id)
-    return false;
-  else
-    return skip_token (token_id);
-}
-
-/* Checks the current token - if id is same as expected, skips and returns it,
- * otherwise diagnoses error and returns null. */
-template <typename ManagedTokenSource>
-const_TokenPtr
-Parser<ManagedTokenSource>::expect_token (TokenId token_id)
-{
-  const_TokenPtr t = lexer.peek_token ();
-  if (t->get_id () == token_id)
-    {
-      lexer.skip_token ();
-      return t;
-    }
-  else
-    {
-      Error error (t->get_locus (), "expecting %qs but %qs found",
-		   get_token_description (token_id),
-		   t->get_token_description ());
-      add_error (std::move (error));
-
-      return const_TokenPtr ();
-    }
-}
-
-/* Checks the current token - if same as expected, skips and returns it,
- * otherwise diagnoses error and returns null. */
-template <typename ManagedTokenSource>
-const_TokenPtr
-Parser<ManagedTokenSource>::expect_token (const_TokenPtr token_expect)
-{
-  const_TokenPtr t = lexer.peek_token ();
-  if (t->get_id () == token_expect->get_id ()
-      && (!t->should_have_str () || t->get_str () == token_expect->get_str ()))
-    {
-      lexer.skip_token ();
-      return t;
-    }
-  else
-    {
-      Error error (t->get_locus (), "expecting %qs but %qs found",
-		   token_expect->get_token_description (),
-		   t->get_token_description ());
-      add_error (std::move (error));
-
-      return const_TokenPtr ();
-    }
-}
-
-// Skips all tokens until EOF or }. Don't use.
-template <typename ManagedTokenSource>
-void
-Parser<ManagedTokenSource>::skip_after_end ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-
-  while (t->get_id () != END_OF_FILE && t->get_id () != RIGHT_CURLY)
-    {
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-    }
-
-  if (t->get_id () == RIGHT_CURLY)
-    {
-      lexer.skip_token ();
-    }
-}
-
-/* A slightly more aware error-handler that skips all tokens until it reaches
- * the end of the block scope (i.e. when left curly brackets = right curly
- * brackets). Note: assumes currently in the middle of a block. Use
- * skip_after_next_block to skip based on the assumption that the block
- * has not been entered yet. */
-template <typename ManagedTokenSource>
-void
-Parser<ManagedTokenSource>::skip_after_end_block ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-  int curly_count = 1;
-
-  while (curly_count > 0 && t->get_id () != END_OF_FILE)
-    {
-      switch (t->get_id ())
-	{
-	case LEFT_CURLY:
-	  curly_count++;
-	  break;
-	case RIGHT_CURLY:
-	  curly_count--;
-	  break;
-	default:
-	  break;
-	}
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-    }
-}
-
-/* Skips tokens until the end of the next block. i.e. assumes that the block
- * has not been entered yet. */
-template <typename ManagedTokenSource>
-void
-Parser<ManagedTokenSource>::skip_after_next_block ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-
-  // initial loop - skip until EOF if no left curlies encountered
-  while (t->get_id () != END_OF_FILE && t->get_id () != LEFT_CURLY)
-    {
-      lexer.skip_token ();
-
-      t = lexer.peek_token ();
-    }
-
-  // if next token is left, skip it and then skip after the block ends
-  if (t->get_id () == LEFT_CURLY)
-    {
-      lexer.skip_token ();
-
-      skip_after_end_block ();
-    }
-  // otherwise, do nothing as EOF
-}
-
-/* Skips all tokens until ] (the end of an attribute) - does not skip the ]
- * (as designed for attribute body use) */
-template <typename ManagedTokenSource>
-void
-Parser<ManagedTokenSource>::skip_after_end_attribute ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-
-  while (t->get_id () != RIGHT_SQUARE && t->get_id () != END_OF_FILE)
-    {
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-    }
-
-  // Don't skip the RIGHT_SQUARE token
-}
-
-/* Pratt parser impl of parse_expr. FIXME: this is only provisional and
- * probably will be changed. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::parse_expr (int right_binding_power,
-					AST::AttrVec outer_attrs,
-					ParseRestrictions restrictions)
-{
-  const_TokenPtr current_token = lexer.peek_token ();
-  // Special hack because we are allowed to return nullptr, in that case we
-  // don't want to skip the token, since we don't actually parse it. But if
-  // null isn't allowed it indicates an error, and we want to skip past that.
-  // So return early if it is one of the tokens that ends an expression
-  // (or at least cannot start a new expression).
-  if (restrictions.expr_can_be_null)
-    {
-      TokenId id = current_token->get_id ();
-      if (id == SEMICOLON || id == RIGHT_PAREN || id == RIGHT_CURLY
-	  || id == RIGHT_SQUARE || id == COMMA || id == LEFT_CURLY)
-	return nullptr;
-    }
-
-  ParseRestrictions null_denotation_restrictions = restrictions;
-  null_denotation_restrictions.expr_can_be_stmt = false;
-
-  // parse null denotation (unary part of expression)
-  std::unique_ptr<AST::Expr> expr
-    = null_denotation ({}, null_denotation_restrictions);
-  if (expr == nullptr)
-    return nullptr;
-
-  return left_denotations (std::move (expr), right_binding_power,
-			   std::move (outer_attrs), restrictions);
-}
-
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::left_denotations (std::unique_ptr<AST::Expr> expr,
-					      int right_binding_power,
-					      AST::AttrVec outer_attrs,
-					      ParseRestrictions restrictions)
-{
-  if (expr == nullptr)
-    {
-      // DEBUG
-      rust_debug ("null denotation is null; returning null for parse_expr");
-      return nullptr;
-    }
-
-  const_TokenPtr current_token = lexer.peek_token ();
-
-  if (restrictions.expr_can_be_stmt && !expr->is_expr_without_block ()
-      && current_token->get_id () != DOT
-      && current_token->get_id () != QUESTION_MARK)
-    {
-      rust_debug ("statement expression with block");
-      expr->set_outer_attrs (std::move (outer_attrs));
-      return expr;
-    }
-
-  restrictions.expr_can_be_stmt = false;
-
-  // stop parsing if find lower priority token - parse higher priority first
-  while (right_binding_power < left_binding_power (current_token))
-    {
-      lexer.skip_token ();
-
-      // FIXME attributes should generally be applied to the null denotation.
-      expr = left_denotation (current_token, std::move (expr),
-			      std::move (outer_attrs), restrictions);
-
-      if (expr == nullptr)
-	{
-	  // DEBUG
-	  rust_debug ("left denotation is null; returning null for parse_expr");
-
-	  return nullptr;
-	}
-
-      current_token = lexer.peek_token ();
-    }
-
-  return expr;
-}
-
-// Parse expression with lowest left binding power.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::parse_expr (AST::AttrVec outer_attrs,
-					ParseRestrictions restrictions)
-{
-  return parse_expr (LBP_LOWEST, std::move (outer_attrs), restrictions);
-}
-
-/* Determines action to take when finding token at beginning of expression. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::null_denotation (AST::AttrVec outer_attrs,
-					     ParseRestrictions restrictions)
-{
-  /* note: tok is previous character in input stream, not current one, as
-   * parse_expr skips it before passing it in */
-
-  /* as a Pratt parser (which works by decomposing expressions into a null
-   * denotation and then a left denotation), null denotations handle primaries
-   * and unary operands (but only prefix unary operands) */
-
-  auto tok = lexer.peek_token ();
-
-  switch (tok->get_id ())
-    {
-    case IDENTIFIER:
-    case SELF:
-    case SELF_ALIAS:
-    case DOLLAR_SIGN:
-    case CRATE:
-    case SUPER:
-    case SCOPE_RESOLUTION:
-      {
-	// DEBUG
-	rust_debug ("beginning null denotation identifier handling");
-
-	/* best option: parse as path, then extract identifier, macro,
-	 * struct/enum, or just path info from it */
-	AST::PathInExpression path = parse_path_in_expression ();
-
-	return null_denotation_path (std::move (path), std::move (outer_attrs),
-				     restrictions);
-      }
-    default:
-      if (tok->get_id () == LEFT_SHIFT)
-	{
-	  lexer.split_current_token (LEFT_ANGLE, LEFT_ANGLE);
-	  tok = lexer.peek_token ();
-	}
-
-      lexer.skip_token ();
-      return null_denotation_not_path (std::move (tok), std::move (outer_attrs),
-				       restrictions);
-    }
-}
-
-// Handling of expresions that start with a path for `null_denotation`.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::null_denotation_path (
-  AST::PathInExpression path, AST::AttrVec outer_attrs,
-  ParseRestrictions restrictions)
-{
-  rust_debug ("parsing null denotation after path");
-
-  // HACK: always make "self" by itself a path (regardless of next
-  // tokens)
-  if (path.is_single_segment () && path.get_segments ()[0].is_lower_self_seg ())
-    {
-      // HACK: add outer attrs to path
-      path.set_outer_attrs (std::move (outer_attrs));
-      return std::unique_ptr<AST::PathInExpression> (
-	new AST::PathInExpression (std::move (path)));
-    }
-
-  // branch on next token
-  const_TokenPtr t = lexer.peek_token ();
-  switch (t->get_id ())
-    {
-    case EXCLAM:
-      // macro
-      return parse_macro_invocation_partial (std::move (path),
-					     std::move (outer_attrs));
-    case LEFT_CURLY:
-      {
-	bool not_a_block = lexer.peek_token (1)->get_id () == IDENTIFIER
-			   && (lexer.peek_token (2)->get_id () == COMMA
-			       || (lexer.peek_token (2)->get_id () == COLON
-				   && (lexer.peek_token (4)->get_id () == COMMA
-				       || !Parse::Utils::can_tok_start_type (
-					 lexer.peek_token (3)->get_id ()))));
-
-	/* definitely not a block:
-	 *  path '{' ident ','
-	 *  path '{' ident ':' [anything] ','
-	 *  path '{' ident ':' [not a type]
-	 * otherwise, assume block expr and thus path */
-	// DEBUG
-	rust_debug ("values of lookahead: '%s' '%s' '%s' '%s' ",
-		    lexer.peek_token (1)->get_token_description (),
-		    lexer.peek_token (2)->get_token_description (),
-		    lexer.peek_token (3)->get_token_description (),
-		    lexer.peek_token (4)->get_token_description ());
-
-	rust_debug ("can be struct expr: '%s', not a block: '%s'",
-		    restrictions.can_be_struct_expr ? "true" : "false",
-		    not_a_block ? "true" : "false");
-
-	// struct/enum expr struct
-	if (!restrictions.can_be_struct_expr && !not_a_block)
-	  {
-	    // HACK: add outer attrs to path
-	    path.set_outer_attrs (std::move (outer_attrs));
-	    return std::unique_ptr<AST::PathInExpression> (
-	      new AST::PathInExpression (std::move (path)));
-	  }
-	return parse_struct_expr_struct_partial (std::move (path),
-						 std::move (outer_attrs));
-      }
-    case LEFT_PAREN:
-      // struct/enum expr tuple
-      if (!restrictions.can_be_struct_expr)
-	{
-	  // assume path is returned
-	  // HACK: add outer attributes to path
-	  path.set_outer_attrs (std::move (outer_attrs));
-	  return std::unique_ptr<AST::PathInExpression> (
-	    new AST::PathInExpression (std::move (path)));
-	}
-      return parse_struct_expr_tuple_partial (std::move (path),
-					      std::move (outer_attrs));
-    default:
-      // assume path is returned if not single segment
-      if (path.is_single_segment ())
-	{
-	  // FIXME: This should probably be returned as a path.
-	  /* HACK: may have to become permanent, but this is my current
-	   * identifier expression */
-	  return std::unique_ptr<AST::IdentifierExpr> (new AST::IdentifierExpr (
-	    path.get_segments ()[0].get_ident_segment ().as_string (), {},
-	    path.get_locus ()));
-	}
-      // HACK: add outer attrs to path
-      path.set_outer_attrs (std::move (outer_attrs));
-      return std::unique_ptr<AST::PathInExpression> (
-	new AST::PathInExpression (std::move (path)));
-    }
-  rust_unreachable ();
-}
-
-// Handling of expresions that do not start with a path for `null_denotation`.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::null_denotation_not_path (
-  const_TokenPtr tok, AST::AttrVec outer_attrs, ParseRestrictions restrictions)
-{
-  switch (tok->get_id ())
-    {
-    // FIXME: Handle in null_denotation_path?
-    case LEFT_SHIFT:
-    case LEFT_ANGLE:
-      {
-	// qualified path
-	// HACK: add outer attrs to path
-	AST::QualifiedPathInExpression path
-	  = parse_qualified_path_in_expression (tok->get_locus ());
-	path.set_outer_attrs (std::move (outer_attrs));
-	return std::unique_ptr<AST::QualifiedPathInExpression> (
-	  new AST::QualifiedPathInExpression (std::move (path)));
-      }
-    // FIXME: delegate to parse_literal_expr instead? would have to rejig
-    // tokens and whatever.
-    // FIXME: for literal exprs, outer attrs should be passed in, and later
-    // error if it does not make up the entire statement.
-    case INT_LITERAL:
-      // we should check the range, but ignore for now
-      // encode as int?
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (tok->get_str (), AST::Literal::INT,
-			      tok->get_type_hint (), {}, tok->get_locus ()));
-    case FLOAT_LITERAL:
-      // encode as float?
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (tok->get_str (), AST::Literal::FLOAT,
-			      tok->get_type_hint (), {}, tok->get_locus ()));
-    case STRING_LITERAL:
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (tok->get_str (), AST::Literal::STRING,
-			      tok->get_type_hint (), {}, tok->get_locus ()));
-    case BYTE_STRING_LITERAL:
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (tok->get_str (), AST::Literal::BYTE_STRING,
-			      tok->get_type_hint (), {}, tok->get_locus ()));
-    case RAW_STRING_LITERAL:
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (tok->get_str (), AST::Literal::RAW_STRING,
-			      tok->get_type_hint (), {}, tok->get_locus ()));
-    case CHAR_LITERAL:
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (tok->get_str (), AST::Literal::CHAR,
-			      tok->get_type_hint (), {}, tok->get_locus ()));
-    case BYTE_CHAR_LITERAL:
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (tok->get_str (), AST::Literal::BYTE,
-			      tok->get_type_hint (), {}, tok->get_locus ()));
-    case TRUE_LITERAL:
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (Values::Keywords::TRUE_LITERAL,
-			      AST::Literal::BOOL, tok->get_type_hint (), {},
-			      tok->get_locus ()));
-    case FALSE_LITERAL:
-      return std::unique_ptr<AST::LiteralExpr> (
-	new AST::LiteralExpr (Values::Keywords::FALSE_LITERAL,
-			      AST::Literal::BOOL, tok->get_type_hint (), {},
-			      tok->get_locus ()));
-    case LEFT_PAREN:
-      return parse_grouped_or_tuple_expr (std::move (outer_attrs),
-					  tok->get_locus ());
-
-    /*case PLUS: { // unary plus operator
-	// invoke parse_expr recursively with appropriate priority, etc. for
-    below AST::Expr* expr = parse_expr(LBP_UNARY_PLUS);
-
-	if (expr == nullptr)
-	    return nullptr;
-	// can only apply to integer and float expressions
-	if (expr->get_type() != integer_type_node || expr->get_type() !=
-    float_type_node) { rust_error_at(tok->get_locus(), "operand of unary
-    plus must be int or float but it is %s", print_type(expr->get_type()));
-    return nullptr;
-	}
-
-	return Tree(expr, tok->get_locus());
-    }*/
-    // Rust has no unary plus operator
-    case MINUS:
-      { // unary minus
-	ParseRestrictions entered_from_unary;
-	entered_from_unary.entered_from_unary = true;
-	if (!restrictions.can_be_struct_expr)
-	  entered_from_unary.can_be_struct_expr = false;
-	std::unique_ptr<AST::Expr> expr
-	  = parse_expr (LBP_UNARY_MINUS, {}, entered_from_unary);
-
-	if (expr == nullptr)
-	  return nullptr;
-	// can only apply to integer and float expressions
-	/*if (expr.get_type() != integer_type_node || expr.get_type() !=
-	float_type_node) { rust_error_at(tok->get_locus(), "operand of unary
-	minus must be int or float but it is %s",
-	print_type(expr.get_type())); return Tree::error();
-	}*/
-	/* FIXME: when implemented the "get type" method on expr, ensure it is
-	 * int or float type (except unsigned int). Actually, this would
-	 * probably have to be done in semantic analysis (as type checking).
-	 */
-
-	/* FIXME: allow outer attributes on these expressions by having an
-	 * outer attrs parameter in function*/
-	return std::unique_ptr<AST::NegationExpr> (
-	  new AST::NegationExpr (std::move (expr), NegationOperator::NEGATE,
-				 std::move (outer_attrs), tok->get_locus ()));
-      }
-    case EXCLAM:
-      { // logical or bitwise not
-	ParseRestrictions entered_from_unary;
-	entered_from_unary.entered_from_unary = true;
-	if (!restrictions.can_be_struct_expr)
-	  entered_from_unary.can_be_struct_expr = false;
-	std::unique_ptr<AST::Expr> expr
-	  = parse_expr (LBP_UNARY_EXCLAM, {}, entered_from_unary);
-
-	if (expr == nullptr)
-	  return nullptr;
-	// can only apply to boolean expressions
-	/*if (expr.get_type() != boolean_type_node) {
-	    rust_error_at(tok->get_locus(),
-	      "operand of logical not must be a boolean but it is %s",
-	      print_type(expr.get_type()));
-	    return Tree::error();
-	}*/
-	/* FIXME: type checking for boolean or integer expressions in semantic
-	 * analysis */
-
-	// FIXME: allow outer attributes on these expressions
-	return std::unique_ptr<AST::NegationExpr> (
-	  new AST::NegationExpr (std::move (expr), NegationOperator::NOT,
-				 std::move (outer_attrs), tok->get_locus ()));
-      }
-    case ASTERISK:
-      {
-	/* pointer dereference only - HACK: as struct expressions should
-	 * always be value expressions, cannot be dereferenced */
-	ParseRestrictions entered_from_unary;
-	entered_from_unary.entered_from_unary = true;
-	entered_from_unary.can_be_struct_expr = false;
-	std::unique_ptr<AST::Expr> expr
-	  = parse_expr (LBP_UNARY_ASTERISK, {}, entered_from_unary);
-	// FIXME: allow outer attributes on expression
-	return std::unique_ptr<AST::DereferenceExpr> (
-	  new AST::DereferenceExpr (std::move (expr), std::move (outer_attrs),
-				    tok->get_locus ()));
-      }
-    case AMP:
-      {
-	// (single) "borrow" expression - shared (mutable) or immutable
-	std::unique_ptr<AST::Expr> expr = nullptr;
-	Mutability mutability = Mutability::Imm;
-	bool raw_borrow = false;
-
-	ParseRestrictions entered_from_unary;
-	entered_from_unary.entered_from_unary = true;
-	if (!restrictions.can_be_struct_expr)
-	  entered_from_unary.can_be_struct_expr = false;
-
-	auto is_mutability = [] (const_TokenPtr token) {
-	  return token->get_id () == CONST || token->get_id () == MUT;
-	};
-
-	auto t = lexer.peek_token ();
-	// Weak raw keyword, we look (1) ahead and treat it as an identifier if
-	// there is no mut nor const.
-	if (t->get_id () == IDENTIFIER
-	    && t->get_str () == Values::WeakKeywords::RAW
-	    && is_mutability (lexer.peek_token (1)))
-	  {
-	    lexer.skip_token ();
-	    switch (lexer.peek_token ()->get_id ())
-	      {
-	      case MUT:
-		mutability = Mutability::Mut;
-		break;
-	      case CONST:
-		mutability = Mutability::Imm;
-		break;
-	      default:
-		rust_error_at (lexer.peek_token ()->get_locus (),
-			       "raw borrow should be either const or mut");
-	      }
-	    lexer.skip_token ();
-	    expr = parse_expr (LBP_UNARY_AMP_MUT, {}, entered_from_unary);
-	    raw_borrow = true;
-	  }
-	else if (t->get_id () == MUT)
-	  {
-	    lexer.skip_token ();
-	    expr = parse_expr (LBP_UNARY_AMP_MUT, {}, entered_from_unary);
-	    mutability = Mutability::Mut;
-	    raw_borrow = false;
-	  }
-	else
-	  {
-	    expr = parse_expr (LBP_UNARY_AMP, {}, entered_from_unary);
-	    raw_borrow = false;
-	  }
-
-	// FIXME: allow outer attributes on expression
-	return std::unique_ptr<AST::BorrowExpr> (
-	  new AST::BorrowExpr (std::move (expr), mutability, raw_borrow, false,
-			       std::move (outer_attrs), tok->get_locus ()));
-      }
-    case LOGICAL_AND:
-      {
-	// (double) "borrow" expression - shared (mutable) or immutable
-	std::unique_ptr<AST::Expr> expr = nullptr;
-	Mutability mutability = Mutability::Imm;
-
-	ParseRestrictions entered_from_unary;
-	entered_from_unary.entered_from_unary = true;
-
-	if (lexer.peek_token ()->get_id () == MUT)
-	  {
-	    lexer.skip_token ();
-	    expr = parse_expr (LBP_UNARY_AMP_MUT, {}, entered_from_unary);
-	    mutability = Mutability::Mut;
-	  }
-	else
-	  {
-	    expr = parse_expr (LBP_UNARY_AMP, {}, entered_from_unary);
-	    mutability = Mutability::Imm;
-	  }
-
-	// FIXME: allow outer attributes on expression
-	return std::unique_ptr<AST::BorrowExpr> (
-	  new AST::BorrowExpr (std::move (expr), mutability, false, true,
-			       std::move (outer_attrs), tok->get_locus ()));
-      }
-    case OR:
-    case PIPE:
-    case MOVE:
-      // closure expression
-      return parse_closure_expr_pratt (tok, std::move (outer_attrs));
-    case DOT_DOT:
-      // either "range to" or "range full" expressions
-      return parse_nud_range_exclusive_expr (tok, std::move (outer_attrs));
-    case DOT_DOT_EQ:
-      // range to inclusive expr
-      return parse_range_to_inclusive_expr (tok, std::move (outer_attrs));
-    case RETURN_KW:
-      // FIXME: is this really a null denotation expression?
-      return parse_return_expr (std::move (outer_attrs), tok->get_locus ());
-    case TRY:
-      // FIXME: is this really a null denotation expression?
-      return parse_try_expr (std::move (outer_attrs), tok->get_locus ());
-    case BREAK:
-      // FIXME: is this really a null denotation expression?
-      return parse_break_expr (std::move (outer_attrs), tok->get_locus ());
-    case CONTINUE:
-      return parse_continue_expr (std::move (outer_attrs), tok->get_locus ());
-    case LEFT_CURLY:
-      // ok - this is an expression with block for once.
-      return parse_block_expr (std::move (outer_attrs), tl::nullopt,
-			       tok->get_locus ());
-    case IF:
-      // if or if let, so more lookahead to find out
-      if (lexer.peek_token ()->get_id () == LET)
-	{
-	  // if let expr
-	  return parse_if_let_expr (std::move (outer_attrs), tok->get_locus ());
-	}
-      else
-	{
-	  // if expr
-	  return parse_if_expr (std::move (outer_attrs), tok->get_locus ());
-	}
-    case LIFETIME:
-      return parse_labelled_loop_expr (tok, std::move (outer_attrs));
-    case LOOP:
-      return parse_loop_expr (std::move (outer_attrs), tl::nullopt,
-			      tok->get_locus ());
-    case WHILE:
-      if (lexer.peek_token ()->get_id () == LET)
-	{
-	  return parse_while_let_loop_expr (std::move (outer_attrs));
-	}
-      else
-	{
-	  return parse_while_loop_expr (std::move (outer_attrs), tl::nullopt,
-					tok->get_locus ());
-	}
-    case FOR:
-      return parse_for_loop_expr (std::move (outer_attrs), tl::nullopt);
-    case MATCH_KW:
-      // also an expression with block
-      return parse_match_expr (std::move (outer_attrs), tok->get_locus ());
-    case LEFT_SQUARE:
-      // array definition expr (not indexing)
-      return parse_array_expr (std::move (outer_attrs), tok->get_locus ());
-    case UNSAFE:
-      return parse_unsafe_block_expr (std::move (outer_attrs),
-				      tok->get_locus ());
-    case BOX:
-      return parse_box_expr (std::move (outer_attrs), tok->get_locus ());
-    case UNDERSCORE:
-      add_error (
-	Error (tok->get_locus (),
-	       "use of %qs is not allowed on the right-side of an assignment",
-	       tok->get_token_description ()));
-      return nullptr;
-    case CONST:
-      return parse_const_block_expr (std::move (outer_attrs),
-				     tok->get_locus ());
-    default:
-      if (!restrictions.expr_can_be_null)
-	add_error (Error (tok->get_locus (),
-			  "found unexpected token %qs in null denotation",
-			  tok->get_token_description ()));
-      return nullptr;
-    }
-}
-
-/* Called for each token that can appear in infix (between) position. Can be
- * operators or other punctuation. Returns a function pointer to member
- * function that implements the left denotation for the token given. */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::Expr>
-Parser<ManagedTokenSource>::left_denotation (const_TokenPtr tok,
-					     std::unique_ptr<AST::Expr> left,
-					     AST::AttrVec outer_attrs,
-					     ParseRestrictions restrictions)
-{
-  // Token passed in has already been skipped, so peek gives "next" token
-  switch (tok->get_id ())
-    {
-    // FIXME: allow for outer attributes to be applied
-    case QUESTION_MARK:
-      {
-	location_t left_locus = left->get_locus ();
-	// error propagation expression - unary postfix
-	return std::unique_ptr<AST::ErrorPropagationExpr> (
-	  new AST::ErrorPropagationExpr (std::move (left),
-					 std::move (outer_attrs), left_locus));
-      }
-    case PLUS:
-      // sum expression - binary infix
-      /*return parse_binary_plus_expr (tok, std::move (left),
-				     std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (tok, std::move (left),
-					       std::move (outer_attrs),
-					       ArithmeticOrLogicalOperator::ADD,
-					       restrictions);
-    case MINUS:
-      // difference expression - binary infix
-      /*return parse_binary_minus_expr (tok, std::move (left),
-				      std::move (outer_attrs),
-	 restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::SUBTRACT, restrictions);
-    case ASTERISK:
-      // product expression - binary infix
-      /*return parse_binary_mult_expr (tok, std::move (left),
-				     std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::MULTIPLY, restrictions);
-    case DIV:
-      // quotient expression - binary infix
-      /*return parse_binary_div_expr (tok, std::move (left),
-				    std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::DIVIDE, restrictions);
-    case PERCENT:
-      // modulo expression - binary infix
-      /*return parse_binary_mod_expr (tok, std::move (left),
-				    std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::MODULUS, restrictions);
-    case AMP:
-      // logical or bitwise and expression - binary infix
-      /*return parse_bitwise_and_expr (tok, std::move (left),
-				     std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::BITWISE_AND, restrictions);
-    case PIPE:
-      // logical or bitwise or expression - binary infix
-      /*return parse_bitwise_or_expr (tok, std::move (left),
-				    std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::BITWISE_OR, restrictions);
-    case CARET:
-      // logical or bitwise xor expression - binary infix
-      /*return parse_bitwise_xor_expr (tok, std::move (left),
-				     std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::BITWISE_XOR, restrictions);
-    case LEFT_SHIFT:
-      // left shift expression - binary infix
-      /*return parse_left_shift_expr (tok, std::move (left),
-				    std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::LEFT_SHIFT, restrictions);
-    case RIGHT_SHIFT:
-      // right shift expression - binary infix
-      /*return parse_right_shift_expr (tok, std::move (left),
-				     std::move (outer_attrs), restrictions);*/
-      return parse_arithmetic_or_logical_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	ArithmeticOrLogicalOperator::RIGHT_SHIFT, restrictions);
-    case EQUAL_EQUAL:
-      // equal to expression - binary infix (no associativity)
-      /*return parse_binary_equal_expr (tok, std::move (left),
-				      std::move (outer_attrs),
-	 restrictions);*/
-      return parse_comparison_expr (tok, std::move (left),
-				    std::move (outer_attrs),
-				    ComparisonOperator::EQUAL, restrictions);
-    case NOT_EQUAL:
-      // not equal to expression - binary infix (no associativity)
-      /*return parse_binary_not_equal_expr (tok, std::move (left),
-					  std::move (outer_attrs),
-					  restrictions);*/
-      return parse_comparison_expr (tok, std::move (left),
-				    std::move (outer_attrs),
-				    ComparisonOperator::NOT_EQUAL,
-				    restrictions);
-    case RIGHT_ANGLE:
-      // greater than expression - binary infix (no associativity)
-      /*return parse_binary_greater_than_expr (tok, std::move (left),
-					     std::move (outer_attrs),
-					     restrictions);*/
-      return parse_comparison_expr (tok, std::move (left),
-				    std::move (outer_attrs),
-				    ComparisonOperator::GREATER_THAN,
-				    restrictions);
-    case LEFT_ANGLE:
-      // less than expression - binary infix (no associativity)
-      /*return parse_binary_less_than_expr (tok, std::move (left),
-					  std::move (outer_attrs),
-					  restrictions);*/
-      return parse_comparison_expr (tok, std::move (left),
-				    std::move (outer_attrs),
-				    ComparisonOperator::LESS_THAN,
-				    restrictions);
-    case GREATER_OR_EQUAL:
-      // greater than or equal to expression - binary infix (no associativity)
-      /*return parse_binary_greater_equal_expr (tok, std::move (left),
-					      std::move (outer_attrs),
-					      restrictions);*/
-      return parse_comparison_expr (tok, std::move (left),
-				    std::move (outer_attrs),
-				    ComparisonOperator::GREATER_OR_EQUAL,
-				    restrictions);
-    case LESS_OR_EQUAL:
-      // less than or equal to expression - binary infix (no associativity)
-      /*return parse_binary_less_equal_expr (tok, std::move (left),
-					   std::move (outer_attrs),
-					   restrictions);*/
-      return parse_comparison_expr (tok, std::move (left),
-				    std::move (outer_attrs),
-				    ComparisonOperator::LESS_OR_EQUAL,
-				    restrictions);
-    case OR:
-      // lazy logical or expression - binary infix
-      return parse_lazy_or_expr (tok, std::move (left), std::move (outer_attrs),
-				 restrictions);
-    case LOGICAL_AND:
-      // lazy logical and expression - binary infix
-      return parse_lazy_and_expr (tok, std::move (left),
-				  std::move (outer_attrs), restrictions);
-    case AS:
-      /* type cast expression - kind of binary infix (RHS is actually a
-       * TypeNoBounds) */
-      return parse_type_cast_expr (tok, std::move (left),
-				   std::move (outer_attrs), restrictions);
-    case EQUAL:
-      // assignment expression - binary infix (note right-to-left
-      // associativity)
-      return parse_assig_expr (tok, std::move (left), std::move (outer_attrs),
-			       restrictions);
-    case PLUS_EQ:
-      /* plus-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_plus_assig_expr (tok, std::move (left),
-				    std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (tok, std::move (left),
-					     std::move (outer_attrs),
-					     CompoundAssignmentOperator::ADD,
-					     restrictions);
-    case MINUS_EQ:
-      /* minus-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_minus_assig_expr (tok, std::move (left),
-				     std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::SUBTRACT, restrictions);
-    case ASTERISK_EQ:
-      /* multiply-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_mult_assig_expr (tok, std::move (left),
-				    std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::MULTIPLY, restrictions);
-    case DIV_EQ:
-      /* division-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_div_assig_expr (tok, std::move (left),
-				   std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (tok, std::move (left),
-					     std::move (outer_attrs),
-					     CompoundAssignmentOperator::DIVIDE,
-					     restrictions);
-    case PERCENT_EQ:
-      /* modulo-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_mod_assig_expr (tok, std::move (left),
-				   std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::MODULUS, restrictions);
-    case AMP_EQ:
-      /* bitwise and-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_and_assig_expr (tok, std::move (left),
-				   std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::BITWISE_AND, restrictions);
-    case PIPE_EQ:
-      /* bitwise or-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_or_assig_expr (tok, std::move (left),
-				  std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::BITWISE_OR, restrictions);
-    case CARET_EQ:
-      /* bitwise xor-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_xor_assig_expr (tok, std::move (left),
-				   std::move (outer_attrs), restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::BITWISE_XOR, restrictions);
-    case LEFT_SHIFT_EQ:
-      /* left shift-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_left_shift_assig_expr (tok, std::move (left),
-					  std::move (outer_attrs),
-					  restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::LEFT_SHIFT, restrictions);
-    case RIGHT_SHIFT_EQ:
-      /* right shift-assignment expression - binary infix (note right-to-left
-       * associativity) */
-      /*return parse_right_shift_assig_expr (tok, std::move (left),
-					   std::move (outer_attrs),
-					   restrictions);*/
-      return parse_compound_assignment_expr (
-	tok, std::move (left), std::move (outer_attrs),
-	CompoundAssignmentOperator::RIGHT_SHIFT, restrictions);
-    case DOT_DOT:
-      /* range exclusive expression - binary infix (no associativity)
-       * either "range" or "range from" */
-      return parse_led_range_exclusive_expr (tok, std::move (left),
-					     std::move (outer_attrs),
-					     restrictions);
-    case DOT_DOT_EQ:
-      /* range inclusive expression - binary infix (no associativity)
-       * unambiguously RangeInclusiveExpr */
-      return parse_range_inclusive_expr (tok, std::move (left),
-					 std::move (outer_attrs), restrictions);
-    case SCOPE_RESOLUTION:
-      // path expression - binary infix? FIXME should this even be parsed
-      // here?
-      add_error (
-	Error (tok->get_locus (),
-	       "found scope resolution operator in left denotation "
-	       "function - this should probably be handled elsewhere"));
-
-      return nullptr;
-    case DOT:
-      {
-	/* field expression or method call - relies on parentheses after next
-	 * identifier or await if token after is "await" (unary postfix) or
-	 * tuple index if token after is a decimal int literal */
-
-	const_TokenPtr next_tok = lexer.peek_token ();
-	if (next_tok->get_id () == IDENTIFIER
-	    && next_tok->get_str () == Values::Keywords::AWAIT)
-	  {
-	    // await expression
-	    return parse_await_expr (tok, std::move (left),
-				     std::move (outer_attrs));
-	  }
-	else if (next_tok->get_id () == INT_LITERAL)
-	  {
-	    // tuple index expression - TODO check for decimal int literal
-	    return parse_tuple_index_expr (tok, std::move (left),
-					   std::move (outer_attrs),
-					   restrictions);
-	  }
-	else if (next_tok->get_id () == FLOAT_LITERAL)
-	  {
-	    // Lexer has misidentified a tuple index as a float literal
-	    // eg: `(x, (y, z)).1.0` -> 1.0 has been identified as a float
-	    // literal. This means we should split it into three new separate
-	    // tokens, the first tuple index, the dot and the second tuple
-	    // index.
-	    auto current_loc = next_tok->get_locus ();
-	    auto str = next_tok->get_str ();
-	    auto dot_pos = str.find (".");
-	    auto prefix = str.substr (0, dot_pos);
-	    auto suffix = str.substr (dot_pos + 1);
-	    if (dot_pos == str.size () - 1)
-	      lexer.split_current_token (
-		{Token::make_int (current_loc, std::move (prefix),
-				  CORETYPE_PURE_DECIMAL),
-		 Token::make (DOT, current_loc + 1)});
-	    else
-	      lexer.split_current_token (
-		{Token::make_int (current_loc, std::move (prefix),
-				  CORETYPE_PURE_DECIMAL),
-		 Token::make (DOT, current_loc + 1),
-		 Token::make_int (current_loc + 2, std::move (suffix),
-				  CORETYPE_PURE_DECIMAL)});
-	    return parse_tuple_index_expr (tok, std::move (left),
-					   std::move (outer_attrs),
-					   restrictions);
-	  }
-	else if (next_tok->get_id () == IDENTIFIER
-		 && lexer.peek_token (1)->get_id () != LEFT_PAREN
-		 && lexer.peek_token (1)->get_id () != SCOPE_RESOLUTION)
-	  {
-	    /* field expression (or should be) - FIXME: scope resolution right
-	     * after identifier should always be method, I'm pretty sure */
-	    return parse_field_access_expr (tok, std::move (left),
-					    std::move (outer_attrs),
-					    restrictions);
-	  }
-	else
-	  {
-	    // method call (probably)
-	    return parse_method_call_expr (tok, std::move (left),
-					   std::move (outer_attrs),
-					   restrictions);
-	  }
-      }
-    case LEFT_PAREN:
-      // function call - method call is based on dot notation first
-      return parse_function_call_expr (tok, std::move (left),
-				       std::move (outer_attrs), restrictions);
-    case LEFT_SQUARE:
-      // array or slice index expression (pseudo binary infix)
-      return parse_index_expr (tok, std::move (left), std::move (outer_attrs),
-			       restrictions);
-    default:
-      add_error (Error (tok->get_locus (),
-			"found unexpected token %qs in left denotation",
-			tok->get_token_description ()));
-
-      return nullptr;
-    }
-}
-
-/* Returns the left binding power for the given ArithmeticOrLogicalExpr type.
- * TODO make constexpr? Would that even do anything useful? */
-inline binding_powers
-get_lbp_for_arithmetic_or_logical_expr (
-  AST::ArithmeticOrLogicalExpr::ExprType expr_type)
-{
-  switch (expr_type)
-    {
-    case ArithmeticOrLogicalOperator::ADD:
-      return LBP_PLUS;
-    case ArithmeticOrLogicalOperator::SUBTRACT:
-      return LBP_MINUS;
-    case ArithmeticOrLogicalOperator::MULTIPLY:
-      return LBP_MUL;
-    case ArithmeticOrLogicalOperator::DIVIDE:
-      return LBP_DIV;
-    case ArithmeticOrLogicalOperator::MODULUS:
-      return LBP_MOD;
-    case ArithmeticOrLogicalOperator::BITWISE_AND:
-      return LBP_AMP;
-    case ArithmeticOrLogicalOperator::BITWISE_OR:
-      return LBP_PIPE;
-    case ArithmeticOrLogicalOperator::BITWISE_XOR:
-      return LBP_CARET;
-    case ArithmeticOrLogicalOperator::LEFT_SHIFT:
-      return LBP_L_SHIFT;
-    case ArithmeticOrLogicalOperator::RIGHT_SHIFT:
-      return LBP_R_SHIFT;
-    default:
-      // WTF? should not happen, this is an error
-      rust_unreachable ();
-
-      return LBP_PLUS;
-    }
-}
-
-// Parses an arithmetic or logical expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_arithmetic_or_logical_expr (
-  const_TokenPtr, std::unique_ptr<AST::Expr> left, AST::AttrVec,
-  AST::ArithmeticOrLogicalExpr::ExprType expr_type,
-  ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (get_lbp_for_arithmetic_or_logical_expr (expr_type),
-		  AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      expr_type, locus));
-}
-
-// Parses a binary addition expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_binary_plus_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_PLUS, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::ADD, locus));
-}
-
-// Parses a binary subtraction expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_binary_minus_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_MINUS, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::SUBTRACT,
-				      locus));
-}
-
-// Parses a binary multiplication expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_binary_mult_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_MUL, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::MULTIPLY,
-				      locus));
-}
-
-// Parses a binary division expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_binary_div_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_DIV, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::DIVIDE,
-				      locus));
-}
-
-// Parses a binary modulo expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_binary_mod_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_MOD, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::MODULUS,
-				      locus));
-}
-
-/* Parses a binary bitwise (or eager logical) and expression (with Pratt
- * parsing). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_bitwise_and_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_AMP, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::BITWISE_AND,
-				      locus));
-}
-
-/* Parses a binary bitwise (or eager logical) or expression (with Pratt
- * parsing). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_bitwise_or_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_PIPE, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::BITWISE_OR,
-				      locus));
-}
-
-/* Parses a binary bitwise (or eager logical) xor expression (with Pratt
- * parsing). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_bitwise_xor_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_CARET, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::BITWISE_XOR,
-				      locus));
-}
-
-// Parses a binary left shift expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_left_shift_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_L_SHIFT, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::LEFT_SHIFT,
-				      locus));
-}
-
-// Parses a binary right shift expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArithmeticOrLogicalExpr>
-Parser<ManagedTokenSource>::parse_right_shift_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_R_SHIFT, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ArithmeticOrLogicalExpr> (
-    new AST::ArithmeticOrLogicalExpr (std::move (left), std::move (right),
-				      ArithmeticOrLogicalOperator::RIGHT_SHIFT,
-				      locus));
-}
-
-/* Returns the left binding power for the given ComparisonExpr type.
- * TODO make constexpr? Would that even do anything useful? */
-inline binding_powers
-get_lbp_for_comparison_expr (AST::ComparisonExpr::ExprType expr_type)
-{
-  switch (expr_type)
-    {
-    case ComparisonOperator::EQUAL:
-      return LBP_EQUAL;
-    case ComparisonOperator::NOT_EQUAL:
-      return LBP_NOT_EQUAL;
-    case ComparisonOperator::GREATER_THAN:
-      return LBP_GREATER_THAN;
-    case ComparisonOperator::LESS_THAN:
-      return LBP_SMALLER_THAN;
-    case ComparisonOperator::GREATER_OR_EQUAL:
-      return LBP_GREATER_EQUAL;
-    case ComparisonOperator::LESS_OR_EQUAL:
-      return LBP_SMALLER_EQUAL;
-    default:
-      // WTF? should not happen, this is an error
-      rust_unreachable ();
-
-      return LBP_EQUAL;
-    }
-}
-
-/* Parses a ComparisonExpr of given type and LBP. TODO find a way to only
- * specify one and have the other looked up - e.g. specify ExprType and
- * binding power is looked up? */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ComparisonExpr>
-Parser<ManagedTokenSource>::parse_comparison_expr (
-  const_TokenPtr, std::unique_ptr<AST::Expr> left, AST::AttrVec,
-  AST::ComparisonExpr::ExprType expr_type, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (get_lbp_for_comparison_expr (expr_type), AST::AttrVec (),
-		  restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ComparisonExpr> (
-    new AST::ComparisonExpr (std::move (left), std::move (right), expr_type,
-			     locus));
-}
-
-// Parses a binary equal to expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ComparisonExpr>
-Parser<ManagedTokenSource>::parse_binary_equal_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_EQUAL, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ComparisonExpr> (
-    new AST::ComparisonExpr (std::move (left), std::move (right),
-			     ComparisonOperator::EQUAL, locus));
-}
-
-// Parses a binary not equal to expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ComparisonExpr>
-Parser<ManagedTokenSource>::parse_binary_not_equal_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_NOT_EQUAL, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ComparisonExpr> (
-    new AST::ComparisonExpr (std::move (left), std::move (right),
-			     ComparisonOperator::NOT_EQUAL, locus));
-}
-
-// Parses a binary greater than expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ComparisonExpr>
-Parser<ManagedTokenSource>::parse_binary_greater_than_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_GREATER_THAN, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ComparisonExpr> (
-    new AST::ComparisonExpr (std::move (left), std::move (right),
-			     ComparisonOperator::GREATER_THAN, locus));
-}
-
-// Parses a binary less than expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ComparisonExpr>
-Parser<ManagedTokenSource>::parse_binary_less_than_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_SMALLER_THAN, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ComparisonExpr> (
-    new AST::ComparisonExpr (std::move (left), std::move (right),
-			     ComparisonOperator::LESS_THAN, locus));
-}
-
-// Parses a binary greater than or equal to expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ComparisonExpr>
-Parser<ManagedTokenSource>::parse_binary_greater_equal_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_GREATER_EQUAL, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ComparisonExpr> (
-    new AST::ComparisonExpr (std::move (left), std::move (right),
-			     ComparisonOperator::GREATER_OR_EQUAL, locus));
-}
-
-// Parses a binary less than or equal to expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ComparisonExpr>
-Parser<ManagedTokenSource>::parse_binary_less_equal_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_SMALLER_EQUAL, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::ComparisonExpr> (
-    new AST::ComparisonExpr (std::move (left), std::move (right),
-			     ComparisonOperator::LESS_OR_EQUAL, locus));
-}
-
-// Parses a binary lazy boolean or expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::LazyBooleanExpr>
-Parser<ManagedTokenSource>::parse_lazy_or_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_LOGICAL_OR, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::LazyBooleanExpr> (
-    new AST::LazyBooleanExpr (std::move (left), std::move (right),
-			      LazyBooleanOperator::LOGICAL_OR, locus));
-}
-
-// Parses a binary lazy boolean and expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::LazyBooleanExpr>
-Parser<ManagedTokenSource>::parse_lazy_and_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_LOGICAL_AND, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::LazyBooleanExpr> (
-    new AST::LazyBooleanExpr (std::move (left), std::move (right),
-			      LazyBooleanOperator::LOGICAL_AND, locus));
-}
-
-// Parses a pseudo-binary infix type cast expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::TypeCastExpr>
-Parser<ManagedTokenSource>::parse_type_cast_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> expr_to_cast,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED,
-  ParseRestrictions restrictions ATTRIBUTE_UNUSED)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::TypeNoBounds> type = parse_type_no_bounds ();
-  if (type == nullptr)
-    return nullptr;
-  // FIXME: how do I get precedence put in here?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = expr_to_cast->get_locus ();
-
-  return std::unique_ptr<AST::TypeCastExpr> (
-    new AST::TypeCastExpr (std::move (expr_to_cast), std::move (type), locus));
-}
-
-// Parses a binary assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::AssignmentExpr>
-Parser<ManagedTokenSource>::parse_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::AssignmentExpr> (
-    new AST::AssignmentExpr (std::move (left), std::move (right),
-			     std::move (outer_attrs), locus));
-}
-
-/* Returns the left binding power for the given CompoundAssignmentExpr type.
- * TODO make constexpr? Would that even do anything useful? */
-inline binding_powers
-get_lbp_for_compound_assignment_expr (
-  AST::CompoundAssignmentExpr::ExprType expr_type)
-{
-  switch (expr_type)
-    {
-    case CompoundAssignmentOperator::ADD:
-      return LBP_PLUS;
-    case CompoundAssignmentOperator::SUBTRACT:
-      return LBP_MINUS;
-    case CompoundAssignmentOperator::MULTIPLY:
-      return LBP_MUL;
-    case CompoundAssignmentOperator::DIVIDE:
-      return LBP_DIV;
-    case CompoundAssignmentOperator::MODULUS:
-      return LBP_MOD;
-    case CompoundAssignmentOperator::BITWISE_AND:
-      return LBP_AMP;
-    case CompoundAssignmentOperator::BITWISE_OR:
-      return LBP_PIPE;
-    case CompoundAssignmentOperator::BITWISE_XOR:
-      return LBP_CARET;
-    case CompoundAssignmentOperator::LEFT_SHIFT:
-      return LBP_L_SHIFT;
-    case CompoundAssignmentOperator::RIGHT_SHIFT:
-      return LBP_R_SHIFT;
-    default:
-      // WTF? should not happen, this is an error
-      rust_unreachable ();
-
-      return LBP_PLUS;
-    }
-}
-
-// Parses a compound assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_compound_assignment_expr (
-  const_TokenPtr, std::unique_ptr<AST::Expr> left, AST::AttrVec,
-  AST::CompoundAssignmentExpr::ExprType expr_type,
-  ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (get_lbp_for_compound_assignment_expr (expr_type) - 1,
-		  AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     expr_type, locus));
-}
-
-// Parses a binary add-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_plus_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_PLUS_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::ADD, locus));
-}
-
-// Parses a binary minus-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_minus_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_MINUS_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::SUBTRACT,
-				     locus));
-}
-
-// Parses a binary multiplication-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_mult_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_MULT_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::MULTIPLY,
-				     locus));
-}
-
-// Parses a binary division-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_div_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_DIV_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::DIVIDE,
-				     locus));
-}
-
-// Parses a binary modulo-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_mod_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_MOD_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::MODULUS,
-				     locus));
-}
-
-// Parses a binary and-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_and_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_AMP_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::BITWISE_AND,
-				     locus));
-}
-
-// Parses a binary or-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_or_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_PIPE_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::BITWISE_OR,
-				     locus));
-}
-
-// Parses a binary xor-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_xor_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_CARET_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::BITWISE_XOR,
-				     locus));
-}
-
-// Parses a binary left shift-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_left_shift_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_L_SHIFT_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::LEFT_SHIFT,
-				     locus));
-}
-
-// Parses a binary right shift-assignment expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CompoundAssignmentExpr>
-Parser<ManagedTokenSource>::parse_right_shift_assig_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_R_SHIFT_ASSIG - 1, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: ensure right-associativity for this - 'LBP - 1' may do this?
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::CompoundAssignmentExpr> (
-    new AST::CompoundAssignmentExpr (std::move (left), std::move (right),
-				     CompoundAssignmentOperator::RIGHT_SHIFT,
-				     locus));
-}
-
-// Parses a postfix unary await expression (with Pratt parsing).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::AwaitExpr>
-Parser<ManagedTokenSource>::parse_await_expr (
-  const_TokenPtr tok, std::unique_ptr<AST::Expr> expr_to_await,
-  AST::AttrVec outer_attrs)
-{
-  /* skip "await" identifier (as "." has already been consumed in
-   * parse_expression) this assumes that the identifier was already identified
-   * as await */
-  if (!skip_token (IDENTIFIER))
-    {
-      Error error (tok->get_locus (), "failed to skip %<await%> in await expr "
-				      "- this is probably a deep issue");
-      add_error (std::move (error));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // TODO: check inside async block in semantic analysis
-  location_t locus = expr_to_await->get_locus ();
-
-  return std::unique_ptr<AST::AwaitExpr> (
-    new AST::AwaitExpr (std::move (expr_to_await), std::move (outer_attrs),
-			locus));
-}
-
-/* Parses an exclusive range ('..') in left denotation position (i.e.
- * RangeFromExpr or RangeFromToExpr). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::RangeExpr>
-Parser<ManagedTokenSource>::parse_led_range_exclusive_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // FIXME: this probably parses expressions accidently or whatever
-  // try parsing RHS (as tok has already been consumed in parse_expression)
-  // Can be nullptr, in which case it is a RangeFromExpr, otherwise a
-  // RangeFromToExpr.
-  restrictions.expr_can_be_null = true;
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_DOT_DOT, AST::AttrVec (), restrictions);
-
-  location_t locus = left->get_locus ();
-
-  if (right == nullptr)
-    {
-      // range from expr
-      return std::unique_ptr<AST::RangeFromExpr> (
-	new AST::RangeFromExpr (std::move (left), locus));
-    }
-  else
-    {
-      return std::unique_ptr<AST::RangeFromToExpr> (
-	new AST::RangeFromToExpr (std::move (left), std::move (right), locus));
-    }
-  // FIXME: make non-associative
-}
-
-/* Parses an exclusive range ('..') in null denotation position (i.e.
- * RangeToExpr or RangeFullExpr). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::RangeExpr>
-Parser<ManagedTokenSource>::parse_nud_range_exclusive_expr (
-  const_TokenPtr tok, AST::AttrVec outer_attrs ATTRIBUTE_UNUSED)
-{
-  auto restrictions = ParseRestrictions ();
-  restrictions.expr_can_be_null = true;
-
-  // FIXME: this probably parses expressions accidently or whatever
-  // try parsing RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_DOT_DOT, AST::AttrVec (), restrictions);
-
-  location_t locus = tok->get_locus ();
-
-  if (right == nullptr)
-    {
-      // range from expr
-      return std::unique_ptr<AST::RangeFullExpr> (
-	new AST::RangeFullExpr (locus));
-    }
-  else
-    {
-      return std::unique_ptr<AST::RangeToExpr> (
-	new AST::RangeToExpr (std::move (right), locus));
-    }
-  // FIXME: make non-associative
-}
-
-// Parses a full binary range inclusive expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::RangeFromToInclExpr>
-Parser<ManagedTokenSource>::parse_range_inclusive_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> left,
-  AST::AttrVec outer_attrs ATTRIBUTE_UNUSED, ParseRestrictions restrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right
-    = parse_expr (LBP_DOT_DOT_EQ, AST::AttrVec (), restrictions);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: make non-associative
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = left->get_locus ();
-
-  return std::unique_ptr<AST::RangeFromToInclExpr> (
-    new AST::RangeFromToInclExpr (std::move (left), std::move (right), locus));
-}
-
-// Parses an inclusive range-to prefix unary expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::RangeToInclExpr>
-Parser<ManagedTokenSource>::parse_range_to_inclusive_expr (
-  const_TokenPtr tok, AST::AttrVec outer_attrs ATTRIBUTE_UNUSED)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  std::unique_ptr<AST::Expr> right = parse_expr (LBP_DOT_DOT_EQ);
-  if (right == nullptr)
-    return nullptr;
-  // FIXME: make non-associative
-
-  // TODO: check types. actually, do so during semantic analysis
-
-  return std::unique_ptr<AST::RangeToInclExpr> (
-    new AST::RangeToInclExpr (std::move (right), tok->get_locus ()));
-}
-
-// Parses a pseudo-binary infix tuple index expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::TupleIndexExpr>
-Parser<ManagedTokenSource>::parse_tuple_index_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> tuple_expr,
-  AST::AttrVec outer_attrs, ParseRestrictions restrictions ATTRIBUTE_UNUSED)
-{
-  // parse int literal (as token already skipped)
-  const_TokenPtr index_tok = expect_token (INT_LITERAL);
-  if (index_tok == nullptr)
-    {
-      return nullptr;
-    }
-  std::string index = index_tok->get_str ();
-
-  // convert to integer
-  if (!index_tok->is_pure_decimal ())
-    {
-      Error error (index_tok->get_locus (),
-		   "tuple index should be a pure decimal literal");
-      add_error (std::move (error));
-    }
-  int index_int = atoi (index.c_str ());
-
-  location_t locus = tuple_expr->get_locus ();
-
-  return std::unique_ptr<AST::TupleIndexExpr> (
-    new AST::TupleIndexExpr (std::move (tuple_expr), index_int,
-			     std::move (outer_attrs), locus));
-}
-
-// Parses a pseudo-binary infix array (or slice) index expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ArrayIndexExpr>
-Parser<ManagedTokenSource>::parse_index_expr (
-  const_TokenPtr, std::unique_ptr<AST::Expr> array_expr,
-  AST::AttrVec outer_attrs, ParseRestrictions)
-{
-  // parse RHS (as tok has already been consumed in parse_expression)
-  /*std::unique_ptr<AST::Expr> index_expr
-    = parse_expr (LBP_ARRAY_REF, AST::AttrVec (),
-    restrictions);*/
-  // TODO: conceptually, should treat [] as brackets, so just parse all expr
-  std::unique_ptr<AST::Expr> index_expr = parse_expr ();
-  if (index_expr == nullptr)
-    return nullptr;
-
-  // skip ']' at end of array
-  if (!skip_token (RIGHT_SQUARE))
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = array_expr->get_locus ();
-
-  return std::unique_ptr<AST::ArrayIndexExpr> (
-    new AST::ArrayIndexExpr (std::move (array_expr), std::move (index_expr),
-			     std::move (outer_attrs), locus));
-}
-
-// Parses a pseudo-binary infix struct field access expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::FieldAccessExpr>
-Parser<ManagedTokenSource>::parse_field_access_expr (
-  const_TokenPtr tok ATTRIBUTE_UNUSED, std::unique_ptr<AST::Expr> struct_expr,
-  AST::AttrVec outer_attrs, ParseRestrictions restrictions ATTRIBUTE_UNUSED)
-{
-  /* get field name identifier (assume that this is a field access expr and
-   * not await, for instance) */
-  const_TokenPtr ident_tok = expect_token (IDENTIFIER);
-  if (ident_tok == nullptr)
-    return nullptr;
-
-  Identifier ident{ident_tok};
-
-  location_t locus = struct_expr->get_locus ();
-
-  // TODO: check types. actually, do so during semantic analysis
-  return std::unique_ptr<AST::FieldAccessExpr> (
-    new AST::FieldAccessExpr (std::move (struct_expr), std::move (ident),
-			      std::move (outer_attrs), locus));
-}
-
-// Parses a pseudo-binary infix method call expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MethodCallExpr>
-Parser<ManagedTokenSource>::parse_method_call_expr (
-  const_TokenPtr tok, std::unique_ptr<AST::Expr> receiver_expr,
-  AST::AttrVec outer_attrs, ParseRestrictions)
-{
-  // parse path expr segment
-  AST::PathExprSegment segment = parse_path_expr_segment ();
-  if (segment.is_error ())
-    {
-      Error error (tok->get_locus (),
-		   "failed to parse path expr segment of method call expr");
-      add_error (std::move (error));
-
-      return nullptr;
-    }
-
-  // skip left parentheses
-  if (!skip_token (LEFT_PAREN))
-    {
-      return nullptr;
-    }
-
-  // parse method params (if they exist)
-  std::vector<std::unique_ptr<AST::Expr>> params;
-
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () != RIGHT_PAREN)
-    {
-      std::unique_ptr<AST::Expr> param = parse_expr ();
-      if (param == nullptr)
-	{
-	  Error error (t->get_locus (),
-		       "failed to parse method param in method call");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-      params.push_back (std::move (param));
-
-      if (lexer.peek_token ()->get_id () != COMMA)
-	break;
-
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-    }
-
-  // skip right paren
-  if (!skip_token (RIGHT_PAREN))
-    {
-      return nullptr;
-    }
-
-  // TODO: check types. actually do so in semantic analysis pass.
-  location_t locus = receiver_expr->get_locus ();
-
-  return std::unique_ptr<AST::MethodCallExpr> (
-    new AST::MethodCallExpr (std::move (receiver_expr), std::move (segment),
-			     std::move (params), std::move (outer_attrs),
-			     locus));
-}
-
-// Parses a pseudo-binary infix function call expression.
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CallExpr>
-Parser<ManagedTokenSource>::parse_function_call_expr (
-  const_TokenPtr, std::unique_ptr<AST::Expr> function_expr,
-  AST::AttrVec outer_attrs, ParseRestrictions)
-{
-  // parse function params (if they exist)
-  std::vector<std::unique_ptr<AST::Expr>> params;
-
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () != RIGHT_PAREN)
-    {
-      std::unique_ptr<AST::Expr> param = parse_expr ();
-      if (param == nullptr)
-	{
-	  Error error (t->get_locus (),
-		       "failed to parse function param in function call");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-      params.push_back (std::move (param));
-
-      if (lexer.peek_token ()->get_id () != COMMA)
-	break;
-
-      lexer.skip_token ();
-      t = lexer.peek_token ();
-    }
-
-  // skip ')' at end of param list
-  if (!skip_token (RIGHT_PAREN))
-    {
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // TODO: check types. actually, do so during semantic analysis
-  location_t locus = function_expr->get_locus ();
-
-  return std::unique_ptr<AST::CallExpr> (
-    new AST::CallExpr (std::move (function_expr), std::move (params),
-		       std::move (outer_attrs), locus));
-}
-
-/* Parses a macro invocation with a path in expression already parsed (but not
- * '!' token). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::MacroInvocation>
-Parser<ManagedTokenSource>::parse_macro_invocation_partial (
-  AST::PathInExpression path, AST::AttrVec outer_attrs,
-  ParseRestrictions restrictions)
-{
-  // macro invocation
-  if (!skip_token (EXCLAM))
-    {
-      return nullptr;
-    }
-
-  // convert PathInExpression to SimplePath - if this isn't possible, error
-  AST::SimplePath converted_path = path.as_simple_path ();
-  if (converted_path.is_empty ())
-    {
-      Error error (lexer.peek_token ()->get_locus (),
-		   "failed to parse simple path in macro invocation");
-      add_error (std::move (error));
-
-      return nullptr;
-    }
-
-  auto tok_tree = parse_delim_token_tree ();
-  if (!tok_tree)
-    return nullptr;
-
-  rust_debug ("successfully parsed macro invocation (via partial)");
-
-  location_t macro_locus = converted_path.get_locus ();
-
-  return AST::MacroInvocation::Regular (
-    AST::MacroInvocData (std::move (converted_path),
-			 std::move (tok_tree.value ())),
-    std::move (outer_attrs), macro_locus);
-}
-
-/* Parses a struct expr struct with a path in expression already parsed (but
- * not
- * '{' token). */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::StructExprStruct>
-Parser<ManagedTokenSource>::parse_struct_expr_struct_partial (
-  AST::PathInExpression path, AST::AttrVec outer_attrs)
-{
-  // assume struct expr struct (as struct-enum disambiguation requires name
-  // lookup) again, make statement if final ';'
-  if (!skip_token (LEFT_CURLY))
-    {
-      return nullptr;
-    }
-
-  // parse inner attributes
-  AST::AttrVec inner_attrs = parse_inner_attributes ();
-
-  // branch based on next token
-  const_TokenPtr t = lexer.peek_token ();
-  location_t path_locus = path.get_locus ();
-  switch (t->get_id ())
-    {
-    case RIGHT_CURLY:
-      // struct with no body
-      lexer.skip_token ();
-
-      return std::unique_ptr<AST::StructExprStruct> (
-	new AST::StructExprStruct (std::move (path), std::move (inner_attrs),
-				   std::move (outer_attrs), path_locus));
-    case DOT_DOT:
-      /* technically this would give a struct base-only struct, but this
-       * algorithm should work too. As such, AST type not happening. */
-    case IDENTIFIER:
-    case HASH:
-    case INT_LITERAL:
-      {
-	// struct with struct expr fields
-
-	// parse struct expr fields
-	std::vector<std::unique_ptr<AST::StructExprField>> fields;
-
-	while (t->get_id () != RIGHT_CURLY && t->get_id () != DOT_DOT)
-	  {
-	    std::unique_ptr<AST::StructExprField> field
-	      = parse_struct_expr_field ();
-	    if (field == nullptr)
-	      {
-		Error error (t->get_locus (),
-			     "failed to parse struct (or enum) expr field");
-		add_error (std::move (error));
-
-		return nullptr;
-	      }
-
-	    // DEBUG:
-	    rust_debug ("struct/enum expr field validated to not be null");
-
-	    fields.push_back (std::move (field));
-
-	    // DEBUG:
-	    rust_debug ("struct/enum expr field pushed back");
-
-	    if (lexer.peek_token ()->get_id () != COMMA)
-	      {
-		// DEBUG:
-		rust_debug ("lack of comma detected in struct/enum expr "
-			    "fields - break");
-		break;
-	      }
-	    lexer.skip_token ();
-
-	    // DEBUG:
-	    rust_debug ("struct/enum expr fields comma skipped ");
-
-	    t = lexer.peek_token ();
-	  }
-
-	// DEBUG:
-	rust_debug ("struct/enum expr about to parse struct base ");
-
-	// parse struct base if it exists
-	AST::StructBase struct_base = AST::StructBase::error ();
-	if (lexer.peek_token ()->get_id () == DOT_DOT)
-	  {
-	    location_t dot_dot_location = lexer.peek_token ()->get_locus ();
-	    lexer.skip_token ();
-
-	    // parse required struct base expr
-	    std::unique_ptr<AST::Expr> base_expr = parse_expr ();
-	    if (base_expr == nullptr)
-	      {
-		Error error (lexer.peek_token ()->get_locus (),
-			     "failed to parse struct base expression in struct "
-			     "expression");
-		add_error (std::move (error));
-
-		return nullptr;
-	      }
-
-	    // DEBUG:
-	    rust_debug ("struct/enum expr - parsed and validated base expr");
-
-	    struct_base
-	      = AST::StructBase (std::move (base_expr), dot_dot_location);
-
-	    // DEBUG:
-	    rust_debug ("assigned struct base to new struct base ");
-	  }
-
-	if (!skip_token (RIGHT_CURLY))
-	  {
-	    return nullptr;
-	  }
-
-	// DEBUG:
-	rust_debug (
-	  "struct/enum expr skipped right curly - done and ready to return");
-
-	return std::unique_ptr<AST::StructExprStructFields> (
-	  new AST::StructExprStructFields (std::move (path), std::move (fields),
-					   path_locus, std::move (struct_base),
-					   std::move (inner_attrs),
-					   std::move (outer_attrs)));
-      }
-    default:
-      add_error (
-	Error (t->get_locus (),
-	       "unrecognised token %qs in struct (or enum) expression - "
-	       "expected %<}%>, identifier, integer literal, or %<..%>",
-	       t->get_token_description ()));
-
-      return nullptr;
-    }
-}
-
-/* Parses a struct expr tuple with a path in expression already parsed (but
- * not
- * '(' token).
- * FIXME: this currently outputs a call expr, as they cannot be disambiguated.
- * A better solution would be to just get this to call that function directly.
- * */
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::CallExpr>
-Parser<ManagedTokenSource>::parse_struct_expr_tuple_partial (
-  AST::PathInExpression path, AST::AttrVec outer_attrs)
-{
-  if (!skip_token (LEFT_PAREN))
-    {
-      return nullptr;
-    }
-
-  AST::AttrVec inner_attrs = parse_inner_attributes ();
-
-  std::vector<std::unique_ptr<AST::Expr>> exprs;
-
-  const_TokenPtr t = lexer.peek_token ();
-  while (t->get_id () != RIGHT_PAREN)
-    {
-      // parse expression (required)
-      std::unique_ptr<AST::Expr> expr = parse_expr ();
-      if (expr == nullptr)
-	{
-	  Error error (t->get_locus (), "failed to parse expression in "
-					"struct (or enum) expression tuple");
-	  add_error (std::move (error));
-
-	  return nullptr;
-	}
-      exprs.push_back (std::move (expr));
-
-      if (lexer.peek_token ()->get_id () != COMMA)
-	break;
-
-      lexer.skip_token ();
-
-      t = lexer.peek_token ();
-    }
-
-  if (!skip_token (RIGHT_PAREN))
-    {
-      return nullptr;
-    }
-
-  location_t path_locus = path.get_locus ();
-
-  auto pathExpr = std::unique_ptr<AST::PathInExpression> (
-    new AST::PathInExpression (std::move (path)));
-
-  return std::unique_ptr<AST::CallExpr> (
-    new AST::CallExpr (std::move (pathExpr), std::move (exprs),
-		       std::move (outer_attrs), path_locus));
-}
-
-// Parses a closure expression with pratt parsing (from null denotation).
-template <typename ManagedTokenSource>
-std::unique_ptr<AST::ClosureExpr>
-Parser<ManagedTokenSource>::parse_closure_expr_pratt (const_TokenPtr tok,
-						      AST::AttrVec outer_attrs)
-{
-  // TODO: does this need pratt parsing (for precedence)? probably not, but
-  // idk
-  location_t locus = tok->get_locus ();
-  bool has_move = false;
-  if (tok->get_id () == MOVE)
-    {
-      has_move = true;
-      tok = lexer.peek_token ();
-      lexer.skip_token ();
-      // skip token and reassign
-    }
-
-  // handle parameter list
-  std::vector<AST::ClosureParam> params;
-
-  switch (tok->get_id ())
-    {
-    case OR:
-      // no parameters, don't skip token
-      break;
-    case PIPE:
-      {
-	// actually may have parameters
-	// don't skip token
-	const_TokenPtr t = lexer.peek_token ();
-	while (t->get_id () != PIPE)
-	  {
-	    AST::ClosureParam param = parse_closure_param ();
-	    if (param.is_error ())
-	      {
-		// TODO is this really an error?
-		Error error (t->get_locus (), "could not parse closure param");
-		add_error (std::move (error));
-
-		return nullptr;
-	      }
-	    params.push_back (std::move (param));
-
-	    if (lexer.peek_token ()->get_id () != COMMA)
-	      {
-		if (lexer.peek_token ()->get_id () == OR)
-		  lexer.split_current_token (PIPE, PIPE);
-		// not an error but means param list is done
-		break;
-	      }
-	    // skip comma
-	    lexer.skip_token ();
-
-	    if (lexer.peek_token ()->get_id () == OR)
-	      lexer.split_current_token (PIPE, PIPE);
-
-	    t = lexer.peek_token ();
-	  }
-
-	if (!skip_token (PIPE))
-	  {
-	    return nullptr;
-	  }
-	break;
-      }
-    default:
-      add_error (Error (tok->get_locus (),
-			"unexpected token %qs in closure expression - expected "
-			"%<|%> or %<||%>",
-			tok->get_token_description ()));
-
-      // skip somewhere?
-      return nullptr;
-    }
-
-  // again branch based on next token
-  tok = lexer.peek_token ();
-  if (tok->get_id () == RETURN_TYPE)
-    {
-      // must be return type closure with block expr
-
-      // skip "return type" token
-      lexer.skip_token ();
-
-      // parse actual type, which is required
-      std::unique_ptr<AST::TypeNoBounds> type = parse_type_no_bounds ();
-      if (type == nullptr)
-	{
-	  // error
-	  Error error (tok->get_locus (), "failed to parse type for closure");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      // parse block expr, which is required
-      std::unique_ptr<AST::BlockExpr> block = parse_block_expr ();
-      if (block == nullptr)
-	{
-	  // error
-	  Error error (lexer.peek_token ()->get_locus (),
-		       "failed to parse block expr in closure");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      return std::unique_ptr<AST::ClosureExprInnerTyped> (
-	new AST::ClosureExprInnerTyped (std::move (type), std::move (block),
-					std::move (params), locus, has_move,
-					std::move (outer_attrs)));
-    }
-  else
-    {
-      // must be expr-only closure
-
-      // parse expr, which is required
-      std::unique_ptr<AST::Expr> expr = parse_expr ();
-      if (expr == nullptr)
-	{
-	  Error error (tok->get_locus (),
-		       "failed to parse expression in closure");
-	  add_error (std::move (error));
-
-	  // skip somewhere?
-	  return nullptr;
-	}
-
-      return std::unique_ptr<AST::ClosureExprInner> (
-	new AST::ClosureExprInner (std::move (expr), std::move (params), locus,
-				   has_move, std::move (outer_attrs)));
-    }
-}
-
-// Returns true if the next token is END, ELSE, or EOF;
-template <typename ManagedTokenSource>
-bool
-Parser<ManagedTokenSource>::done_end_or_else ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-  return (t->get_id () == RIGHT_CURLY || t->get_id () == ELSE
-	  || t->get_id () == END_OF_FILE);
-}
-
-// Returns true if the next token is END or EOF.
-template <typename ManagedTokenSource>
-bool
-Parser<ManagedTokenSource>::done_end ()
-{
-  const_TokenPtr t = lexer.peek_token ();
-  return (t->get_id () == RIGHT_CURLY || t->get_id () == END_OF_FILE);
-}
 } // namespace Rust
+
+#include "rust-parse-impl-utils.hxx"
+#include "rust-parse-impl-attribute.hxx"
+#include "rust-parse-impl-ttree.hxx"
+#include "rust-parse-impl-macro.hxx"
+#include "rust-parse-impl-path.hxx"
+#include "rust-parse-impl-pattern.hxx"
+#include "rust-parse-impl-expr.hxx"
